Bind a column from one table to another

Bind a column from one table to another

Usage

hts_cbind_var(
  lhs_table,
  rhs_var,
  hts_data,
  variable_list = variable_list,
  return_weight_cols = FALSE,
  cbind_ids = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id"),
  cbind_wts = c("hh_weight", "person_weight", "day_weight", "trip_weight", "hh_weight")
)

Arguments

lhs_table

Table to bind a column to in data.table format

rhs_var

Variable to bind to the lhs_table.

hts_data

List of household, person, vehicle, day, and trip tables in data.table format.

variable_list

A variable list with descriptions and table locations of variables.

return_weight_cols

If true binds weight variable along with rhs_var to lhs_table. Default is FALSE.

cbind_ids

list of unique identifiers for each table in hts_data

cbind_wts

list of weight for each table in hts_data

Value

Inputted table with inputted variable binded.

Examples

require(data.table)
hts_cbind_var(lhs_table = trip,
 rhs_var = 'speed_mph',
 hts_data = test_data,
  variable_list = variable_list)
#> Joining speed_mph to table on hh_id, person_id, day_id, trip_id, speed_mph, trip_weight
#> Key: <hh_id, person_id, day_id, trip_id, speed_mph, trip_weight>
#>           hh_id  person_id       day_id      trip_id speed_mph trip_weight
#>           <int>      <num>        <num>        <num>     <num>       <int>
#>     1: 23000173 2300017301 230001730102 2.300017e+12 21.480572         968
#>     2: 23000173 2300017301 230001730102 2.300017e+12 23.720178         906
#>     3: 23000173 2300017301 230001730102 2.300017e+12  6.927242         387
#>     4: 23000173 2300017301 230001730102 2.300017e+12  6.146634         684
#>     5: 23000173 2300017301 230001730103 2.300017e+12 34.657421          64
#>    ---                                                                    
#> 15816: 23422311 2342231101 234223110101 2.342231e+12 25.812478         624
#> 15817: 23422311 2342231102 234223110201 2.342231e+12 16.040986         322
#> 15818: 23422311 2342231102 234223110201 2.342231e+12 26.901306          22
#> 15819: 23422311 2342231102 234223110201 2.342231e+12 16.720703         596
#> 15820: 23422311 2342231102 234223110201 2.342231e+12 25.812478          95
#>        travel_date distance_miles mode_type mode_1 mode_2 num_travelers
#>             <Date>          <num>     <int>  <int>  <int>         <int>
#>     1:  2023-04-20       4.755560         8      7    995             2
#>     2:  2023-04-20       4.856047         8      7    995             3
#>     3:  2023-04-20       1.516296         1      1    995             1
#>     4:  2023-04-20       1.393237         1      1    995             1
#>     5:  2023-04-21       8.933913         8      7    995             1
#>    ---                                                                 
#> 15816:  2023-06-01      20.219774         8      6    995             2
#> 15817:  2023-06-01       6.683744         8      7    995             1
#> 15818:  2023-06-01      20.175979         8      7    995             1
#> 15819:  2023-06-01       5.573568         8      6    995             2
#> 15820:  2023-06-01      20.219774         8      6    995             2
#>        d_purpose_category
#>                     <int>
#>     1:                  9
#>     2:                  1
#>     3:                  9
#>     4:                  1
#>     5:                  3
#>    ---                   
#> 15816:                  1
#> 15817:                 10
#> 15818:                 13
#> 15819:                  6
#> 15820:                  1
hts_cbind_var(lhs_table = trip,
 rhs_var = 'speed_mph',
  hts_data = test_data,
variable_list = variable_list,
 return_weight_cols = TRUE)
#> Joining speed_mph to table on hh_id, person_id, day_id, trip_id, speed_mph, trip_weight
#> Key: <hh_id, person_id, day_id, trip_id, speed_mph, trip_weight>
#>           hh_id  person_id       day_id      trip_id speed_mph trip_weight
#>           <int>      <num>        <num>        <num>     <num>       <int>
#>     1: 23000173 2300017301 230001730102 2.300017e+12 21.480572         968
#>     2: 23000173 2300017301 230001730102 2.300017e+12 23.720178         906
#>     3: 23000173 2300017301 230001730102 2.300017e+12  6.927242         387
#>     4: 23000173 2300017301 230001730102 2.300017e+12  6.146634         684
#>     5: 23000173 2300017301 230001730103 2.300017e+12 34.657421          64
#>    ---                                                                    
#> 15816: 23422311 2342231101 234223110101 2.342231e+12 25.812478         624
#> 15817: 23422311 2342231102 234223110201 2.342231e+12 16.040986         322
#> 15818: 23422311 2342231102 234223110201 2.342231e+12 26.901306          22
#> 15819: 23422311 2342231102 234223110201 2.342231e+12 16.720703         596
#> 15820: 23422311 2342231102 234223110201 2.342231e+12 25.812478          95
#>        travel_date distance_miles mode_type mode_1 mode_2 num_travelers
#>             <Date>          <num>     <int>  <int>  <int>         <int>
#>     1:  2023-04-20       4.755560         8      7    995             2
#>     2:  2023-04-20       4.856047         8      7    995             3
#>     3:  2023-04-20       1.516296         1      1    995             1
#>     4:  2023-04-20       1.393237         1      1    995             1
#>     5:  2023-04-21       8.933913         8      7    995             1
#>    ---                                                                 
#> 15816:  2023-06-01      20.219774         8      6    995             2
#> 15817:  2023-06-01       6.683744         8      7    995             1
#> 15818:  2023-06-01      20.175979         8      7    995             1
#> 15819:  2023-06-01       5.573568         8      6    995             2
#> 15820:  2023-06-01      20.219774         8      6    995             2
#>        d_purpose_category
#>                     <int>
#>     1:                  9
#>     2:                  1
#>     3:                  9
#>     4:                  1
#>     5:                  3
#>    ---                   
#> 15816:                  1
#> 15817:                 10
#> 15818:                 13
#> 15819:                  6
#> 15820:                  1