Filter datasets to only keep specified ids

Filter datasets to only keep specified ids

Usage

hts_filter_data(hts_data, ids, id_name)

Arguments

hts_data

List of containing household travel data tables.

ids

List of ids to keep in all of the tables

id_name

Name of id being used for filtering (e.g., hh_id, person_id)

Value

Inputted list of tables filtered to the specified ids.

Examples

require(data.table)

data(test_data)

hts_filter_data(hts_data = test_data,
                ids = hh[num_people > 5, hh_id],
                id_name = 'hh_id')
#> $hh
#>       hh_id sample_segment num_trips num_people income_detailed income_followup
#> 1: 23077623              2        25          7               9             995
#> 2: 23214339              3        20         13               6             995
#> 3: 23225402             21         6          9               6             995
#> 4: 23230310             10        13          6               6             995
#> 5: 23238765             21         3          6               5             995
#> 6: 23242420             21        37          6               7             995
#> 7: 23275913              4        12          6               7             995
#> 8: 23293124              4        21          6              10             995
#> 9: 23411644             14        24          6               9             995
#>     home_lon home_lat home_county residence_type hh_weight
#> 1: -82.74274 35.80280           2              2       737
#> 2: -95.55556 37.30230           2              4       337
#> 3: -97.03704 39.67768           2              5       830
#> 4: -91.41141 33.21722           3              1        55
#> 5: -92.23223 37.64565           2              1       807
#> 6: -83.38338 38.75976           1              1       376
#> 7: -98.15816 39.72673           1              1       351
#> 8: -91.71171 39.95796           3              1       163
#> 9: -83.12312 38.17117           2              1       747
#> 
#> $person
#>     ethnicity_1 ethnicity_2 ethnicity_3 ethnicity_4 ethnicity_997 ethnicity_999
#>  1:           1           0           0           0             0             0
#>  2:         995         995         995         995           995           995
#>  3:           1           0           0           0             0             0
#>  4:           0           1           0           0             0             0
#>  5:           0           0           1           0             0             0
#>  6:           1           0           0           0             0             0
#>  7:           1           0           0           0             0             0
#>  8:           1           0           0           0             0             0
#>  9:           0           0           0           0             0             1
#> 10:           0           0           1           0             0             0
#> 11:           1           0           0           0             0             0
#> 12:           0           1           0           0             0             0
#> 13:           1           0           0           0             0             0
#> 14:           0           0           1           0             0             0
#> 15:           0           0           1           0             0             0
#> 16:           0           0           0           1             0             0
#> 17:           0           0           1           0             0             0
#> 18:           0           0           0           0             0             1
#> 19:           0           0           1           0             0             0
#> 20:           0           0           0           0             0             1
#> 21:           1           0           0           0             0             0
#> 22:           1           0           0           0             0             0
#> 23:           1           0           0           0             0             0
#> 24:           1           0           0           0             0             0
#> 25:           1           0           0           0             0             0
#> 26:           1           0           0           0             0             0
#> 27:         995         995         995         995           995           995
#> 28:         995         995         995         995           995           995
#> 29:         995         995         995         995           995           995
#> 30:           1           0           0           0             0             0
#> 31:           1           0           0           0             0             0
#> 32:         995         995         995         995           995           995
#> 33:         995         995         995         995           995           995
#> 34:         995         995         995         995           995           995
#> 35:         995         995         995         995           995           995
#> 36:           1           0           0           0             0             0
#> 37:           1           0           0           0             0             0
#> 38:           1           0           0           0             0             0
#> 39:         995         995         995         995           995           995
#> 40:         995         995         995         995           995           995
#> 41:         995         995         995         995           995           995
#> 42:           0           0           0           0             0             1
#> 43:           0           0           0           0             0             1
#> 44:         995         995         995         995           995           995
#> 45:         995         995         995         995           995           995
#> 46:         995         995         995         995           995           995
#> 47:         995         995         995         995           995           995
#> 48:           1           0           0           0             0             0
#> 49:           1           0           0           0             0             0
#> 50:         995         995         995         995           995           995
#> 51:         995         995         995         995           995           995
#> 52:         995         995         995         995           995           995
#> 53:           1           0           0           0             0             0
#> 54:           1           0           0           0             0             0
#> 55:           1           0           0           0             0             0
#> 56:         995         995         995         995           995           995
#> 57:         995         995         995         995           995           995
#> 58:         995         995         995         995           995           995
#> 59:         995         995         995         995           995           995
#> 60:           1           0           0           0             0             0
#> 61:           1           0           0           0             0             0
#> 62:         995         995         995         995           995           995
#> 63:         995         995         995         995           995           995
#> 64:         995         995         995         995           995           995
#> 65:         995         995         995         995           995           995
#>     ethnicity_1 ethnicity_2 ethnicity_3 ethnicity_4 ethnicity_997 ethnicity_999
#>     race_1 race_2 race_3 race_4 race_5 race_997 race_999    hh_id num_trips
#>  1:      0      0      0      0      1        0        0 23077623         6
#>  2:    995    995    995    995    995      995      995 23077623         0
#>  3:      0      0      0      0      1        0        0 23077623         1
#>  4:      0      0      0      1      0        0        0 23077623         6
#>  5:      0      0      0      1      0        0        0 23077623         0
#>  6:      0      0      0      0      1        0        0 23077623         6
#>  7:      0      0      1      0      0        0        0 23077623         6
#>  8:      0      0      0      0      0        0        1 23214339         6
#>  9:      0      0      0      0      0        0        1 23214339         2
#> 10:      0      0      0      0      0        1        0 23214339         2
#> 11:      0      0      0      0      1        0        0 23214339         1
#> 12:      0      0      0      0      1        0        0 23214339         1
#> 13:      0      0      0      0      1        0        0 23214339         1
#> 14:      0      0      0      0      1        0        0 23214339         1
#> 15:      0      0      0      0      1        0        0 23214339         1
#> 16:      0      0      0      0      1        0        0 23214339         1
#> 17:      1      0      0      0      0        0        0 23214339         1
#> 18:      0      0      0      0      0        0        1 23214339         1
#> 19:      0      0      0      0      1        0        0 23214339         1
#> 20:      0      0      0      0      1        0        0 23214339         1
#> 21:      0      0      0      0      1        0        0 23225402         0
#> 22:      0      0      0      0      1        0        0 23225402         3
#> 23:      0      0      0      0      1        0        0 23225402         3
#> 24:      0      0      0      0      1        0        0 23225402         0
#> 25:      0      0      0      0      1        0        0 23225402         0
#> 26:      0      0      0      0      1        0        0 23225402         0
#> 27:    995    995    995    995    995      995      995 23225402         0
#> 28:    995    995    995    995    995      995      995 23225402         0
#> 29:    995    995    995    995    995      995      995 23225402         0
#> 30:      0      0      0      0      1        0        0 23230310         4
#> 31:      0      0      0      0      1        0        0 23230310         0
#> 32:    995    995    995    995    995      995      995 23230310         3
#> 33:    995    995    995    995    995      995      995 23230310         2
#> 34:    995    995    995    995    995      995      995 23230310         2
#> 35:    995    995    995    995    995      995      995 23230310         2
#> 36:      0      0      0      0      1        0        0 23238765         3
#> 37:      0      0      0      0      1        0        0 23238765         0
#> 38:      0      0      0      0      1        0        0 23238765         0
#> 39:    995    995    995    995    995      995      995 23238765         0
#> 40:    995    995    995    995    995      995      995 23238765         0
#> 41:    995    995    995    995    995      995      995 23238765         0
#> 42:      0      0      0      0      0        0        1 23242420         6
#> 43:      0      0      0      0      0        0        1 23242420         7
#> 44:    995    995    995    995    995      995      995 23242420         6
#> 45:    995    995    995    995    995      995      995 23242420         6
#> 46:    995    995    995    995    995      995      995 23242420         6
#> 47:    995    995    995    995    995      995      995 23242420         6
#> 48:      0      0      1      0      0        0        0 23275913         4
#> 49:      0      0      1      0      0        0        0 23275913         0
#> 50:    995    995    995    995    995      995      995 23275913         2
#> 51:    995    995    995    995    995      995      995 23275913         2
#> 52:    995    995    995    995    995      995      995 23275913         2
#> 53:      0      0      1      0      0        0        0 23275913         2
#> 54:      0      0      0      0      1        0        0 23293124         4
#> 55:      0      0      0      0      1        0        0 23293124         5
#> 56:    995    995    995    995    995      995      995 23293124         3
#> 57:    995    995    995    995    995      995      995 23293124         3
#> 58:    995    995    995    995    995      995      995 23293124         6
#> 59:    995    995    995    995    995      995      995 23293124         0
#> 60:      0      0      0      0      1        0        0 23411644         4
#> 61:      0      0      0      0      1        0        0 23411644         6
#> 62:    995    995    995    995    995      995      995 23411644         4
#> 63:    995    995    995    995    995      995      995 23411644         2
#> 64:    995    995    995    995    995      995      995 23411644         2
#> 65:    995    995    995    995    995      995      995 23411644         6
#>     race_1 race_2 race_3 race_4 race_5 race_997 race_999    hh_id num_trips
#>      person_id age gender employment education person_weight
#>  1: 2307762301   7      2          1         7           697
#>  2: 2307762302   2      2        995       995           752
#>  3: 2307762303   7      1          1         7           793
#>  4: 2307762304   9      1          5         6           801
#>  5: 2307762305   4      1          6         3           840
#>  6: 2307762306  10      2          5         7           527
#>  7: 2307762307  11      1          5         7            94
#>  8: 2321433901   5      2          1         7            98
#>  9: 2321433902   5      1          1         5           129
#> 10: 2321433903   5      1          1         7           681
#> 11: 2321433904   9      2          5         3           665
#> 12: 2321433905   8      1          5         2           362
#> 13: 2321433906   5    999          2         1           405
#> 14: 2321433907   4    999          7         5           860
#> 15: 2321433908   4    999          7         7           631
#> 16: 2321433909   5    999          2         5           535
#> 17: 2321433910   5    999          2         6            85
#> 18: 2321433911   6    999          1         5           602
#> 19: 2321433912   6    999          2         4           664
#> 20: 2321433913   6    999          2         6           903
#> 21: 2322540201   7      2          5         3           617
#> 22: 2322540202   5      1          5         3           419
#> 23: 2322540203   7      1          1         6           671
#> 24: 2322540204   5      2          5         6           212
#> 25: 2322540205   4      2          5         2           492
#> 26: 2322540206  10      1          5         3           578
#> 27: 2322540207   1    999        995       995           918
#> 28: 2322540208   5    995          1       995           801
#> 29: 2322540209   1    995        995       995           903
#> 30: 2323031001   6      1          1         6           802
#> 31: 2323031002   5      2          1         2           824
#> 32: 2323031003   2      2        995       995           561
#> 33: 2323031004   2      1        995       995           261
#> 34: 2323031005   2      2        995       995           617
#> 35: 2323031006   2      1        995       995           100
#> 36: 2323876501   7      1          1         6           669
#> 37: 2323876502   8      2          5         3           505
#> 38: 2323876503   4      2          6         3           104
#> 39: 2323876504   2      1        995       995           813
#> 40: 2323876505   2      1        995       995           279
#> 41: 2323876506   2      1        995       995           410
#> 42: 2324242001   6    999          5         2           401
#> 43: 2324242002   6    999          1         2           999
#> 44: 2324242003   3    999          5       995           602
#> 45: 2324242004   2    999        995       995           236
#> 46: 2324242005   2    999        995       995           426
#> 47: 2324242006   1    999        995       995           346
#> 48: 2327591301   7      2          1         7           498
#> 49: 2327591302   9      1          5         2           389
#> 50: 2327591303   2      2        995       995           100
#> 51: 2327591304   2      2        995       995           320
#> 52: 2327591305   1      1        995       995           672
#> 53: 2327591306   6      1          1         2           526
#> 54: 2329312401   6      2          1         7           122
#> 55: 2329312402   6      1          5         7           647
#> 56: 2329312403   2      1        995       995            66
#> 57: 2329312404   2      2        995       995           698
#> 58: 2329312405   1      1        995       995           470
#> 59: 2329312406   1      2        995       995           641
#> 60: 2341164401   6      1          1         5           657
#> 61: 2341164402   6      2          1         6           689
#> 62: 2341164403   2      1        995       995           191
#> 63: 2341164404   2      1        995       995           707
#> 64: 2341164405   2      2        995       995           226
#> 65: 2341164406   2      1        995       995           556
#>      person_id age gender employment education person_weight
#> 
#> $day
#>     delivery_2 delivery_3 delivery_4 delivery_5 delivery_6 delivery_7
#>  1:          0          0          1          1          0          0
#>  2:        995        995        995        995        995        995
#>  3:        995        995        995        995        995        995
#>  4:        995        995        995        995        995        995
#>  5:        995        995        995        995        995        995
#>  6:        995        995        995        995        995        995
#>  7:        995        995        995        995        995        995
#>  8:          0          0          0          1          0          0
#>  9:        995        995        995        995        995        995
#> 10:        995        995        995        995        995        995
#> 11:        995        995        995        995        995        995
#> 12:        995        995        995        995        995        995
#> 13:        995        995        995        995        995        995
#> 14:        995        995        995        995        995        995
#> 15:        995        995        995        995        995        995
#> 16:        995        995        995        995        995        995
#> 17:        995        995        995        995        995        995
#> 18:        995        995        995        995        995        995
#> 19:        995        995        995        995        995        995
#> 20:        995        995        995        995        995        995
#> 21:          0          0          0          0          0          0
#> 22:        995        995        995        995        995        995
#> 23:        995        995        995        995        995        995
#> 24:        995        995        995        995        995        995
#> 25:        995        995        995        995        995        995
#> 26:        995        995        995        995        995        995
#> 27:        995        995        995        995        995        995
#> 28:          1          0          0          1          0          0
#> 29:        995        995        995        995        995        995
#> 30:        995        995        995        995        995        995
#> 31:        995        995        995        995        995        995
#> 32:        995        995        995        995        995        995
#> 33:        995        995        995        995        995        995
#> 34:          0          0          0          1          0          0
#> 35:        995        995        995        995        995        995
#> 36:        995        995        995        995        995        995
#> 37:        995        995        995        995        995        995
#> 38:        995        995        995        995        995        995
#> 39:        995        995        995        995        995        995
#> 40:          0          1          0          1          0          0
#> 41:        995        995        995        995        995        995
#> 42:        995        995        995        995        995        995
#> 43:        995        995        995        995        995        995
#> 44:        995        995        995        995        995        995
#> 45:        995        995        995        995        995        995
#> 46:          0          0          0          0          0          0
#> 47:        995        995        995        995        995        995
#> 48:        995        995        995        995        995        995
#> 49:        995        995        995        995        995        995
#> 50:        995        995        995        995        995        995
#> 51:        995        995        995        995        995        995
#> 52:          0          1          1          1          0          0
#> 53:        995        995        995        995        995        995
#> 54:        995        995        995        995        995        995
#> 55:        995        995        995        995        995        995
#> 56:        995        995        995        995        995        995
#> 57:        995        995        995        995        995        995
#> 58:          0          0          0          1          0          0
#> 59:        995        995        995        995        995        995
#> 60:        995        995        995        995        995        995
#> 61:        995        995        995        995        995        995
#> 62:        995        995        995        995        995        995
#> 63:        995        995        995        995        995        995
#>     delivery_2 delivery_3 delivery_4 delivery_5 delivery_6 delivery_7
#>     delivery_8 delivery_996    hh_id num_trips  person_id       day_id
#>  1:          0            0 23077623         6 2307762301 230776230101
#>  2:        995          995 23077623         0 2307762302 230776230201
#>  3:        995          995 23077623         1 2307762303 230776230301
#>  4:        995          995 23077623         6 2307762304 230776230401
#>  5:        995          995 23077623         0 2307762305 230776230501
#>  6:        995          995 23077623         6 2307762306 230776230601
#>  7:        995          995 23077623         6 2307762307 230776230701
#>  8:          0            0 23214339         6 2321433901 232143390101
#>  9:        995          995 23214339         2 2321433902 232143390201
#> 10:        995          995 23214339         2 2321433903 232143390301
#> 11:        995          995 23214339         1 2321433904 232143390401
#> 12:        995          995 23214339         1 2321433905 232143390501
#> 13:        995          995 23214339         1 2321433906 232143390601
#> 14:        995          995 23214339         1 2321433907 232143390701
#> 15:        995          995 23214339         1 2321433908 232143390801
#> 16:        995          995 23214339         1 2321433909 232143390901
#> 17:        995          995 23214339         1 2321433910 232143391001
#> 18:        995          995 23214339         1 2321433911 232143391101
#> 19:        995          995 23214339         1 2321433912 232143391201
#> 20:        995          995 23214339         1 2321433913 232143391301
#> 21:          0            1 23225402         0 2322540201 232254020101
#> 22:        995          995 23225402         3 2322540202 232254020201
#> 23:        995          995 23225402         3 2322540203 232254020301
#> 24:        995          995 23225402         0 2322540204 232254020401
#> 25:        995          995 23225402         0 2322540205 232254020501
#> 26:        995          995 23225402         0 2322540206 232254020601
#> 27:        995          995 23225402         0 2322540207 232254020701
#> 28:          0            0 23230310         4 2323031001 232303100101
#> 29:        995          995 23230310         0 2323031002 232303100201
#> 30:        995          995 23230310         3 2323031003 232303100301
#> 31:        995          995 23230310         2 2323031004 232303100401
#> 32:        995          995 23230310         2 2323031005 232303100501
#> 33:        995          995 23230310         2 2323031006 232303100601
#> 34:          0            0 23238765         3 2323876501 232387650101
#> 35:        995          995 23238765         0 2323876502 232387650201
#> 36:        995          995 23238765         0 2323876503 232387650301
#> 37:        995          995 23238765         0 2323876504 232387650401
#> 38:        995          995 23238765         0 2323876505 232387650501
#> 39:        995          995 23238765         0 2323876506 232387650601
#> 40:          0            0 23242420         6 2324242001 232424200101
#> 41:        995          995 23242420         7 2324242002 232424200201
#> 42:        995          995 23242420         6 2324242003 232424200301
#> 43:        995          995 23242420         6 2324242004 232424200401
#> 44:        995          995 23242420         6 2324242005 232424200501
#> 45:        995          995 23242420         6 2324242006 232424200601
#> 46:          0            1 23275913         4 2327591301 232759130101
#> 47:        995          995 23275913         0 2327591302 232759130201
#> 48:        995          995 23275913         2 2327591303 232759130301
#> 49:        995          995 23275913         2 2327591304 232759130401
#> 50:        995          995 23275913         2 2327591305 232759130501
#> 51:        995          995 23275913         2 2327591306 232759130601
#> 52:          0            0 23293124         4 2329312401 232931240101
#> 53:        995          995 23293124         5 2329312402 232931240201
#> 54:        995          995 23293124         3 2329312403 232931240301
#> 55:        995          995 23293124         3 2329312404 232931240401
#> 56:        995          995 23293124         6 2329312405 232931240501
#> 57:        995          995 23293124         0 2329312406 232931240601
#> 58:          0            0 23411644         4 2341164401 234116440101
#> 59:        995          995 23411644         6 2341164402 234116440201
#> 60:        995          995 23411644         4 2341164403 234116440301
#> 61:        995          995 23411644         2 2341164404 234116440401
#> 62:        995          995 23411644         2 2341164405 234116440501
#> 63:        995          995 23411644         6 2341164406 234116440601
#>     delivery_8 delivery_996    hh_id num_trips  person_id       day_id
#>     travel_date begin_day end_day day_weight
#>  1:  2023-05-03         1       1         29
#>  2:  2023-05-03         1       1        229
#>  3:  2023-05-03         3       1        599
#>  4:  2023-05-03         1       1        228
#>  5:  2023-05-03       997     997        935
#>  6:  2023-05-03         1       1        813
#>  7:  2023-05-03         1       1        453
#>  8:  2023-05-29         1       1         90
#>  9:  2023-05-29         1       1        309
#> 10:  2023-05-29         1       1        338
#> 11:  2023-05-29         4       7        168
#> 12:  2023-05-29         4       7        634
#> 13:  2023-05-29         4       7         66
#> 14:  2023-05-29         4       7         91
#> 15:  2023-05-29         4       7        646
#> 16:  2023-05-29         1       7        355
#> 17:  2023-05-29         1       7        242
#> 18:  2023-05-29         1       7        853
#> 19:  2023-05-29         1       7        840
#> 20:  2023-05-29         1       7        796
#> 21:  2023-05-29         1       1        251
#> 22:  2023-05-29         1       1        485
#> 23:  2023-05-29         1       1        377
#> 24:  2023-05-29         1       1        780
#> 25:  2023-05-29         1       1        227
#> 26:  2023-05-29         1       1        799
#> 27:  2023-05-29         1       1        540
#> 28:  2023-05-31         1       1        240
#> 29:  2023-05-31       997     997        742
#> 30:  2023-05-31         1       1        943
#> 31:  2023-05-31         1       1        417
#> 32:  2023-05-31         1       1        582
#> 33:  2023-05-31         1       1        651
#> 34:  2023-05-29         1       1        755
#> 35:  2023-05-29         1       1        711
#> 36:  2023-05-29         1       1        986
#> 37:  2023-05-29         1       1         70
#> 38:  2023-05-29         1       1        599
#> 39:  2023-05-29         1       1        866
#> 40:  2023-05-24         1       1        788
#> 41:  2023-05-24         1       1        401
#> 42:  2023-05-24         1       1        752
#> 43:  2023-05-24         1       1        534
#> 44:  2023-05-24         1       1        353
#> 45:  2023-05-24         1       1         27
#> 46:  2023-06-08         1       1        997
#> 47:  2023-06-08         1       1        896
#> 48:  2023-06-08         1       1        670
#> 49:  2023-06-08         1       1        639
#> 50:  2023-06-08         1       1        264
#> 51:  2023-06-08         1       1        374
#> 52:  2023-06-08         1       1        639
#> 53:  2023-06-08         1       1        606
#> 54:  2023-06-08         1       1        118
#> 55:  2023-06-08         1       1        146
#> 56:  2023-06-08         1       1        374
#> 57:  2023-06-08         1       1        938
#> 58:  2023-05-31         1       3        322
#> 59:  2023-05-31         1       1        785
#> 60:  2023-05-31         1       1        518
#> 61:  2023-05-31         1       1        381
#> 62:  2023-05-31         1       1        715
#> 63:  2023-05-31         1       1        597
#>     travel_date begin_day end_day day_weight
#> 
#> $trip
#>         hh_id  person_id       day_id travel_date      trip_id  speed_mph
#>   1: 23077623 2307762301 230776230101  2023-05-03 2.307762e+12         NA
#>   2: 23077623 2307762301 230776230101  2023-05-03 2.307762e+12  2.6685051
#>   3: 23077623 2307762301 230776230101  2023-05-03 2.307762e+12         NA
#>   4: 23077623 2307762301 230776230101  2023-05-03 2.307762e+12         NA
#>   5: 23077623 2307762301 230776230101  2023-05-03 2.307762e+12  0.3321791
#>  ---                                                                     
#> 157: 23411644 2341164406 234116440601  2023-05-31 2.341164e+12  7.6175542
#> 158: 23411644 2341164406 234116440601  2023-05-31 2.341164e+12         NA
#> 159: 23411644 2341164406 234116440601  2023-05-31 2.341164e+12  6.3479618
#> 160: 23411644 2341164406 234116440601  2023-05-31 2.341164e+12 17.7310895
#> 161: 23411644 2341164406 234116440601  2023-05-31 2.341164e+12 17.7310895
#>      mode_type d_purpose_category trip_weight
#>   1:         1                 11         610
#>   2:        13                 11         560
#>   3:         1                  6         708
#>   4:         1                 11         422
#>   5:        13                 11         277
#>  ---                                         
#> 157:        13                 11         594
#> 158:         1                  4         900
#> 159:        10                  1         448
#> 160:         8                  9         836
#> 161:         8                  1         819
#> 
#> $vehicle
#>        hh_id vehicle_id fuel_type hh_weight
#>  1: 23077623 2307762301         1       737
#>  2: 23077623 2307762302         1       737
#>  3: 23077623 2307762303         1       737
#>  4: 23214339 2321433901         1       337
#>  5: 23225402 2322540201         1       830
#>  6: 23225402 2322540202         1       830
#>  7: 23225402 2322540203         1       830
#>  8: 23225402 2322540204         1       830
#>  9: 23225402 2322540205         5       830
#> 10: 23225402 2322540206         1       830
#> 11: 23225402 2322540207         1       830
#> 12: 23225402 2322540208         1       830
#> 13: 23230310 2323031001         2        55
#> 14: 23230310 2323031002         1        55
#> 15: 23230310 2323031003         1        55
#> 16: 23238765 2323876501         1       807
#> 17: 23238765 2323876502         2       807
#> 18: 23238765 2323876503         1       807
#> 19: 23242420 2324242001         1       376
#> 20: 23275913 2327591301         1       351
#> 21: 23275913 2327591302         1       351
#> 22: 23293124 2329312401         1       163
#> 23: 23293124 2329312402         3       163
#> 24: 23411644 2341164401         1       747
#> 25: 23411644 2341164402         1       747
#> 26: 23411644 2341164403         5       747
#>        hh_id vehicle_id fuel_type hh_weight
#>