Melts checkbox variables into a single variable

Melts checkbox variables into a single variable

hts_melt_vars(
  shared_name = NULL,
  wide_dt = NULL,
  shared_name_vars = NULL,
  variables_dt = variable_list,
  hts_data = hts_data,
  remove_missing = TRUE,
  missing_values = c("Missing Response", "995"),
  checkbox_label_sep = ":",
  to_single_row = FALSE
)

Arguments

shared_name

Shared name of the checkbox variable to melt (e.g., 'race_1', and 'race_2' have shared name of 'race'). Defaults to NULL.

wide_dt

Table containing the checkbox variables to melt in data.table format.

shared_name_vars

List of the checkbox variables with the same shared name. Defaults to NULL.

variables_dt

List of variable locations and descriptions in data.table format.

hts_data

List containing household, person, day, trip, and vehicle datasets in data.table format.

remove_missing

Boolean to remove rows with missing values. Defaults to TRUE.

missing_values

Missing values to remove. Defaults to 'Missing Response' and 995.

checkbox_label_sep

Character to use to split description of checkbox variable in the variable list. Defaults to ':'

to_single_row

Boolean if treating multiple checkbox selections as 'Two or more'. Defaults to FALSE.

Value

Inputted data table with checkbox variables melted into a single variable, with a 'variable' column to indicate original variable names and a 'value' column with the original value of the checkbox variable.

Examples

require(data.table)
require(stringr)
hts_melt_vars(shared_name = 'race',
              wide_dt = person,
              hts_data = list('hh' = hh,
                              'person' = person,
                              'day' = day,
                              'trip' = trip,
                              'vehicle' = vehicle))
#>           hh_id  person_id person_weight variable value
#>     1: 23000173 2300017301           139   race_1     0
#>     2: 23000173 2300017304           747   race_1     0
#>     3: 23000738 2300073801           298   race_1     0
#>     4: 23000858 2300085801           865   race_1     0
#>     5: 23000858 2300085802            87   race_1     0
#>    ---                                                 
#> 11679: 23416563 2341656302           889 race_999     0
#> 11680: 23418712 2341871201           476 race_999     0
#> 11681: 23419141 2341914101           136 race_999     0
#> 11682: 23419141 2341914102           170 race_999     0
#> 11683: 23419703 2341970301           645 race_999     0
#>                             race
#>     1: African American or Black
#>     2: African American or Black
#>     3: African American or Black
#>     4: African American or Black
#>     5: African American or Black
#>    ---                          
#> 11679:      Prefer not to answer
#> 11680:      Prefer not to answer
#> 11681:      Prefer not to answer
#> 11682:      Prefer not to answer
#> 11683:      Prefer not to answer