Prepare variable to be summarized

Prepare variable to be summarized

hts_prep_byvar(
  summarize_by = NULL,
  variables_dt = variables_list,
  hts_data,
  ...
)

Arguments

summarize_by

Name of the variable to be prepped. Default is NULL.

variables_dt

List of variable locations and descriptions in data.table format.

hts_data

List containing household, person, day, trip, and vehicle datasets in data.table format.

...

Additional parameters to pass to link{hts_melt_vars}

Value

Data table containing the variable to be summarized and other key columns.

Examples

hts_prep_byvar(summarize_by = 'age',
               variables_dt = variable_list,
               hts_data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#>          hh_id  person_id person_weight age
#>    1: 23000173 2300017301           139   7
#>    2: 23000173 2300017302            17   3
#>    3: 23000173 2300017303           400   2
#>    4: 23000173 2300017304           747   6
#>    5: 23000738 2300073801           298   8
#>   ---                                      
#> 2038: 23418712 2341871203           809   2
#> 2039: 23419141 2341914101           136   6
#> 2040: 23419141 2341914102           170   5
#> 2041: 23419141 2341914103           490   1
#> 2042: 23419703 2341970301           645   9
hts_prep_byvar(summarize_by = 'race',
               variables_dt = variable_list,
               hts_data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#>          hh_id  person_id person_weight                      race
#>    1: 23025240 2302524001           243 African American or Black
#>    2: 23029096 2302909601           772 African American or Black
#>    3: 23033218 2303321802           170 African American or Black
#>    4: 23034871 2303487101           977 African American or Black
#>    5: 23035194 2303519401           630 African American or Black
#>   ---                                                            
#> 1665: 23411086 2341108601           311      Prefer not to answer
#> 1666: 23411086 2341108602           827      Prefer not to answer
#> 1667: 23413086 2341308603           514      Prefer not to answer
#> 1668: 23413585 2341358501           288      Prefer not to answer
#> 1669: 23413585 2341358502           603      Prefer not to answer