Prepare datasets to make summaries

Prepare datasets to make summaries

Usage

hts_prep_data(
  summarize_var = NULL,
  summarize_by = NULL,
  variables_dt = variable_list,
  data = hts_data,
  id_cols = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id"),
  weighted = TRUE,
  wt_cols = c("hh_weight", "person_weight", "day_weight", "trip_weight", "hh_weight"),
  remove_outliers = TRUE,
  threshold = 0.975,
  remove_missing = TRUE,
  missing_values = c("Missing Response", "995"),
  not_imputable = -1,
  strataname = NULL
)

Arguments

summarize_var

Name of the variable to summarize. Default is NULL

summarize_by

Name of the variable to summarize the summarize_var by. Default is NULL.

variables_dt

List of variable locations and descriptions in data.table format.

data

List of household, person, vehicle, day, and trip tables in data.table format.

id_cols

name of unique identifier for each table in hts_data

weighted

Whether the data is weighted. Default is TRUE.

wt_cols

weight name for each table in hts_data

remove_outliers

Whether to remove outliers for numeric variable. Default is TRUE.

threshold

Threshold to define outliers. Default is 0.975.

remove_missing

Whether to remove missing values from the summary. Default is TRUE.

missing_values

Missing values to remove. Default is 995.

not_imputable

Value representing 'Not imputable' to remove. Default is -1.

strataname

Name of strata name to bring in. Default is NULL.

Value

List containing the categorical and numeric datasets of the summary variables and key columns, and either whether the summarize variable is shared or a breakdown of outliers, depending on if the summarize variable is categorical or numeric.

Examples

require(data.table)
require(stringr)
hts_prep_data(summarize_var = 'age',
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> Creating a summary of person age 
#> $cat
#>          hh_id  person_id age person_weight
#>    1: 23000137 2300013701   5           320
#>    2: 23000137 2300013702   5           827
#>    3: 23000185 2300018501   9           315
#>    4: 23000185 2300018502   9           933
#>    5: 23000738 2300073801   8           489
#>   ---                                      
#> 2026: 23421077 2342107701   8           813
#> 2027: 23422853 2342285301   7           980
#> 2028: 23422853 2342285302   7           679
#> 2029: 23422853 2342285303   2           882
#> 2030: 23422853 2342285304   2           525
#> 
#> $num
#> NULL
#> 
#> $var_is_shared
#> [1] FALSE
#> 
hts_prep_data(summarize_var = 'speed_mph',
              summarize_by = 'age',
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> Creating a summary of trip speed_mph broken down by person age
#> $cat
#>           hh_id  person_id       day_id      trip_id person_weight trip_weight
#>     1: 23000137 2300013701 230001370101 2.300014e+12           320         890
#>     2: 23000137 2300013701 230001370101 2.300014e+12           320         663
#>     3: 23000137 2300013701 230001370101 2.300014e+12           320         892
#>     4: 23000137 2300013701 230001370101 2.300014e+12           320         146
#>     5: 23000137 2300013701 230001370101 2.300014e+12           320         630
#>    ---                                                                        
#> 16279: 23422853 2342285303 234228530301 2.342285e+12           882         935
#> 16280: 23422853 2342285303 234228530301 2.342285e+12           882         234
#> 16281: 23422853 2342285303 234228530301 2.342285e+12           882         379
#> 16282: 23422853 2342285304 234228530401 2.342285e+12           525         210
#> 16283: 23422853 2342285304 234228530401 2.342285e+12           525         468
#>        speed_mph age
#>     1:      1-10   5
#>     2:      1-10   5
#>     3:      1-10   5
#>     4:      1-10   5
#>     5:      1-10   5
#>    ---              
#> 16279:      1-10   2
#> 16280:      1-10   2
#> 16281:      1-10   2
#> 16282:      1-10   2
#> 16283:      1-10   2
#> 
#> $num
#>           hh_id  person_id       day_id      trip_id person_weight trip_weight
#>     1: 23000137 2300013701 230001370101 2.300014e+12           320         890
#>     2: 23000137 2300013701 230001370101 2.300014e+12           320         663
#>     3: 23000137 2300013701 230001370101 2.300014e+12           320         892
#>     4: 23000137 2300013701 230001370101 2.300014e+12           320         146
#>     5: 23000137 2300013701 230001370101 2.300014e+12           320         630
#>    ---                                                                        
#> 16279: 23422853 2342285303 234228530301 2.342285e+12           882         935
#> 16280: 23422853 2342285303 234228530301 2.342285e+12           882         234
#> 16281: 23422853 2342285303 234228530301 2.342285e+12           882         379
#> 16282: 23422853 2342285304 234228530401 2.342285e+12           525         210
#> 16283: 23422853 2342285304 234228530401 2.342285e+12           525         468
#>        speed_mph age
#>     1:  4.764584   5
#>     2:  7.425306   5
#>     3:  5.372348   5
#>     4:  3.232303   5
#>     5:  3.106531   5
#>    ---              
#> 16279:  8.094207   2
#> 16280:  7.888320   2
#> 16281:  6.378893   2
#> 16282:  8.094207   2
#> 16283:  7.888320   2
#> 
#> $outliers
#>    threshold num_removed min_outlier max_outlier
#> 1:     0.975         418    123.5646    466317.9
#> 
                            
                            
hts_prep_data(summarize_var = 'employment',
              summarize_by = c('age', 'race'),
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> Creating a summary of person employment broken down by person age and person race
#> $cat
#>          hh_id  person_id person_weight employment age                 race
#>    1: 23000137 2300013701           320          1   5                White
#>    2: 23000137 2300013702           827          1   5 Prefer not to answer
#>    3: 23000185 2300018501           315          5   9                White
#>    4: 23000185 2300018502           933          1   9                White
#>    5: 23000738 2300073801           489          1   8                White
#>   ---                                                                      
#> 1657: 23419373 2341937301           108          1   8                White
#> 1658: 23419703 2341970301           501          5   9                White
#> 1659: 23421077 2342107701           813          1   8                White
#> 1660: 23422853 2342285301           980          1   7                Asian
#> 1661: 23422853 2342285302           679          1   7                White
#> 
#> $num
#> NULL
#> 
#> $var_is_shared
#> [1] FALSE
#>