Remove missing data for summary variables

Remove missing data for summary variables

Usage

hts_remove_missing_data(
  hts_data,
  variables_dt,
  summarize_var,
  ids = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id"),
  summarize_by = NULL,
  missing_values = c("Missing Response", "995"),
  not_imputable = -1
)

Arguments

hts_data

List containing household, person, day, trip, and vehicle datasets in data.table format.

variables_dt

A variable list with descriptions and table locations of variables.

summarize_var

Variable to be summarized that has it's missing data removed.

ids

names of unique identifiers for each table in hts_data

summarize_by

Variable being summarized by that has it's missing data removed. Default is NULL.

missing_values

Missing values that will be removed. Defaults are 995 and 'Missing Response'.

not_imputable

Value meaning not_imputable that will be removed. Default is -1.

Value

Inputted list of datasets without missing values for specified variables.

Examples

require(data.table)
hts_remove_missing_data(hts_data = list('hh' = hh,
                                        'person' = person,
                                        'day' = day,
                                        'trip' = trip,
                                        'vehicle' = vehicle),
                        variables_dt = variable_list,
                        summarize_var = 'speed_mph',
                        summarize_by = 'mode_type')
#> $hh
#>          hh_id sample_segment num_trips num_people income_detailed
#>    1: 23000137             16       102          2               8
#>    2: 23000185             19         3          2               7
#>    3: 23000738              4        16          1             999
#>    4: 23000946             16        28          1               8
#>    5: 23001247             16        58          3              10
#>   ---                                                             
#>  996: 23414947             14         2          2               6
#>  997: 23419373              5         3          1               9
#>  998: 23419703             14         2          1               2
#>  999: 23421077             14         0          1               6
#> 1000: 23422853              5         7          4              10
#>       income_followup  home_lon home_lat home_county residence_type hh_weight
#>    1:             995 -90.53053 39.20821           1            995       249
#>    2:             995 -91.69169 35.22122           3              1       420
#>    3:             999 -97.47748 39.07508           1              1       501
#>    4:             995 -94.21421 37.73674           3              1       519
#>    5:             995 -80.00000 33.35736           1              2       493
#>   ---                                                                        
#>  996:             995 -80.54054 34.56957           2              1       418
#>  997:             995 -86.24625 35.80981           2              1       753
#>  998:             995 -90.95095 34.45746           1              1       294
#>  999:             995 -93.47347 37.69469           3              1       342
#> 1000:             995 -92.61261 36.16717           1              1       976
#> 
#> $person
#>       ethnicity_1 ethnicity_2 ethnicity_3 ethnicity_4 ethnicity_997
#>    1:           1           0           0           0             0
#>    2:           0           0           0           0             0
#>    3:           1           0           0           0             0
#>    4:           1           0           0           0             0
#>    5:           1           0           0           0             0
#>   ---                                                              
#> 2026:           1           0           0           0             0
#> 2027:           1           0           0           0             0
#> 2028:           1           0           0           0             0
#> 2029:         995         995         995         995           995
#> 2030:         995         995         995         995           995
#>       ethnicity_999 race_1 race_2 race_3 race_4 race_5 race_997 race_999
#>    1:             0      0      0      0      0      1        0        0
#>    2:             1      0      0      0      0      0        0        1
#>    3:             0      0      0      0      0      1        0        0
#>    4:             0      0      0      0      0      1        0        0
#>    5:             0      0      0      0      0      1        0        0
#>   ---                                                                   
#> 2026:             0      0      0      0      0      1        0        0
#> 2027:             0      0      0      1      0      0        0        0
#> 2028:             0      0      0      0      0      1        0        0
#> 2029:           995    995    995    995    995    995      995      995
#> 2030:           995    995    995    995    995    995      995      995
#>          hh_id num_trips  person_id age gender employment education
#>    1: 23000137        59 2300013701   5      1          1         6
#>    2: 23000137        43 2300013702   5      1          1         6
#>    3: 23000185         1 2300018501   9      1          5         3
#>    4: 23000185         2 2300018502   9      2          1         3
#>    5: 23000738        16 2300073801   8      1          1         7
#>   ---                                                              
#> 2026: 23421077         0 2342107701   8      2          1         4
#> 2027: 23422853         2 2342285301   7      1          1         6
#> 2028: 23422853         0 2342285302   7      2          1         7
#> 2029: 23422853         3 2342285303   2      2        995       995
#> 2030: 23422853         2 2342285304   2      1        995       995
#>       person_weight
#>    1:           320
#>    2:           827
#>    3:           315
#>    4:           933
#>    5:           489
#>   ---              
#> 2026:           813
#> 2027:           980
#> 2028:           679
#> 2029:           882
#> 2030:           525
#> 
#> $day
#>       delivery_2 delivery_3 delivery_4 delivery_5 delivery_6 delivery_7
#>    1:          0          1          0          0          0          0
#>    2:          0          0          0          0          0          0
#>    3:          0          0          0          0          0          0
#>    4:          0          0          0          0          0          0
#>    5:          0          0          0          0          0          0
#>   ---                                                                  
#> 4413:          0          0          0          0          0          0
#> 4414:          0          0          0          0          0          0
#> 4415:        995        995        995        995        995        995
#> 4416:        995        995        995        995        995        995
#> 4417:        995        995        995        995        995        995
#>       delivery_8 delivery_996    hh_id num_trips  person_id       day_id
#>    1:          0            0 23000137        13 2300013701 230001370101
#>    2:          0            1 23000137        15 2300013702 230001370201
#>    3:          0            1 23000137         2 2300013701 230001370102
#>    4:          0            1 23000137         0 2300013702 230001370202
#>    5:          0            1 23000137         7 2300013701 230001370103
#>   ---                                                                   
#> 4413:          0            1 23421077         0 2342107701 234210770101
#> 4414:          0            1 23422853         2 2342285301 234228530101
#> 4415:        995          995 23422853         0 2342285302 234228530201
#> 4416:        995          995 23422853         3 2342285303 234228530301
#> 4417:        995          995 23422853         2 2342285304 234228530401
#>       travel_date begin_day end_day day_weight
#>    1:  2023-04-25         1       1        861
#>    2:  2023-04-25         1       1        991
#>    3:  2023-04-26         1       1        489
#>    4:  2023-04-26         1       1        558
#>    5:  2023-04-27         1       1        119
#>   ---                                         
#> 4413:  2023-06-07         1       1        574
#> 4414:  2023-06-05         1       1        262
#> 4415:  2023-06-05         1       1        737
#> 4416:  2023-06-05         1       1        367
#> 4417:  2023-06-05         1       1        169
#> 
#> $trip
#>           hh_id  person_id       day_id travel_date      trip_id speed_mph
#>     1: 23000137 2300013701 230001370101  2023-04-25 2.300014e+12  4.764584
#>     2: 23000137 2300013701 230001370101  2023-04-25 2.300014e+12  7.425306
#>     3: 23000137 2300013701 230001370101  2023-04-25 2.300014e+12  5.372348
#>     4: 23000137 2300013701 230001370101  2023-04-25 2.300014e+12  3.232303
#>     5: 23000137 2300013701 230001370101  2023-04-25 2.300014e+12  3.106531
#>    ---                                                                    
#> 16996: 23422853 2342285303 234228530301  2023-06-05 2.342285e+12  8.094207
#> 16997: 23422853 2342285303 234228530301  2023-06-05 2.342285e+12  7.888320
#> 16998: 23422853 2342285303 234228530301  2023-06-05 2.342285e+12  6.378893
#> 16999: 23422853 2342285304 234228530401  2023-06-05 2.342285e+12  8.094207
#> 17000: 23422853 2342285304 234228530401  2023-06-05 2.342285e+12  7.888320
#>        mode_type d_purpose_category trip_weight
#>     1:         1                 11         890
#>     2:        13                 11         663
#>     3:         1                 10         892
#>     4:         1                  1         146
#>     5:         1                 11         630
#>    ---                                         
#> 16996:         8                  6         935
#> 16997:        10                  4         234
#> 16998:        10                  1         379
#> 16999:         8                  4         210
#> 17000:        10                  4         468
#> 
#> $vehicle
#>          hh_id vehicle_id fuel_type hh_weight
#>    1: 23000185 2300018501         1       420
#>    2: 23000185 2300018502         1       420
#>    3: 23000738 2300073801         5       501
#>    4: 23000946 2300094601         1       519
#>    5: 23000946 2300094602         4       519
#>   ---                                        
#> 1387: 23421077 2342107704         1       342
#> 1388: 23421077 2342107705         1       342
#> 1389: 23422853 2342285301         1       976
#> 1390: 23422853 2342285302         4       976
#> 1391: 23422853 2342285303         1       976
#>