Remove outliers from a numeric variable

Remove outliers from a numeric variable

hts_remove_outliers(var_dt, numvar = NULL, threshold = 0.975)

Arguments

var_dt

Dataset with a numeric variable to remove outliers from in data.table format.

numvar

Numeric variable to remove outliers from. Default is NULL.

threshold

Threshold to define what an outlier is. Default is .975.

Value

List of outliers removed and the dataset without the outliers.

Examples

require(data.table)
hts_remove_outliers(var_dt = trip, numvar = 'speed_mph')
#> $outlier_description
#>    threshold num_removed min_outlier max_outlier
#> 1:     0.975         386    110.2297    228233.1
#> 
#> $dt
#>           hh_id  person_id       day_id travel_date      trip_id speed_mph
#>     1: 23000173 2300017301 230001730102  2023-04-20 2.300017e+12 21.480572
#>     2: 23000173 2300017301 230001730102  2023-04-20 2.300017e+12 23.720178
#>     3: 23000173 2300017301 230001730102  2023-04-20 2.300017e+12  6.927242
#>     4: 23000173 2300017301 230001730102  2023-04-20 2.300017e+12  6.146634
#>     5: 23000173 2300017301 230001730103  2023-04-21 2.300017e+12 34.657421
#>    ---                                                                    
#> 15031: 23419141 2341914103 234191410301  2023-05-31 2.341914e+12  8.904454
#> 15032: 23419141 2341914103 234191410301  2023-05-31 2.341914e+12 43.284089
#> 15033: 23419141 2341914103 234191410301  2023-05-31 2.341914e+12 25.518845
#> 15034: 23419703 2341970301 234197030101  2023-05-31 2.341970e+12 13.131100
#> 15035: 23419703 2341970301 234197030101  2023-05-31 2.341970e+12 16.413875
#>        mode_type d_purpose_category trip_weight
#>     1:         8                  9         356
#>     2:         8                  1         679
#>     3:         1                  9         758
#>     4:         1                  1         314
#>     5:         8                  3         883
#>    ---                                         
#> 15031:         8                  1         950
#> 15032:         8                  6         123
#> 15033:         8                  1         720
#> 15034:         8                 13         523
#> 15035:         8                  1         298
#>