Summarize a categorical variable

Summarize a categorical variable

Usage

hts_summary_cat(
  prepped_dt,
  summarize_var = NULL,
  summarize_by = NULL,
  weighted = TRUE,
  se = FALSE,
  wtname = NULL,
  strataname = NULL,
  checkbox_valname = "value",
  checkbox_yesval = 1,
  summarize_vartype = "categorical",
  id_cols = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id")
)

Arguments

prepped_dt

Dataset containing the summary variables and key columns in data.table format.

summarize_var

Name of the categorical variable to summarize. Default is NULL.

summarize_by

Name of the variable to summarize the summarize_var by. Default is NULL.

weighted

Whether the data is weighted. Default is TRUE.

se

Whether to calculate standard error. Default is FALSE.

wtname

Name of the weight column to use. Default is NULL.

strataname

Name of strata name to bring in. Default is NULL.

checkbox_valname

Name of the column with the checkbox value. Default is 'value'. Must be provided if summarize_var is a checkbox variable.

checkbox_yesval

Value of checkbox_valname that indicates it was selected. Default is 1. Must be provided if summarize_var is a checkbox variable.

summarize_vartype

String; one of either 'categorical' (when the variable being summarized is categorical) or 'checkbox' (when the variable being summarized is derived from a multiple response, aka select-all-that-apply question).

id_cols

names of possible ids in prepped_dt to return unique counts of

Value

List of unweighted and weighted categorical summaries including counts and proportions.

Examples

require(data.table)
require(stringr)
require(dplyr)
require(srvyr)
DT = hts_prep_data(summarize_var = 'age',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
#> Creating a summary of person age 
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'age',
                wtname = 'person_weight')
#> $unwtd
#>     age count        prop
#>  1:   1    79 0.039818548
#>  2:   2   194 0.097782258
#>  3:   3    33 0.016633065
#>  4:   4    99 0.049899194
#>  5:   5   389 0.196068548
#>  6:   6   354 0.178427419
#>  7:   7   234 0.117943548
#>  8:   8   259 0.130544355
#>  9:   9   218 0.109879032
#> 10:  10   119 0.059979839
#> 11:  11     6 0.003024194
#> 
#> $wtd
#>     age count        prop    est
#>  1:   1    79 0.042732174  43036
#>  2:   2   194 0.092259038  92915
#>  3:   3    33 0.018525285  18657
#>  4:   4    99 0.049041316  49390
#>  5:   5   389 0.198440091 199851
#>  6:   6   354 0.171938517 173161
#>  7:   7   234 0.117023960 117856
#>  8:   8   259 0.134759857 135718
#>  9:   9   218 0.109464706 110243
#> 10:  10   119 0.062543317  62988
#> 11:  11     6 0.003271738   3295
#> 
#> $weight_name
#> [1] "person_weight"
#> 
                
DT = hts_prep_data(summarize_var = 'race',
                   summarize_by = 'employment',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
#> Creating a summary of person race broken down by person employment
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'race',
                summarize_by = 'employment',
                wtname = 'person_weight',
                summarize_vartype = 'checkbox')
#> $unwtd
#>     employment                                      race count        prop
#>  1:          1                 African American or Black    24 0.029268293
#>  2:          1          American Indian or Alaska Native     6 0.007317073
#>  3:          1                                     Asian   203 0.247560976
#>  4:          1 Native Hawaiian or other Pacific Islander    11 0.013414634
#>  5:          1                                     White   505 0.615853659
#>  6:          1                                Other race    13 0.015853659
#>  7:          1                      Prefer not to answer    92 0.112195122
#>  8:          2                 African American or Black    10 0.068965517
#>  9:          2          American Indian or Alaska Native     2 0.013793103
#> 10:          2                                     Asian    25 0.172413793
#> 11:          2 Native Hawaiian or other Pacific Islander     1 0.006896552
#> 12:          2                                     White   101 0.696551724
#> 13:          2                      Prefer not to answer    16 0.110344828
#> 14:          3                 African American or Black     1 0.009803922
#> 15:          3          American Indian or Alaska Native     3 0.029411765
#> 16:          3                                     Asian    15 0.147058824
#> 17:          3                                     White    76 0.745098039
#> 18:          3                                Other race     2 0.019607843
#> 19:          3                      Prefer not to answer     9 0.088235294
#> 20:          5                 African American or Black    10 0.021834061
#> 21:          5          American Indian or Alaska Native    11 0.024017467
#> 22:          5                                     Asian    69 0.150655022
#> 23:          5 Native Hawaiian or other Pacific Islander     1 0.002183406
#> 24:          5                                     White   343 0.748908297
#> 25:          5                                Other race     8 0.017467249
#> 26:          5                      Prefer not to answer    35 0.076419214
#> 27:          6                 African American or Black     8 0.103896104
#> 28:          6                                     Asian     9 0.116883117
#> 29:          6                                     White    51 0.662337662
#> 30:          6                                Other race     2 0.025974026
#> 31:          6                      Prefer not to answer    10 0.129870130
#> 32:          7                 African American or Black     3 0.200000000
#> 33:          7          American Indian or Alaska Native     1 0.066666667
#> 34:          7                                     Asian     4 0.266666667
#> 35:          7                                     White     8 0.533333333
#> 36:          7                                Other race     1 0.066666667
#> 37:          7                      Prefer not to answer     3 0.200000000
#> 38:          8                                     Asian     1 0.111111111
#> 39:          8                                     White     4 0.444444444
#> 40:          8                                Other race     1 0.111111111
#> 41:          8                      Prefer not to answer     3 0.333333333
#>     employment                                      race count        prop
#> 
#> $wtd
#>     employment                                      race count        prop
#>  1:          1                 African American or Black    24 0.028694995
#>  2:          1          American Indian or Alaska Native     6 0.009275864
#>  3:          1                                     Asian   203 0.258106330
#>  4:          1 Native Hawaiian or other Pacific Islander    11 0.017659958
#>  5:          1                                     White   505 0.620537498
#>  6:          1                                Other race    13 0.017940159
#>  7:          1                      Prefer not to answer    92 0.096883680
#>  8:          2                 African American or Black    10 0.063301044
#>  9:          2          American Indian or Alaska Native     2 0.015717331
#> 10:          2                                     Asian    25 0.154313159
#> 11:          2 Native Hawaiian or other Pacific Islander     1 0.006475810
#> 12:          2                                     White   101 0.680675103
#> 13:          2                      Prefer not to answer    16 0.135196028
#> 14:          3                 African American or Black     1 0.015622480
#> 15:          3          American Indian or Alaska Native     3 0.021448154
#> 16:          3                                     Asian    15 0.145460410
#> 17:          3                                     White    76 0.747802774
#> 18:          3                                Other race     2 0.013183358
#> 19:          3                      Prefer not to answer     9 0.099520239
#> 20:          5                 African American or Black    10 0.025855586
#> 21:          5          American Indian or Alaska Native    11 0.026347325
#> 22:          5                                     Asian    69 0.144806982
#> 23:          5 Native Hawaiian or other Pacific Islander     1 0.003297548
#> 24:          5                                     White   343 0.739481318
#> 25:          5                                Other race     8 0.021636542
#> 26:          5                      Prefer not to answer    35 0.080426285
#> 27:          6                 African American or Black     8 0.090334120
#> 28:          6                                     Asian     9 0.091319111
#> 29:          6                                     White    51 0.699473807
#> 30:          6                                Other race     2 0.022395604
#> 31:          6                      Prefer not to answer    10 0.133025739
#> 32:          7                 African American or Black     3 0.121598770
#> 33:          7          American Indian or Alaska Native     1 0.037509608
#> 34:          7                                     Asian     4 0.186625673
#> 35:          7                                     White     8 0.506994620
#> 36:          7                                Other race     1 0.037509608
#> 37:          7                      Prefer not to answer     3 0.295772483
#> 38:          8                                     Asian     1 0.166990763
#> 39:          8                                     White     4 0.401798736
#> 40:          8                                Other race     1 0.095527467
#> 41:          8                      Prefer not to answer     3 0.335683034
#>     employment                                      race count        prop
#>        est wtd_group_n
#>  1:  11777      410420
#>  2:   3807      410420
#>  3: 105932      410420
#>  4:   7248      410420
#>  5: 254681      410420
#>  6:   7363      410420
#>  7:  39763      410420
#>  8:   4692       74122
#>  9:   1165       74122
#> 10:  11438       74122
#> 11:    480       74122
#> 12:  50453       74122
#> 13:  10021       74122
#> 14:    775       49608
#> 15:   1064       49608
#> 16:   7216       49608
#> 17:  37097       49608
#> 18:    654       49608
#> 19:   4937       49608
#> 20:   6257      241998
#> 21:   6376      241998
#> 22:  35043      241998
#> 23:    798      241998
#> 24: 178953      241998
#> 25:   5236      241998
#> 26:  19463      241998
#> 27:   3485       38579
#> 28:   3523       38579
#> 29:  26985       38579
#> 30:    864       38579
#> 31:   5132       38579
#> 32:    791        6505
#> 33:    244        6505
#> 34:   1214        6505
#> 35:   3298        6505
#> 36:    244        6505
#> 37:   1924        6505
#> 38:    687        4114
#> 39:   1653        4114
#> 40:    393        4114
#> 41:   1381        4114
#>        est wtd_group_n
#> 
#> $weight_name
#> [1] "person_weight"
#> 
                
DT = hts_prep_data(summarize_var = 'employment',
                summarize_by = c('race', 'income_detailed', 'gender'),
                variables_dt = variable_list,
                data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))$cat
#> Creating a summary of person employment broken down by person race and hh income_detailed and person gender

hts_summary_cat(prepped_dt = DT,
                summarize_var = 'employment',
                summarize_by = c('race', 'income_detailed', 'gender'),
                wtname = 'person_weight')$unwtd
#>                           race income_detailed gender employment count
#>   1: African American or Black               1      1          5     2
#>   2: African American or Black               1      1          6     1
#>   3: African American or Black               1      2          6     1
#>   4: African American or Black               2      1          1     1
#>   5: African American or Black               2      1          5     2
#>  ---                                                                  
#> 394:      Prefer not to answer             999    999          1    24
#> 395:      Prefer not to answer             999    999          2     1
#> 396:      Prefer not to answer             999    999          3     1
#> 397:      Prefer not to answer             999    999          5     5
#> 398:      Prefer not to answer             999    999          6     1
#>           prop
#>   1: 0.6666667
#>   2: 0.3333333
#>   3: 1.0000000
#>   4: 0.2000000
#>   5: 0.4000000
#>  ---          
#> 394: 0.7500000
#> 395: 0.0312500
#> 396: 0.0312500
#> 397: 0.1562500
#> 398: 0.0312500