Summarize a numeric variable

Summarize a numeric variable

hts_summary_num(
  prepped_dt,
  summarize_var = NULL,
  summarize_by = NULL,
  weighted = TRUE,
  se = FALSE,
  wtname = NULL,
  strataname = NULL
)

Arguments

prepped_dt

A prepared dataset in data.table format with the variable to summarize, the variable to summarize by, and the weights, if used.

summarize_var

Name of the variable to summarize. Default is NULL.

summarize_by

Name of the variable to summarize the summarize_var by. Default is NULL.

weighted

Whether the data is weighted. Default is TRUE.

se

Whether to calculate standard error. Default is FALSE. Will be set to FALSE if weighted is FALSE.

wtname

Name of the weight column to use. Default is NULL. Must be specified when weighted = TRUE.

strataname

Name of strata name to bring in. Default is NULL.

Value

List of unweighted and weighted numeric summaries including count, min, max, mean, se, and median.

Examples

require(data.table)
require(stringr)
require(dplyr)
require(srvyr)
DT = hts_prep_data(summarize_var = 'speed_mph',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$num
#> Creating a summary of trip speed_mph 
hts_summary_num(prepped_dt = DT,
                summarize_var = 'speed_mph',
                wtname = 'trip_weight')
#> $unwtd
#>    count min      max     mean   median
#> 1: 15035   0 109.2184 12.89507 9.728796
#> 
#> $wtd
#>    count min      max     mean   median
#> 1: 15035   0 109.2184 12.84671 9.760141
#> 
#> $weight_name
#> [1] "trip_weight"
#> 
DT = hts_prep_data(summarize_var = 'speed_mph',
                   summarize_by = 'age', 
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$num
#> Creating a summary of trip speed_mph broken down by person age
hts_summary_num(prepped_dt = DT,
                summarize_var = 'speed_mph',
                summarize_by = 'age',
                wtname = 'trip_weight')
#> $unwtd
#>     age count        min       max     mean    median
#>  1:   1   311 0.35959216 101.19362 14.55031 10.280068
#>  2:   2   844 0.00000000  82.89088 12.26536 10.190479
#>  3:   3    89 0.00000000  98.43097 12.02007  9.323293
#>  4:   4   984 0.00000000 109.21842 11.97881  7.763898
#>  5:   5  3704 0.00000000 108.02176 12.21873  8.234977
#>  6:   6  3685 0.00000000 101.19362 12.50943  9.558940
#>  7:   7  1940 0.01815516 106.72662 13.84846 10.808561
#>  8:   8  1644 0.02554711 104.77523 14.14443 11.208044
#>  9:   9  1379 0.00000000 105.10755 13.40606 11.036031
#> 10:  10   433 0.07317172  76.07189 13.43742 11.760465
#> 11:  11    22 0.15677999  88.79454 16.50646 11.099190
#> 
#> $wtd
#>     age count        min       max     mean    median
#>  1:   1   311 0.35959216 101.19362 14.06340  9.556707
#>  2:   2   844 0.00000000  82.89088 12.69728 10.496980
#>  3:   3    89 0.00000000  98.43097 13.16383 12.224664
#>  4:   4   984 0.00000000 109.21842 11.79167  7.763898
#>  5:   5  3704 0.00000000 108.02176 12.04057  8.095823
#>  6:   6  3685 0.00000000 101.19362 12.47361  9.474190
#>  7:   7  1940 0.01815516 106.72662 13.84666 10.899718
#>  8:   8  1644 0.02554711 104.77523 14.20364 11.197504
#>  9:   9  1379 0.00000000 105.10755 13.34549 11.127732
#> 10:  10   433 0.07317172  76.07189 13.44185 12.035237
#> 11:  11    22 0.15677999  88.79454 14.11659 11.099190
#> 
#> $weight_name
#> [1] "trip_weight"
#>