Test person dataset

A dataset containing person level attributes of almost 2,000 persons from the 2023 Puget Sound Regional Council Household Travel Survey. The variables are as follows:

person

Format

person

A data frame with 1,999 rows and 19 columns:

hh_id

8 digit household ID

person_id

10 digit person ID

ethnicity_1

Not of Hispnic, Latino, or Spanish origin, response codes

ethnicity_2

Mexican, Mexican American, Chicano, response codes

ethnicity_3

Puerto Rican, response codes

ethnicity_4

Cuban, response codes

ethnicity_997

Another Hispanic, Latino, or Spanish origin, response codes

ethnicity_999

Prefer not to answer, response codes

race_1

African American or Black, response codes

race_2

Americian Indian or Alaska Native, response codes

race_3

Asian, response codes

race_4

Native Hawaiian or other Pacific Islander, response codes

race_5

White, response codes

race_997

Other race, response codes

race_999

Prefer not to answer, response codes

age

Age, response codes

gender

Gender, response codes

employment

Employment status, response codes

education

Highest level of education completed, response codes

Source

Export tables compiled from internal RSG database on 11/22/2023