Filter datasets to only keep specified ids

Filter datasets to only keep specified ids

Usage

hts_filter_data(hts_data, ids, id_name)

Arguments

hts_data

List of containing household travel data tables.

ids

List of ids to keep in all of the tables

id_name

Name of id being used for filtering (e.g., hh_id, person_id)

Value

Inputted list of tables filtered to the specified ids.

Examples

require(data.table)

data(test_data)

hts_filter_data(hts_data = test_data,
                ids = hh[num_people > 5, hh_id],
                id_name = 'hh_id')
#> $hh
#>       hh_id sample_segment num_trips num_people income_detailed income_followup
#> 1: 23016029              8        20          7               5             995
#> 2: 23121381             14       101          6               9             995
#> 3: 23122159             21         5          6               2             995
#> 4: 23150818             16         3          6               1             995
#> 5: 23233061             21        10          6               8             995
#> 6: 23233651             21        20          8               6             995
#> 7: 23240801             21         2          6               6             995
#> 8: 23411644             14        24          6               9             995
#> 9: 23416563             14        16          6               7             995
#>    residence_type hh_weight
#> 1:              2       974
#> 2:              1       826
#> 3:              1       938
#> 4:              1       602
#> 5:              1        39
#> 6:              1       345
#> 7:              1       783
#> 8:              1       154
#> 9:              2       928
#> 
#> $person
#>     ethnicity_1 ethnicity_2 ethnicity_3 ethnicity_4 ethnicity_997 ethnicity_999
#>  1:           1           0           0           0             0             0
#>  2:           1           0           0           0             0             0
#>  3:         995         995         995         995           995           995
#>  4:         995         995         995         995           995           995
#>  5:         995         995         995         995           995           995
#>  6:           1           0           0           0             0             0
#>  7:         995         995         995         995           995           995
#>  8:           1           0           0           0             0             0
#>  9:           1           0           0           0             0             0
#> 10:         995         995         995         995           995           995
#> 11:         995         995         995         995           995           995
#> 12:         995         995         995         995           995           995
#> 13:         995         995         995         995           995           995
#> 14:           0           1           0           0             0             0
#> 15:         995         995         995         995           995           995
#> 16:         995         995         995         995           995           995
#> 17:         995         995         995         995           995           995
#> 18:         995         995         995         995           995           995
#> 19:         995         995         995         995           995           995
#> 20:           1           0           0           0             0             0
#> 21:         995         995         995         995           995           995
#> 22:         995         995         995         995           995           995
#> 23:         995         995         995         995           995           995
#> 24:         995         995         995         995           995           995
#> 25:         995         995         995         995           995           995
#> 26:           1           0           0           0             0             0
#> 27:           1           0           0           0             0             0
#> 28:           1           0           0           0             0             0
#> 29:         995         995         995         995           995           995
#> 30:         995         995         995         995           995           995
#> 31:         995         995         995         995           995           995
#> 32:           1           0           0           0             0             0
#> 33:           1           0           0           0             0             0
#> 34:           1           0           0           0             0             0
#> 35:           1           0           0           0             0             0
#> 36:         995         995         995         995           995           995
#> 37:         995         995         995         995           995           995
#> 38:         995         995         995         995           995           995
#> 39:         995         995         995         995           995           995
#> 40:           0           0           0           0             0             1
#> 41:         995         995         995         995           995           995
#> 42:           0           0           0           0             0             1
#> 43:         995         995         995         995           995           995
#> 44:         995         995         995         995           995           995
#> 45:         995         995         995         995           995           995
#> 46:           1           0           0           0             0             0
#> 47:           1           0           0           0             0             0
#> 48:         995         995         995         995           995           995
#> 49:         995         995         995         995           995           995
#> 50:         995         995         995         995           995           995
#> 51:         995         995         995         995           995           995
#> 52:           1           0           0           0             0             0
#> 53:           1           0           0           0             0             0
#> 54:         995         995         995         995           995           995
#> 55:         995         995         995         995           995           995
#> 56:         995         995         995         995           995           995
#> 57:         995         995         995         995           995           995
#>     ethnicity_1 ethnicity_2 ethnicity_3 ethnicity_4 ethnicity_997 ethnicity_999
#>     race_1 race_2 race_3 race_4 race_5 race_997 race_999    hh_id num_trips
#>  1:      0      0      0      0      1        0        0 23016029         5
#>  2:      0      0      0      0      1        0        0 23016029         0
#>  3:    995    995    995    995    995      995      995 23016029         2
#>  4:    995    995    995    995    995      995      995 23016029         2
#>  5:    995    995    995    995    995      995      995 23016029         5
#>  6:      0      0      0      0      1        0        0 23016029         2
#>  7:    995    995    995    995    995      995      995 23016029         4
#>  8:      0      0      0      0      1        0        0 23121381        45
#>  9:      0      0      0      0      1        0        0 23121381        15
#> 10:    995    995    995    995    995      995      995 23121381         0
#> 11:    995    995    995    995    995      995      995 23121381        13
#> 12:    995    995    995    995    995      995      995 23121381        12
#> 13:    995    995    995    995    995      995      995 23121381        16
#> 14:      0      0      0      0      1        0        0 23122159         5
#> 15:    995    995    995    995    995      995      995 23122159         0
#> 16:    995    995    995    995    995      995      995 23122159         0
#> 17:    995    995    995    995    995      995      995 23122159         0
#> 18:    995    995    995    995    995      995      995 23122159         0
#> 19:    995    995    995    995    995      995      995 23122159         0
#> 20:      1      0      0      0      0        0        0 23150818         3
#> 21:    995    995    995    995    995      995      995 23150818         0
#> 22:    995    995    995    995    995      995      995 23150818         0
#> 23:    995    995    995    995    995      995      995 23150818         0
#> 24:    995    995    995    995    995      995      995 23150818         0
#> 25:    995    995    995    995    995      995      995 23150818         0
#> 26:      0      0      0      0      1        0        0 23233061         0
#> 27:      0      0      0      0      1        0        0 23233061         2
#> 28:      0      0      1      0      0        0        0 23233061         4
#> 29:    995    995    995    995    995      995      995 23233061         2
#> 30:    995    995    995    995    995      995      995 23233061         2
#> 31:    995    995    995    995    995      995      995 23233061         0
#> 32:      0      0      0      0      1        0        0 23233651         0
#> 33:      0      1      0      0      0        0        0 23233651         2
#> 34:      0      0      0      0      1        0        0 23233651         3
#> 35:      0      0      0      0      1        0        0 23233651         3
#> 36:    995    995    995    995    995      995      995 23233651         3
#> 37:    995    995    995    995    995      995      995 23233651         3
#> 38:    995    995    995    995    995      995      995 23233651         3
#> 39:    995    995    995    995    995      995      995 23233651         3
#> 40:      0      0      0      0      1        0        0 23240801         0
#> 41:    995    995    995    995    995      995      995 23240801         2
#> 42:      0      0      0      0      1        0        0 23240801         0
#> 43:    995    995    995    995    995      995      995 23240801         0
#> 44:    995    995    995    995    995      995      995 23240801         0
#> 45:    995    995    995    995    995      995      995 23240801         0
#> 46:      0      0      0      0      1        0        0 23411644         4
#> 47:      0      0      0      0      1        0        0 23411644         6
#> 48:    995    995    995    995    995      995      995 23411644         4
#> 49:    995    995    995    995    995      995      995 23411644         2
#> 50:    995    995    995    995    995      995      995 23411644         2
#> 51:    995    995    995    995    995      995      995 23411644         6
#> 52:      1      0      0      0      0        0        0 23416563         4
#> 53:      1      0      0      0      0        0        0 23416563         4
#> 54:    995    995    995    995    995      995      995 23416563         2
#> 55:    995    995    995    995    995      995      995 23416563         2
#> 56:    995    995    995    995    995      995      995 23416563         2
#> 57:    995    995    995    995    995      995      995 23416563         2
#>     race_1 race_2 race_3 race_4 race_5 race_997 race_999    hh_id num_trips
#>      person_id age gender employment education person_weight
#>  1: 2301602901   5      1          5         2           411
#>  2: 2301602902   6      2          1         2           578
#>  3: 2301602903   2      1        995       995            17
#>  4: 2301602904   2      1        995       995           490
#>  5: 2301602905   1      1        995       995           342
#>  6: 2301602906   7      1          1         2           754
#>  7: 2301602907   2      1        995       995           249
#>  8: 2312138101   6      1          5         6           186
#>  9: 2312138102   6      2          1         7           296
#> 10: 2312138103   3      1          2       995           369
#> 11: 2312138104   2      2        995       995           499
#> 12: 2312138105   2      1        995       995           226
#> 13: 2312138106   2      2        995       995           823
#> 14: 2312215901   5      1          1         3           859
#> 15: 2312215902   2    995        995       995           280
#> 16: 2312215903   2    995        995       995           685
#> 17: 2312215904   2    995        995       995           592
#> 18: 2312215905   5    995          1       995           240
#> 19: 2312215906   5    995          1       995           710
#> 20: 2315081801   4      2          8         7           231
#> 21: 2315081802   4    995          5       995           331
#> 22: 2315081803   4    995          5       995           214
#> 23: 2315081804   4    995          2       995           105
#> 24: 2315081805   4    995          2       995           317
#> 25: 2315081806   4    995          5       995           825
#> 26: 2323306101  10      1          5         6           114
#> 27: 2323306102   7      2          1         4           297
#> 28: 2323306103   6      1          1         2           861
#> 29: 2323306104   2      1        995       995           161
#> 30: 2323306105   2      2        995       995           817
#> 31: 2323306106  10    995          5       995           193
#> 32: 2323365101   7      1          1         3           207
#> 33: 2323365102   8      2          1         2           178
#> 34: 2323365103   5      2          6         2           211
#> 35: 2323365104   5      1          6         2           590
#> 36: 2323365105   2      2        995       995           626
#> 37: 2323365106   2      1        995       995           433
#> 38: 2323365107   2      1        995       995           447
#> 39: 2323365108   1      1        995       995           404
#> 40: 2324080101   8      2          1         6           801
#> 41: 2324080102   2    999        995       995           894
#> 42: 2324080103   7      1          5         3           330
#> 43: 2324080104   2    999        995       995           983
#> 44: 2324080105   2    999        995       995           327
#> 45: 2324080106   2    999        995       995            66
#> 46: 2341164401   6      1          1         5           950
#> 47: 2341164402   6      2          1         6           785
#> 48: 2341164403   2      1        995       995           203
#> 49: 2341164404   2      1        995       995           719
#> 50: 2341164405   2      2        995       995           991
#> 51: 2341164406   2      1        995       995           553
#> 52: 2341656301   6      1          1         5           812
#> 53: 2341656302   6      2          1         3           889
#> 54: 2341656303   2      1        995       995           258
#> 55: 2341656304   2      1        995       995            42
#> 56: 2341656305   2      2        995       995           955
#> 57: 2341656306   3      2          5       995           950
#>      person_id age gender employment education person_weight
#> 
#> $day
#>     delivery_2 delivery_3 delivery_4 delivery_5 delivery_6 delivery_7
#>  1:          0          0          0          0          0          0
#>  2:        995        995        995        995        995        995
#>  3:        995        995        995        995        995        995
#>  4:        995        995        995        995        995        995
#>  5:        995        995        995        995        995        995
#>  6:        995        995        995        995        995        995
#>  7:        995        995        995        995        995        995
#>  8:          0          0          0          0          0          0
#>  9:          0          0          0          0          0          0
#> 10:        995        995        995        995        995        995
#> 11:        995        995        995        995        995        995
#> 12:        995        995        995        995        995        995
#> 13:        995        995        995        995        995        995
#> 14:          0          0          0          0          0          0
#> 15:          0          0          0          0          0          0
#> 16:        995        995        995        995        995        995
#> 17:        995        995        995        995        995        995
#> 18:        995        995        995        995        995        995
#> 19:        995        995        995        995        995        995
#> 20:          0          0          0          0          0          0
#> 21:          0          0          0          0          0          0
#> 22:        995        995        995        995        995        995
#> 23:        995        995        995        995        995        995
#> 24:        995        995        995        995        995        995
#> 25:        995        995        995        995        995        995
#> 26:          0          0          0          1          0          0
#> 27:          0          0          0          0          0          0
#> 28:        995        995        995        995        995        995
#> 29:        995        995        995        995        995        995
#> 30:        995        995        995        995        995        995
#> 31:        995        995        995        995        995        995
#> 32:          0          0          0          1          0          0
#> 33:          0          0          0          0          0          0
#> 34:        995        995        995        995        995        995
#> 35:        995        995        995        995        995        995
#> 36:        995        995        995        995        995        995
#> 37:        995        995        995        995        995        995
#> 38:          0          0          0          1          0          0
#> 39:          0          0          0          1          0          0
#> 40:        995        995        995        995        995        995
#> 41:        995        995        995        995        995        995
#> 42:        995        995        995        995        995        995
#> 43:        995        995        995        995        995        995
#> 44:          0          0          0          0          0          0
#> 45:          0          0          0          0          0          0
#> 46:        995        995        995        995        995        995
#> 47:        995        995        995        995        995        995
#> 48:        995        995        995        995        995        995
#> 49:        995        995        995        995        995        995
#> 50:          0          0          0          1          0          0
#> 51:          0          0          0          1          0          0
#> 52:          0          0          0          0          0          0
#> 53:        995        995        995        995        995        995
#> 54:        995        995        995        995        995        995
#> 55:        995        995        995        995        995        995
#> 56:        995        995        995        995        995        995
#> 57:          0          0          0          0          0          0
#> 58:        995        995        995        995        995        995
#> 59:        995        995        995        995        995        995
#> 60:        995        995        995        995        995        995
#> 61:        995        995        995        995        995        995
#> 62:        995        995        995        995        995        995
#> 63:        995        995        995        995        995        995
#> 64:        995        995        995        995        995        995
#> 65:          0          0          0          0          0          0
#> 66:        995        995        995        995        995        995
#> 67:        995        995        995        995        995        995
#> 68:        995        995        995        995        995        995
#> 69:        995        995        995        995        995        995
#> 70:        995        995        995        995        995        995
#> 71:          0          0          0          1          0          0
#> 72:        995        995        995        995        995        995
#> 73:        995        995        995        995        995        995
#> 74:        995        995        995        995        995        995
#> 75:        995        995        995        995        995        995
#> 76:        995        995        995        995        995        995
#> 77:          0          0          0          0          0          0
#> 78:        995        995        995        995        995        995
#> 79:        995        995        995        995        995        995
#> 80:        995        995        995        995        995        995
#> 81:        995        995        995        995        995        995
#> 82:        995        995        995        995        995        995
#>     delivery_2 delivery_3 delivery_4 delivery_5 delivery_6 delivery_7
#>     delivery_8 delivery_996    hh_id num_trips  person_id       day_id
#>  1:          0            1 23016029         5 2301602901 230160290101
#>  2:        995          995 23016029         0 2301602902 230160290201
#>  3:        995          995 23016029         2 2301602903 230160290301
#>  4:        995          995 23016029         2 2301602904 230160290401
#>  5:        995          995 23016029         5 2301602905 230160290501
#>  6:        995          995 23016029         2 2301602906 230160290601
#>  7:        995          995 23016029         4 2301602907 230160290701
#>  8:          0            1 23121381        11 2312138101 231213810101
#>  9:          0            1 23121381         0 2312138102 231213810201
#> 10:        995          995 23121381        NA 2312138103 231213810301
#> 11:        995          995 23121381         3 2312138104 231213810401
#> 12:        995          995 23121381         4 2312138105 231213810501
#> 13:        995          995 23121381         3 2312138106 231213810601
#> 14:          0            1 23121381         5 2312138101 231213810102
#> 15:          0            1 23121381         0 2312138102 231213810202
#> 16:        995          995 23121381        NA 2312138103 231213810302
#> 17:        995          995 23121381        NA 2312138104 231213810402
#> 18:        995          995 23121381        NA 2312138105 231213810502
#> 19:        995          995 23121381        NA 2312138106 231213810602
#> 20:          0            1 23121381         2 2312138101 231213810103
#> 21:          0            1 23121381         4 2312138102 231213810203
#> 22:        995          995 23121381        NA 2312138103 231213810303
#> 23:        995          995 23121381         2 2312138104 231213810403
#> 24:        995          995 23121381         2 2312138105 231213810503
#> 25:        995          995 23121381         2 2312138106 231213810603
#> 26:          0            0 23121381         7 2312138101 231213810104
#> 27:          0            1 23121381         3 2312138102 231213810204
#> 28:        995          995 23121381        NA 2312138103 231213810304
#> 29:        995          995 23121381         4 2312138104 231213810404
#> 30:        995          995 23121381        NA 2312138105 231213810504
#> 31:        995          995 23121381         1 2312138106 231213810604
#> 32:          0            0 23121381        12 2312138101 231213810105
#> 33:          0            1 23121381         3 2312138102 231213810205
#> 34:        995          995 23121381         0 2312138103 231213810305
#> 35:        995          995 23121381         1 2312138104 231213810405
#> 36:        995          995 23121381         5 2312138105 231213810505
#> 37:        995          995 23121381         8 2312138106 231213810605
#> 38:          0            0 23121381         5 2312138101 231213810106
#> 39:          0            0 23121381         3 2312138102 231213810206
#> 40:        995          995 23121381        NA 2312138103 231213810306
#> 41:        995          995 23121381         1 2312138104 231213810406
#> 42:        995          995 23121381         1 2312138105 231213810506
#> 43:        995          995 23121381         1 2312138106 231213810606
#> 44:          0            1 23121381         3 2312138101 231213810107
#> 45:          0            1 23121381         2 2312138102 231213810207
#> 46:        995          995 23121381        NA 2312138103 231213810307
#> 47:        995          995 23121381         2 2312138104 231213810407
#> 48:        995          995 23121381        NA 2312138105 231213810507
#> 49:        995          995 23121381         1 2312138106 231213810607
#> 50:          0            0 23122159         5 2312215901 231221590101
#> 51:          0            0 23150818         3 2315081801 231508180101
#> 52:          0            1 23233061         0 2323306101 232330610101
#> 53:        995          995 23233061         2 2323306102 232330610201
#> 54:        995          995 23233061         4 2323306103 232330610301
#> 55:        995          995 23233061         2 2323306104 232330610401
#> 56:        995          995 23233061         2 2323306105 232330610501
#> 57:          0            1 23233651         0 2323365101 232336510101
#> 58:        995          995 23233651         2 2323365102 232336510201
#> 59:        995          995 23233651         3 2323365103 232336510301
#> 60:        995          995 23233651         3 2323365104 232336510401
#> 61:        995          995 23233651         3 2323365105 232336510501
#> 62:        995          995 23233651         3 2323365106 232336510601
#> 63:        995          995 23233651         3 2323365107 232336510701
#> 64:        995          995 23233651         3 2323365108 232336510801
#> 65:          0            1 23240801         0 2324080101 232408010101
#> 66:        995          995 23240801         2 2324080102 232408010201
#> 67:        995          995 23240801         0 2324080103 232408010301
#> 68:        995          995 23240801         0 2324080104 232408010401
#> 69:        995          995 23240801         0 2324080105 232408010501
#> 70:        995          995 23240801         0 2324080106 232408010601
#> 71:          0            0 23411644         4 2341164401 234116440101
#> 72:        995          995 23411644         6 2341164402 234116440201
#> 73:        995          995 23411644         4 2341164403 234116440301
#> 74:        995          995 23411644         2 2341164404 234116440401
#> 75:        995          995 23411644         2 2341164405 234116440501
#> 76:        995          995 23411644         6 2341164406 234116440601
#> 77:          0            1 23416563         4 2341656301 234165630101
#> 78:        995          995 23416563         4 2341656302 234165630201
#> 79:        995          995 23416563         2 2341656303 234165630301
#> 80:        995          995 23416563         2 2341656304 234165630401
#> 81:        995          995 23416563         2 2341656305 234165630501
#> 82:        995          995 23416563         2 2341656306 234165630601
#>     delivery_8 delivery_996    hh_id num_trips  person_id       day_id
#>     travel_date begin_day end_day day_weight
#>  1:  2023-04-27         1       1        504
#>  2:  2023-04-27         1       1        376
#>  3:  2023-04-27         1       1        115
#>  4:  2023-04-27         1       1        306
#>  5:  2023-04-27         1       1        358
#>  6:  2023-04-27         1       1        197
#>  7:  2023-04-27         1       1        546
#>  8:  2023-05-05         1       1        896
#>  9:  2023-05-05         1       1        662
#> 10:  2023-05-05       995     995        228
#> 11:  2023-05-05       995     995        580
#> 12:  2023-05-05       995     995        541
#> 13:  2023-05-05       995     995        909
#> 14:  2023-05-06         1       1        159
#> 15:  2023-05-06         1       1        464
#> 16:  2023-05-06       995     995        518
#> 17:  2023-05-06       995     995        838
#> 18:  2023-05-06       995     995        916
#> 19:  2023-05-06       995     995        368
#> 20:  2023-05-07         1       1        442
#> 21:  2023-05-07         1       1        362
#> 22:  2023-05-07       995     995        904
#> 23:  2023-05-07       995     995        243
#> 24:  2023-05-07       995     995        820
#> 25:  2023-05-07       995     995        827
#> 26:  2023-05-08         1       1        169
#> 27:  2023-05-08         1       1        169
#> 28:  2023-05-08       995     995        495
#> 29:  2023-05-08       995     995        813
#> 30:  2023-05-08       995     995        436
#> 31:  2023-05-08       995     995        177
#> 32:  2023-05-09         1       1        388
#> 33:  2023-05-09         1       1        688
#> 34:  2023-05-09         1       1        497
#> 35:  2023-05-09         1       1        934
#> 36:  2023-05-09         1       1        614
#> 37:  2023-05-09         1       1        414
#> 38:  2023-05-10         1       1        384
#> 39:  2023-05-10         1       1        620
#> 40:  2023-05-10         1     995        748
#> 41:  2023-05-10         1     995        695
#> 42:  2023-05-10         1     995        397
#> 43:  2023-05-10         1     995        544
#> 44:  2023-05-11         1       1         50
#> 45:  2023-05-11         1       1        385
#> 46:  2023-05-11       995     995        247
#> 47:  2023-05-11       995     995        740
#> 48:  2023-05-11       995     995        496
#> 49:  2023-05-11       995     995        713
#> 50:  2023-05-03         1       1        768
#> 51:  2023-05-24         1       1        659
#> 52:  2023-05-25         1       1        111
#> 53:  2023-05-25         1       1        695
#> 54:  2023-05-25         1       1        279
#> 55:  2023-05-25         1       1        661
#> 56:  2023-05-25         1       1        708
#> 57:  2023-05-29         1       1        631
#> 58:  2023-05-29         1       1        167
#> 59:  2023-05-29         1       1        641
#> 60:  2023-05-29         1       1        574
#> 61:  2023-05-29         1       1        543
#> 62:  2023-05-29         1       1         96
#> 63:  2023-05-29         1       1        982
#> 64:  2023-05-29         1       1        288
#> 65:  2023-05-25         1       1         66
#> 66:  2023-05-25         1       1        451
#> 67:  2023-05-25         1       1        711
#> 68:  2023-05-25         1       1        621
#> 69:  2023-05-25         1       1        466
#> 70:  2023-05-25         1       1        209
#> 71:  2023-05-31         1       3        923
#> 72:  2023-05-31         1       1        435
#> 73:  2023-05-31         1       1        468
#> 74:  2023-05-31         1       1        966
#> 75:  2023-05-31         1       1        350
#> 76:  2023-05-31         1       1        160
#> 77:  2023-06-05         1       1        438
#> 78:  2023-06-05         1       1        922
#> 79:  2023-06-05         1       1        554
#> 80:  2023-06-05         1       1        451
#> 81:  2023-06-05         1       1        971
#> 82:  2023-06-05         1       1         47
#>     travel_date begin_day end_day day_weight
#> 
#> $trip
#>         hh_id  person_id       day_id travel_date      trip_id  speed_mph
#>   1: 23016029 2301602901 230160290101  2023-04-27 2.301603e+12  5.9952706
#>   2: 23016029 2301602901 230160290101  2023-04-27 2.301603e+12 21.5112706
#>   3: 23016029 2301602901 230160290101  2023-04-27 2.301603e+12  3.5931191
#>   4: 23016029 2301602901 230160290101  2023-04-27 2.301603e+12 11.9905413
#>   5: 23016029 2301602901 230160290101  2023-04-27 2.301603e+12  5.9952706
#>  ---                                                                     
#> 197: 23416563 2341656304 234165630401  2023-06-05 2.341656e+12  2.4496880
#> 198: 23416563 2341656305 234165630501  2023-06-05 2.341656e+12  1.5417278
#> 199: 23416563 2341656305 234165630501  2023-06-05 2.341656e+12  2.5695464
#> 200: 23416563 2341656306 234165630601  2023-06-05 2.341656e+12  0.8465739
#> 201: 23416563 2341656306 234165630601  2023-06-05 2.341656e+12  4.2869540
#>      mode_type d_purpose_category trip_weight
#>   1:         8                  6         804
#>   2:         8                 10         523
#>   3:         8                  1         622
#>   4:         8                  6         633
#>   5:         8                  1         413
#>  ---                                         
#> 197:        10                  1         649
#> 198:         1                  4         503
#> 199:         1                  1         338
#> 200:         8                  6         302
#> 201:         8                  1         599
#> 
#> $vehicle
#>        hh_id vehicle_id fuel_type hh_weight
#>  1: 23016029 2301602901         1       974
#>  2: 23016029 2301602902         1       974
#>  3: 23016029 2301602903         1       974
#>  4: 23121381 2312138101         1       826
#>  5: 23121381 2312138102         1       826
#>  6: 23121381 2312138103         1       826
#>  7: 23122159 2312215901         1       938
#>  8: 23122159 2312215902         1       938
#>  9: 23122159 2312215903         1       938
#> 10: 23150818 2315081801         1       602
#> 11: 23150818 2315081802         1       602
#> 12: 23150818 2315081803         4       602
#> 13: 23233061 2323306101         1        39
#> 14: 23233061 2323306102         1        39
#> 15: 23233061 2323306103         5        39
#> 16: 23233061 2323306104         1        39
#> 17: 23233061 2323306105         1        39
#> 18: 23233651 2323365101         1       345
#> 19: 23233651 2323365102         1       345
#> 20: 23233651 2323365103         1       345
#> 21: 23240801 2324080101         1       783
#> 22: 23240801 2324080102         1       783
#> 23: 23240801 2324080103         1       783
#> 24: 23411644 2341164401         1       154
#> 25: 23411644 2341164402         1       154
#> 26: 23411644 2341164403         5       154
#> 27: 23416563 2341656301         1       928
#> 28: 23416563 2341656302         1       928
#>        hh_id vehicle_id fuel_type hh_weight
#>