Filter datasets to only keep specified ids

Filter datasets to only keep specified ids

Usage

hts_filter_data(hts_data, ids, id_name)

Arguments

hts_data

List of containing household travel data tables.

ids

List of ids to keep in all of the tables

id_name

Name of id being used for filtering (e.g., hh_id, person_id)

Value

Inputted list of tables filtered to the specified ids.

Examples

require(data.table)

data(test_data)

hts_filter_data(hts_data = test_data,
                ids = hh[num_people > 5, hh_id],
                id_name = 'hh_id')
#> $hh
#>       hh_id sample_segment num_trips num_people income_detailed income_followup
#>       <int>          <int>     <int>      <int>           <int>           <int>
#> 1: 23001496             16        19          6              10             995
#> 2: 23016029              8        20          7               5             995
#> 3: 23049474             12        12          6               8             995
#> 4: 23094796              6        24          6              10             995
#> 5: 23151045             19         4          7               9             995
#> 6: 23191358              7         8          6               5             995
#> 7: 23233651             21        20          8               6             995
#> 8: 23275913              4        12          6               7             995
#> 9: 23416563             14        16          6               7             995
#>     home_lon home_lat home_county residence_type hh_weight
#>        <num>    <num>       <int>          <int>     <int>
#> 1: -87.82783 40.00000           3              1       348
#> 2: -85.60561 35.55055           2              2       140
#> 3: -96.09610 34.48549           1              1       420
#> 4: -87.22723 34.56256           1              1       394
#> 5: -88.42843 35.03904           1              3       279
#> 6: -90.39039 36.67167           1              4       368
#> 7: -81.56156 34.94094           2              1        50
#> 8: -87.02703 35.86587           3              1       521
#> 9: -88.36837 39.86687           1              2        94
#> 
#> $person
#>     ethnicity_1 ethnicity_2 ethnicity_3 ethnicity_4 ethnicity_997 ethnicity_999
#>           <int>       <int>       <int>       <int>         <int>         <int>
#>  1:           1           0           0           0             0             0
#>  2:           1           0           0           0             0             0
#>  3:           1           0           0           0             0             0
#>  4:         995         995         995         995           995           995
#>  5:         995         995         995         995           995           995
#>  6:         995         995         995         995           995           995
#>  7:           1           0           0           0             0             0
#>  8:           1           0           0           0             0             0
#>  9:         995         995         995         995           995           995
#> 10:         995         995         995         995           995           995
#> 11:         995         995         995         995           995           995
#> 12:           1           0           0           0             0             0
#> 13:         995         995         995         995           995           995
#> 14:           1           0           0           0             0             0
#> 15:           0           0           0           0             0             1
#> 16:           0           0           0           0             0             1
#> 17:           0           0           0           0             0             1
#> 18:           0           0           0           0             0             1
#> 19:           0           0           0           0             0             1
#> 20:           1           0           0           0             0             0
#> 21:           1           0           0           0             0             0
#> 22:         995         995         995         995           995           995
#> 23:         995         995         995         995           995           995
#> 24:         995         995         995         995           995           995
#> 25:         995         995         995         995           995           995
#> 26:           1           0           0           0             0             0
#> 27:           1           0           0           0             0             0
#> 28:           1           0           0           0             0             0
#> 29:           1           0           0           0             0             0
#> 30:           1           0           0           0             0             0
#> 31:           1           0           0           0             0             0
#> 32:           1           0           0           0             0             0
#> 33:           0           1           0           0             0             0
#> 34:           0           1           0           0             0             0
#> 35:         995         995         995         995           995           995
#> 36:         995         995         995         995           995           995
#> 37:         995         995         995         995           995           995
#> 38:           0           1           0           0             0             0
#> 39:           1           0           0           0             0             0
#> 40:           1           0           0           0             0             0
#> 41:           1           0           0           0             0             0
#> 42:           1           0           0           0             0             0
#> 43:         995         995         995         995           995           995
#> 44:         995         995         995         995           995           995
#> 45:         995         995         995         995           995           995
#> 46:         995         995         995         995           995           995
#> 47:           1           0           0           0             0             0
#> 48:           1           0           0           0             0             0
#> 49:         995         995         995         995           995           995
#> 50:         995         995         995         995           995           995
#> 51:         995         995         995         995           995           995
#> 52:           1           0           0           0             0             0
#> 53:           1           0           0           0             0             0
#> 54:           1           0           0           0             0             0
#> 55:         995         995         995         995           995           995
#> 56:         995         995         995         995           995           995
#> 57:         995         995         995         995           995           995
#> 58:         995         995         995         995           995           995
#>     ethnicity_1 ethnicity_2 ethnicity_3 ethnicity_4 ethnicity_997 ethnicity_999
#>     race_1 race_2 race_3 race_4 race_5 race_997 race_999    hh_id num_trips
#>      <int>  <int>  <int>  <int>  <int>    <int>    <int>    <int>     <int>
#>  1:      0      0      0      0      1        0        0 23001496         9
#>  2:      0      0      0      0      1        0        0 23001496         2
#>  3:      0      0      1      0      0        0        0 23001496         2
#>  4:    995    995    995    995    995      995      995 23001496         2
#>  5:    995    995    995    995    995      995      995 23001496         2
#>  6:    995    995    995    995    995      995      995 23001496         2
#>  7:      0      0      0      0      1        0        0 23016029         5
#>  8:      0      0      0      0      1        0        0 23016029         0
#>  9:    995    995    995    995    995      995      995 23016029         2
#> 10:    995    995    995    995    995      995      995 23016029         2
#> 11:    995    995    995    995    995      995      995 23016029         5
#> 12:      0      0      0      0      1        0        0 23016029         2
#> 13:    995    995    995    995    995      995      995 23016029         4
#> 14:      0      0      1      1      1        0        0 23049474         2
#> 15:      0      0      0      0      0        0        1 23049474         2
#> 16:      0      0      0      0      0        0        1 23049474         2
#> 17:      0      0      0      0      0        0        1 23049474         2
#> 18:      0      0      0      0      0        0        1 23049474         2
#> 19:      0      0      0      0      0        0        1 23049474         2
#> 20:      0      0      0      0      1        0        0 23094796         2
#> 21:      0      0      0      0      1        0        0 23094796        12
#> 22:    995    995    995    995    995      995      995 23094796         2
#> 23:    995    995    995    995    995      995      995 23094796         2
#> 24:    995    995    995    995    995      995      995 23094796         2
#> 25:    995    995    995    995    995      995      995 23094796         4
#> 26:      0      0      1      0      1        0        0 23151045         1
#> 27:      0      0      1      0      1        0        0 23151045         1
#> 28:      0      0      1      0      1        0        0 23151045         1
#> 29:      0      0      0      0      1        0        0 23151045         1
#> 30:      0      0      0      0      1        0        0 23151045         0
#> 31:      0      0      1      0      0        0        0 23151045         0
#> 32:      0      0      1      0      1        0        0 23151045         0
#> 33:      0      0      0      0      1        0        0 23191358         0
#> 34:      0      0      0      0      1        0        0 23191358         4
#> 35:    995    995    995    995    995      995      995 23191358         2
#> 36:    995    995    995    995    995      995      995 23191358         0
#> 37:    995    995    995    995    995      995      995 23191358         0
#> 38:      0      0      0      0      1        0        0 23191358         2
#> 39:      0      0      0      0      1        0        0 23233651         0
#> 40:      0      1      0      0      0        0        0 23233651         2
#> 41:      0      0      0      0      1        0        0 23233651         3
#> 42:      0      0      0      0      1        0        0 23233651         3
#> 43:    995    995    995    995    995      995      995 23233651         3
#> 44:    995    995    995    995    995      995      995 23233651         3
#> 45:    995    995    995    995    995      995      995 23233651         3
#> 46:    995    995    995    995    995      995      995 23233651         3
#> 47:      0      0      1      0      0        0        0 23275913         4
#> 48:      0      0      1      0      0        0        0 23275913         0
#> 49:    995    995    995    995    995      995      995 23275913         2
#> 50:    995    995    995    995    995      995      995 23275913         2
#> 51:    995    995    995    995    995      995      995 23275913         2
#> 52:      0      0      1      0      0        0        0 23275913         2
#> 53:      1      0      0      0      0        0        0 23416563         4
#> 54:      1      0      0      0      0        0        0 23416563         4
#> 55:    995    995    995    995    995      995      995 23416563         2
#> 56:    995    995    995    995    995      995      995 23416563         2
#> 57:    995    995    995    995    995      995      995 23416563         2
#> 58:    995    995    995    995    995      995      995 23416563         2
#>     race_1 race_2 race_3 race_4 race_5 race_997 race_999    hh_id num_trips
#>      person_id   age gender employment job_type education person_weight
#>          <num> <int>  <int>      <int>    <int>     <int>         <int>
#>  1: 2300149601     6      2          1        5         5           573
#>  2: 2300149602     6      1          1        3         6           585
#>  3: 2300149603     4      2          1        1         2           973
#>  4: 2300149604     2      2        995      995       995           882
#>  5: 2300149605     2      1        995      995       995           128
#>  6: 2300149606     2      1        995      995       995           230
#>  7: 2301602901     5      1          5      995         2           446
#>  8: 2301602902     6      2          1        1         2           296
#>  9: 2301602903     2      1        995      995       995           493
#> 10: 2301602904     2      1        995      995       995           983
#> 11: 2301602905     1      1        995      995       995           298
#> 12: 2301602906     7      1          1        1         2           167
#> 13: 2301602907     2      1        995      995       995           250
#> 14: 2304947401     5      2          1        3         6           341
#> 15: 2304947402     8      1          1        1       999           336
#> 16: 2304947403     4      2          1        1       999           224
#> 17: 2304947404     8      2          1        2       999           309
#> 18: 2304947405     4      2          1        1       999           853
#> 19: 2304947406    10      2          5      995       999           654
#> 20: 2309479601     7      2          1        1         6           446
#> 21: 2309479602     7      1          5      995         7           293
#> 22: 2309479603     2    999        995      995       995           899
#> 23: 2309479604     2    999        995      995       995           812
#> 24: 2309479605     2    999        995      995       995           545
#> 25: 2309479606     2    999        995      995       995           791
#> 26: 2315104501     4      2          7        3         3            25
#> 27: 2315104502     4      1          1        1         6           528
#> 28: 2315104503     4      1          2        1         6           534
#> 29: 2315104504     7      1          8      995         2            17
#> 30: 2315104505     7      2          1        2         7           469
#> 31: 2315104506    10      2          5      995         6           478
#> 32: 2315104507    10      1          5      995         7           420
#> 33: 2319135801     5      1          5      995         2           872
#> 34: 2319135802     7      2          1        2         2           823
#> 35: 2319135803     2    999        995      995       995           101
#> 36: 2319135804     2    999        995      995       995           443
#> 37: 2319135805     1    999        995      995       995           878
#> 38: 2319135806     8      1          2        1         1           492
#> 39: 2323365101     7      1          1        1         3           167
#> 40: 2323365102     8      2          1        1         2           979
#> 41: 2323365103     5      2          6      995         2           749
#> 42: 2323365104     5      1          6      995         2           851
#> 43: 2323365105     2      2        995      995       995           759
#> 44: 2323365106     2      1        995      995       995           153
#> 45: 2323365107     2      1        995      995       995           601
#> 46: 2323365108     1      1        995      995       995           927
#> 47: 2327591301     7      2          1        1         7           755
#> 48: 2327591302     9      1          5      995         2           632
#> 49: 2327591303     2      2        995      995       995           866
#> 50: 2327591304     2      2        995      995       995           967
#> 51: 2327591305     1      1        995      995       995           177
#> 52: 2327591306     6      1          1        4         2           548
#> 53: 2341656301     6      1          1        1         5            72
#> 54: 2341656302     6      2          1        1         3           543
#> 55: 2341656303     2      1        995      995       995           676
#> 56: 2341656304     2      1        995      995       995           692
#> 57: 2341656305     2      2        995      995       995           150
#> 58: 2341656306     3      2          5      995       995           800
#>      person_id   age gender employment job_type education person_weight
#> 
#> $day
#>     delivery_2 delivery_3 delivery_4 delivery_5 delivery_6 delivery_7
#>          <int>      <int>      <int>      <int>      <int>      <int>
#>  1:          0          0          0          1          0          0
#>  2:        995        995        995        995        995        995
#>  3:        995        995        995        995        995        995
#>  4:        995        995        995        995        995        995
#>  5:        995        995        995        995        995        995
#>  6:        995        995        995        995        995        995
#>  7:          0          0          0          0          0          0
#>  8:        995        995        995        995        995        995
#>  9:        995        995        995        995        995        995
#> 10:        995        995        995        995        995        995
#> 11:        995        995        995        995        995        995
#> 12:        995        995        995        995        995        995
#> 13:        995        995        995        995        995        995
#> 14:          0          0          0          0          0          0
#> 15:        995        995        995        995        995        995
#> 16:        995        995        995        995        995        995
#> 17:        995        995        995        995        995        995
#> 18:        995        995        995        995        995        995
#> 19:        995        995        995        995        995        995
#> 20:          0          0          0          1          0          0
#> 21:        995        995        995        995        995        995
#> 22:        995        995        995        995        995        995
#> 23:        995        995        995        995        995        995
#> 24:        995        995        995        995        995        995
#> 25:        995        995        995        995        995        995
#> 26:          0          0          0          1          0          0
#> 27:        995        995        995        995        995        995
#> 28:        995        995        995        995        995        995
#> 29:        995        995        995        995        995        995
#> 30:        995        995        995        995        995        995
#> 31:        995        995        995        995        995        995
#> 32:        995        995        995        995        995        995
#> 33:          0          0          0          1          0          0
#> 34:        995        995        995        995        995        995
#> 35:        995        995        995        995        995        995
#> 36:        995        995        995        995        995        995
#> 37:        995        995        995        995        995        995
#> 38:        995        995        995        995        995        995
#> 39:          0          0          0          0          0          0
#> 40:        995        995        995        995        995        995
#> 41:        995        995        995        995        995        995
#> 42:        995        995        995        995        995        995
#> 43:        995        995        995        995        995        995
#> 44:        995        995        995        995        995        995
#> 45:        995        995        995        995        995        995
#> 46:        995        995        995        995        995        995
#> 47:          0          0          0          0          0          0
#> 48:        995        995        995        995        995        995
#> 49:        995        995        995        995        995        995
#> 50:        995        995        995        995        995        995
#> 51:        995        995        995        995        995        995
#> 52:        995        995        995        995        995        995
#> 53:          0          0          0          0          0          0
#> 54:        995        995        995        995        995        995
#> 55:        995        995        995        995        995        995
#> 56:        995        995        995        995        995        995
#> 57:        995        995        995        995        995        995
#> 58:        995        995        995        995        995        995
#>     delivery_2 delivery_3 delivery_4 delivery_5 delivery_6 delivery_7
#>     delivery_8 delivery_996    hh_id num_trips  person_id       day_id
#>          <int>        <int>    <int>     <int>      <num>        <num>
#>  1:          0            0 23001496         9 2300149601 230014960101
#>  2:        995          995 23001496         2 2300149602 230014960201
#>  3:        995          995 23001496         2 2300149603 230014960301
#>  4:        995          995 23001496         2 2300149604 230014960401
#>  5:        995          995 23001496         2 2300149605 230014960501
#>  6:        995          995 23001496         2 2300149606 230014960601
#>  7:          0            1 23016029         5 2301602901 230160290101
#>  8:        995          995 23016029         0 2301602902 230160290201
#>  9:        995          995 23016029         2 2301602903 230160290301
#> 10:        995          995 23016029         2 2301602904 230160290401
#> 11:        995          995 23016029         5 2301602905 230160290501
#> 12:        995          995 23016029         2 2301602906 230160290601
#> 13:        995          995 23016029         4 2301602907 230160290701
#> 14:          0            1 23049474         2 2304947401 230494740101
#> 15:        995          995 23049474         2 2304947402 230494740201
#> 16:        995          995 23049474         2 2304947403 230494740301
#> 17:        995          995 23049474         2 2304947404 230494740401
#> 18:        995          995 23049474         2 2304947405 230494740501
#> 19:        995          995 23049474         2 2304947406 230494740601
#> 20:          0            0 23094796         2 2309479601 230947960101
#> 21:        995          995 23094796        12 2309479602 230947960201
#> 22:        995          995 23094796         2 2309479603 230947960301
#> 23:        995          995 23094796         2 2309479604 230947960401
#> 24:        995          995 23094796         2 2309479605 230947960501
#> 25:        995          995 23094796         4 2309479606 230947960601
#> 26:          0            0 23151045         1 2315104501 231510450101
#> 27:        995          995 23151045         1 2315104502 231510450201
#> 28:        995          995 23151045         1 2315104503 231510450301
#> 29:        995          995 23151045         1 2315104504 231510450401
#> 30:        995          995 23151045         0 2315104505 231510450501
#> 31:        995          995 23151045         0 2315104506 231510450601
#> 32:        995          995 23151045         0 2315104507 231510450701
#> 33:          0            0 23191358         0 2319135801 231913580101
#> 34:        995          995 23191358         4 2319135802 231913580201
#> 35:        995          995 23191358         2 2319135803 231913580301
#> 36:        995          995 23191358         0 2319135804 231913580401
#> 37:        995          995 23191358         0 2319135805 231913580501
#> 38:        995          995 23191358         2 2319135806 231913580601
#> 39:          0            1 23233651         0 2323365101 232336510101
#> 40:        995          995 23233651         2 2323365102 232336510201
#> 41:        995          995 23233651         3 2323365103 232336510301
#> 42:        995          995 23233651         3 2323365104 232336510401
#> 43:        995          995 23233651         3 2323365105 232336510501
#> 44:        995          995 23233651         3 2323365106 232336510601
#> 45:        995          995 23233651         3 2323365107 232336510701
#> 46:        995          995 23233651         3 2323365108 232336510801
#> 47:          0            1 23275913         4 2327591301 232759130101
#> 48:        995          995 23275913         0 2327591302 232759130201
#> 49:        995          995 23275913         2 2327591303 232759130301
#> 50:        995          995 23275913         2 2327591304 232759130401
#> 51:        995          995 23275913         2 2327591305 232759130501
#> 52:        995          995 23275913         2 2327591306 232759130601
#> 53:          0            1 23416563         4 2341656301 234165630101
#> 54:        995          995 23416563         4 2341656302 234165630201
#> 55:        995          995 23416563         2 2341656303 234165630301
#> 56:        995          995 23416563         2 2341656304 234165630401
#> 57:        995          995 23416563         2 2341656305 234165630501
#> 58:        995          995 23416563         2 2341656306 234165630601
#>     delivery_8 delivery_996    hh_id num_trips  person_id       day_id
#>     travel_date begin_day end_day day_weight
#>          <Date>     <int>   <int>      <int>
#>  1:  2023-04-12         1       1        857
#>  2:  2023-04-12         7       7        158
#>  3:  2023-04-12         1       1         83
#>  4:  2023-04-12         2       2        293
#>  5:  2023-04-12         4       4        142
#>  6:  2023-04-12         7       7         30
#>  7:  2023-04-27         1       1        265
#>  8:  2023-04-27         1       1        782
#>  9:  2023-04-27         1       1         68
#> 10:  2023-04-27         1       1         83
#> 11:  2023-04-27         1       1        111
#> 12:  2023-04-27         1       1        997
#> 13:  2023-04-27         1       1        164
#> 14:  2023-04-20         1       1        923
#> 15:  2023-04-20         1       1        153
#> 16:  2023-04-20         1       1        658
#> 17:  2023-04-20         1       1        658
#> 18:  2023-04-20         1       1        946
#> 19:  2023-04-20         1       1         54
#> 20:  2023-04-19         1       1        617
#> 21:  2023-04-19         1       1        642
#> 22:  2023-04-19         1       1         36
#> 23:  2023-04-19         1       1        204
#> 24:  2023-04-19         1       1        778
#> 25:  2023-04-19         1       1        206
#> 26:  2023-06-01         1     997        195
#> 27:  2023-06-01         3     997        432
#> 28:  2023-06-01         1       3         82
#> 29:  2023-06-01         3     997        826
#> 30:  2023-06-01         1       1        339
#> 31:  2023-06-01         1       1        681
#> 32:  2023-06-01         1       1        791
#> 33:  2023-05-24         1       1        662
#> 34:  2023-05-24         1       1        428
#> 35:  2023-05-24         1       1        157
#> 36:  2023-05-24         1       1        608
#> 37:  2023-05-24         1       1        596
#> 38:  2023-05-24         2       2        591
#> 39:  2023-05-29         1       1        523
#> 40:  2023-05-29         1       1         26
#> 41:  2023-05-29         1       1        578
#> 42:  2023-05-29         1       1        146
#> 43:  2023-05-29         1       1        741
#> 44:  2023-05-29         1       1        348
#> 45:  2023-05-29         1       1        831
#> 46:  2023-05-29         1       1        761
#> 47:  2023-06-08         1       1        687
#> 48:  2023-06-08         1       1        961
#> 49:  2023-06-08         1       1        333
#> 50:  2023-06-08         1       1        967
#> 51:  2023-06-08         1       1         22
#> 52:  2023-06-08         1       1         32
#> 53:  2023-06-05         1       1        756
#> 54:  2023-06-05         1       1        675
#> 55:  2023-06-05         1       1        927
#> 56:  2023-06-05         1       1         96
#> 57:  2023-06-05         1       1        335
#> 58:  2023-06-05         1       1        583
#>     travel_date begin_day end_day day_weight
#> 
#> $trip
#>         hh_id  person_id       day_id travel_date      trip_id  speed_mph
#>         <int>      <num>        <num>      <Date>        <num>      <num>
#>   1: 23001496 2300149601 230014960101  2023-04-12 2.300150e+12         NA
#>   2: 23001496 2300149601 230014960101  2023-04-12 2.300150e+12 12.3087979
#>   3: 23001496 2300149601 230014960101  2023-04-12 2.300150e+12         NA
#>   4: 23001496 2300149601 230014960101  2023-04-12 2.300150e+12 69.5804961
#>   5: 23001496 2300149601 230014960101  2023-04-12 2.300150e+12         NA
#>  ---                                                                     
#> 131: 23416563 2341656304 234165630401  2023-06-05 2.341656e+12  2.4496880
#> 132: 23416563 2341656305 234165630501  2023-06-05 2.341656e+12  1.5417278
#> 133: 23416563 2341656305 234165630501  2023-06-05 2.341656e+12  2.5695464
#> 134: 23416563 2341656306 234165630601  2023-06-05 2.341656e+12  0.8465739
#> 135: 23416563 2341656306 234165630601  2023-06-05 2.341656e+12  4.2869540
#>      distance_miles mode_type mode_1 mode_2 num_travelers d_purpose_category
#>               <num>     <int>  <int>  <int>         <int>              <int>
#>   1:             NA         1     23    995             1                 11
#>   2:      6.1543989        13     23    995             1                 11
#>   3:             NA         1     23    995             1                 10
#>   4:     23.1934987         1      1    995             1                  2
#>   5:             NA         1     23    995             1                 11
#>  ---                                                                        
#> 131:      1.4289847        10     24    995             2                  1
#> 132:      0.2569546         1      1    995             1                  4
#> 133:      0.2569546         1      1    995             1                  1
#> 134:      4.7972523         8      6    995             3                  6
#> 135:      1.4289847         8      6    995             3                  1
#>      trip_weight
#>            <int>
#>   1:         352
#>   2:         773
#>   3:         219
#>   4:         712
#>   5:         609
#>  ---            
#> 131:         478
#> 132:         220
#> 133:         752
#> 134:         788
#> 135:         655
#> 
#> $vehicle
#> Key: <hh_id>
#>        hh_id vehicle_id fuel_type hh_weight
#>        <int>      <num>     <int>     <int>
#>  1: 23001496 2300149601         1       348
#>  2: 23001496 2300149602         1       348
#>  3: 23001496 2300149603         2       348
#>  4: 23016029 2301602901         1       140
#>  5: 23016029 2301602902         1       140
#>  6: 23016029 2301602903         1       140
#>  7: 23049474 2304947401         1       420
#>  8: 23049474 2304947402         1       420
#>  9: 23094796 2309479601         1       394
#> 10: 23094796 2309479602         1       394
#> 11: 23151045 2315104501         1       279
#> 12: 23151045 2315104502         1       279
#> 13: 23151045 2315104503         1       279
#> 14: 23191358 2319135801         1       368
#> 15: 23191358 2319135802         2       368
#> 16: 23191358 2319135803         1       368
#> 17: 23233651 2323365101         1        50
#> 18: 23233651 2323365102         1        50
#> 19: 23233651 2323365103         1        50
#> 20: 23275913 2327591301         1       521
#> 21: 23275913 2327591302         1       521
#> 22: 23416563 2341656301         1        94
#> 23: 23416563 2341656302         1        94
#>        hh_id vehicle_id fuel_type hh_weight
#>