Melts checkbox variables into a single variable

Melts checkbox variables into a single variable

Usage

hts_melt_vars(
  shared_name = NULL,
  wide_dt = NULL,
  shared_name_vars = NULL,
  variables_dt = variable_list,
  data = hts_data,
  ids = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id"),
  remove_missing = TRUE,
  missing_values = c("Missing Response", "995"),
  checkbox_label_sep = ":",
  to_single_row = FALSE
)

Arguments

shared_name

Shared name of the checkbox variable to melt (e.g., 'race_1', and 'race_2' have shared name of 'race'). Defaults to NULL.

wide_dt

Table containing the checkbox variables to melt in data.table format.

shared_name_vars

List of the checkbox variables with the same shared name. Defaults to NULL.

variables_dt

List of variable locations and descriptions in data.table format.

data

List containing household, person, day, trip, and vehicle datasets in data.table format.

ids

unique identifiers appearing in wide_dt

remove_missing

Boolean to remove rows with missing values. Defaults to TRUE.

missing_values

Missing values to remove. Defaults to 'Missing Response' and 995.

checkbox_label_sep

Character to use to split description of checkbox variable in the variable list. Defaults to ':'

to_single_row

Boolean if treating multiple checkbox selections as 'Two or more'. Defaults to FALSE.

Value

Inputted data table with checkbox variables melted into a single variable, with a 'variable' column to indicate original variable names and a 'value' column with the original value of the checkbox variable.

Examples

require(data.table)
require(stringr)
hts_melt_vars(shared_name = 'race',
              wide_dt = person,
              data = list('hh' = hh,
                              'person' = person,
                              'day' = day,
                              'trip' = trip,
                              'vehicle' = vehicle))
#>           hh_id  person_id variable value                      race
#>     1: 23000173 2300017301   race_1     0 African American or Black
#>     2: 23000173 2300017304   race_1     0 African American or Black
#>     3: 23000405 2300040501   race_1     0 African American or Black
#>     4: 23000405 2300040502   race_1     0 African American or Black
#>     5: 23000466 2300046601   race_1     0 African American or Black
#>    ---                                                             
#> 11378: 23417656 2341765602 race_999     0      Prefer not to answer
#> 11379: 23418585 2341858501 race_999     0      Prefer not to answer
#> 11380: 23418585 2341858502 race_999     0      Prefer not to answer
#> 11381: 23422311 2342231101 race_999     1      Prefer not to answer
#> 11382: 23422311 2342231102 race_999     1      Prefer not to answer