Prepare datasets to make summaries

Prepare datasets to make summaries

hts_prep_data(
  summarize_var = NULL,
  summarize_by = NULL,
  variables_dt = variable_list,
  data = hts_data,
  weighted = TRUE,
  remove_outliers = TRUE,
  threshold = 0.975,
  remove_missing = TRUE,
  missing_values = c("Missing Response", "995"),
  not_imputable = -1,
  strataname = NULL
)

Arguments

summarize_var

Name of the variable to summarize. Default is NULL

summarize_by

Name of the variable to summarize the summarize_var by. Default is NULL.

variables_dt

List of variable locations and descriptions in data.table format.

data

List of household, person, vehicle, day, and trip tables in data.table format.

weighted

Whether the data is weighted. Default is TRUE.

remove_outliers

Whether to remove outliers for numeric variable. Default is TRUE.

threshold

Threshold to define outliers. Default is 0.975.

remove_missing

Whether to remove missing values from the summary. Default is TRUE.

missing_values

Missing values to remove. Default is 995.

not_imputable

Value representing 'Not imputable' to remove. Default is -1.

strataname

Name of strata name to bring in. Default is NULL.

Value

List containing the categorical and numeric datasets of the summary variables and key columns, and either whether the summarize variable is shared or a breakdown of outliers, depending on if the summarize variable is categorical or numeric.

Examples

require(data.table)
require(stringr)
hts_prep_data(summarize_var = 'age',
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> Creating a summary of person age 
#> $cat
#>          hh_id  person_id person_weight age
#>    1: 23000173 2300017301           139   7
#>    2: 23000173 2300017302            17   3
#>    3: 23000173 2300017303           400   2
#>    4: 23000173 2300017304           747   6
#>    5: 23000738 2300073801           298   8
#>   ---                                      
#> 2038: 23418712 2341871203           809   2
#> 2039: 23419141 2341914101           136   6
#> 2040: 23419141 2341914102           170   5
#> 2041: 23419141 2341914103           490   1
#> 2042: 23419703 2341970301           645   9
#> 
#> $num
#> NULL
#> 
#> $var_is_shared
#> [1] FALSE
#> 
hts_prep_data(summarize_var = 'speed_mph',
              summarize_by = 'age',
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> Creating a summary of trip speed_mph broken down by person age
#> $cat
#>           hh_id  person_id       day_id      trip_id trip_weight  speed_mph
#>     1: 23000173 2300017301 230001730102 2.300017e+12         356      17-25
#>     2: 23000173 2300017301 230001730102 2.300017e+12         679      17-25
#>     3: 23000173 2300017301 230001730102 2.300017e+12         758        1-9
#>     4: 23000173 2300017301 230001730102 2.300017e+12         314        1-9
#>     5: 23000173 2300017301 230001730103 2.300017e+12         883      33-41
#>    ---                                                                     
#> 15031: 23419141 2341914103 234191410301 2.341914e+12         950        1-9
#> 15032: 23419141 2341914103 234191410301 2.341914e+12         123 43 or more
#> 15033: 23419141 2341914103 234191410301 2.341914e+12         720      25-33
#> 15034: 23419703 2341970301 234197030101 2.341970e+12         523       9-17
#> 15035: 23419703 2341970301 234197030101 2.341970e+12         298       9-17
#>        person_weight age
#>     1:           139   7
#>     2:           139   7
#>     3:           139   7
#>     4:           139   7
#>     5:           139   7
#>    ---                  
#> 15031:           490   1
#> 15032:           490   1
#> 15033:           490   1
#> 15034:           645   9
#> 15035:           645   9
#> 
#> $num
#>           hh_id  person_id       day_id      trip_id trip_weight speed_mph
#>     1: 23000173 2300017301 230001730102 2.300017e+12         356 21.480572
#>     2: 23000173 2300017301 230001730102 2.300017e+12         679 23.720178
#>     3: 23000173 2300017301 230001730102 2.300017e+12         758  6.927242
#>     4: 23000173 2300017301 230001730102 2.300017e+12         314  6.146634
#>     5: 23000173 2300017301 230001730103 2.300017e+12         883 34.657421
#>    ---                                                                    
#> 15031: 23419141 2341914103 234191410301 2.341914e+12         950  8.904454
#> 15032: 23419141 2341914103 234191410301 2.341914e+12         123 43.284089
#> 15033: 23419141 2341914103 234191410301 2.341914e+12         720 25.518845
#> 15034: 23419703 2341970301 234197030101 2.341970e+12         523 13.131100
#> 15035: 23419703 2341970301 234197030101 2.341970e+12         298 16.413875
#>        person_weight age
#>     1:           139   7
#>     2:           139   7
#>     3:           139   7
#>     4:           139   7
#>     5:           139   7
#>    ---                  
#> 15031:           490   1
#> 15032:           490   1
#> 15033:           490   1
#> 15034:           645   9
#> 15035:           645   9
#> 
#> $outliers
#>    threshold num_removed min_outlier max_outlier
#> 1:     0.975         386    110.2297    228233.1
#> 
                            
                            
hts_prep_data(summarize_var = 'employment',
              summarize_by = c('age', 'race'),
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> Creating a summary of person employment broken down by person age and person race
#> $cat
#>          hh_id  person_id person_weight employment age
#>    1: 23000173 2300017301           139          1   7
#>    2: 23000173 2300017304           747          3   6
#>    3: 23000738 2300073801           298          1   8
#>    4: 23000858 2300085801           865          1   6
#>    5: 23000858 2300085802            87          1   7
#>   ---                                                 
#> 1665: 23416563 2341656302           889          1   6
#> 1666: 23418712 2341871201           476          1   6
#> 1667: 23419141 2341914101           136          1   6
#> 1668: 23419141 2341914102           170          5   5
#> 1669: 23419703 2341970301           645          5   9
#>                            race
#>    1:                     White
#>    2:                     White
#>    3:                     White
#>    4:      Prefer not to answer
#>    5:                     White
#>   ---                          
#> 1665: African American or Black
#> 1666:                     Asian
#> 1667:                     White
#> 1668:                     White
#> 1669:                     White
#> 
#> $num
#> NULL
#> 
#> $var_is_shared
#> [1] FALSE
#>