Prepare datasets to make summaries

Prepare datasets to make summaries

Usage

hts_prep_data(
  summarize_var = NULL,
  summarize_by = NULL,
  variables_dt = variable_list,
  data = hts_data,
  id_cols = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id"),
  weighted = TRUE,
  wt_cols = c("hh_weight", "person_weight", "day_weight", "trip_weight", "hh_weight"),
  remove_outliers = TRUE,
  threshold = 0.975,
  remove_missing = TRUE,
  missing_values = c("Missing Response", "995"),
  not_imputable = -1,
  strataname = NULL
)

Arguments

summarize_var

Name of the variable to summarize. Default is NULL

summarize_by

Name of the variable to summarize the summarize_var by. Default is NULL.

variables_dt

List of variable locations and descriptions in data.table format.

data

List of household, person, vehicle, day, and trip tables in data.table format.

id_cols

name of unique identifier for each table in hts_data

weighted

Whether the data is weighted. Default is TRUE.

wt_cols

weight name for each table in hts_data

remove_outliers

Whether to remove outliers for numeric variable. Default is TRUE.

threshold

Threshold to define outliers. Default is 0.975.

remove_missing

Whether to remove missing values from the summary. Default is TRUE.

missing_values

Missing values to remove. Default is 995.

not_imputable

Value representing 'Not imputable' to remove. Default is -1.

strataname

Name of strata name to bring in. Default is NULL.

Value

List containing the categorical and numeric datasets of the summary variables and key columns, and either whether the summarize variable is shared or a breakdown of outliers, depending on if the summarize variable is categorical or numeric.

Examples

require(data.table)
require(stringr)
hts_prep_data(summarize_var = 'age',
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> Creating a summary of person age 
#> $cat
#>          hh_id  person_id age person_weight
#>    1: 23000173 2300017301   7           139
#>    2: 23000173 2300017302   3            17
#>    3: 23000173 2300017303   2           400
#>    4: 23000173 2300017304   6           747
#>    5: 23000738 2300073801   8           298
#>   ---                                      
#> 2038: 23418712 2341871203   2           809
#> 2039: 23419141 2341914101   6           136
#> 2040: 23419141 2341914102   5           170
#> 2041: 23419141 2341914103   1           490
#> 2042: 23419703 2341970301   9           645
#> 
#> $num
#> NULL
#> 
#> $var_is_shared
#> [1] FALSE
#> 
hts_prep_data(summarize_var = 'speed_mph',
              summarize_by = 'age',
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> Creating a summary of trip speed_mph broken down by person age
#> $cat
#>           hh_id  person_id       day_id      trip_id person_weight trip_weight
#>     1: 23000173 2300017301 230001730102 2.300017e+12           139         356
#>     2: 23000173 2300017301 230001730102 2.300017e+12           139         679
#>     3: 23000173 2300017301 230001730102 2.300017e+12           139         758
#>     4: 23000173 2300017301 230001730102 2.300017e+12           139         314
#>     5: 23000173 2300017301 230001730103 2.300017e+12           139         883
#>    ---                                                                        
#> 15031: 23419141 2341914103 234191410301 2.341914e+12           490         950
#> 15032: 23419141 2341914103 234191410301 2.341914e+12           490         123
#> 15033: 23419141 2341914103 234191410301 2.341914e+12           490         720
#> 15034: 23419703 2341970301 234197030101 2.341970e+12           645         523
#> 15035: 23419703 2341970301 234197030101 2.341970e+12           645         298
#>         speed_mph age
#>     1:      17-25   7
#>     2:      17-25   7
#>     3:        1-9   7
#>     4:        1-9   7
#>     5:      33-41   7
#>    ---               
#> 15031:        1-9   1
#> 15032: 43 or more   1
#> 15033:      25-33   1
#> 15034:       9-17   9
#> 15035:       9-17   9
#> 
#> $num
#>           hh_id  person_id       day_id      trip_id person_weight trip_weight
#>     1: 23000173 2300017301 230001730102 2.300017e+12           139         356
#>     2: 23000173 2300017301 230001730102 2.300017e+12           139         679
#>     3: 23000173 2300017301 230001730102 2.300017e+12           139         758
#>     4: 23000173 2300017301 230001730102 2.300017e+12           139         314
#>     5: 23000173 2300017301 230001730103 2.300017e+12           139         883
#>    ---                                                                        
#> 15031: 23419141 2341914103 234191410301 2.341914e+12           490         950
#> 15032: 23419141 2341914103 234191410301 2.341914e+12           490         123
#> 15033: 23419141 2341914103 234191410301 2.341914e+12           490         720
#> 15034: 23419703 2341970301 234197030101 2.341970e+12           645         523
#> 15035: 23419703 2341970301 234197030101 2.341970e+12           645         298
#>        speed_mph age
#>     1: 21.480572   7
#>     2: 23.720178   7
#>     3:  6.927242   7
#>     4:  6.146634   7
#>     5: 34.657421   7
#>    ---              
#> 15031:  8.904454   1
#> 15032: 43.284089   1
#> 15033: 25.518845   1
#> 15034: 13.131100   9
#> 15035: 16.413875   9
#> 
#> $outliers
#>    threshold num_removed min_outlier max_outlier
#> 1:     0.975         386    110.2297    228233.1
#> 
                            
                            
hts_prep_data(summarize_var = 'employment',
              summarize_by = c('age', 'race'),
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> Creating a summary of person employment broken down by person age and person race
#> $cat
#>          hh_id  person_id person_weight employment age
#>    1: 23000173 2300017301           139          1   7
#>    2: 23000173 2300017304           747          3   6
#>    3: 23000738 2300073801           298          1   8
#>    4: 23000858 2300085801           865          1   6
#>    5: 23000858 2300085802            87          1   7
#>   ---                                                 
#> 1665: 23416563 2341656302           889          1   6
#> 1666: 23418712 2341871201           476          1   6
#> 1667: 23419141 2341914101           136          1   6
#> 1668: 23419141 2341914102           170          5   5
#> 1669: 23419703 2341970301           645          5   9
#>                            race
#>    1:                     White
#>    2:                     White
#>    3:                     White
#>    4:      Prefer not to answer
#>    5:                     White
#>   ---                          
#> 1665: African American or Black
#> 1666:                     Asian
#> 1667:                     White
#> 1668:                     White
#> 1669:                     White
#> 
#> $num
#> NULL
#> 
#> $var_is_shared
#> [1] FALSE
#>