Prepare datasets to make summaries

Prepare datasets to make summaries

Usage

hts_prep_data(
  summarize_var = NULL,
  summarize_by = NULL,
  variables_dt = variable_list,
  data = hts_data,
  id_cols = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id"),
  weighted = TRUE,
  wt_cols = c("hh_weight", "person_weight", "day_weight", "trip_weight", "hh_weight"),
  remove_outliers = TRUE,
  threshold = 0.975,
  remove_missing = TRUE,
  missing_values = c("Missing Response", "995"),
  not_imputable = -1,
  strataname = NULL
)

Arguments

summarize_var

Name of the variable to summarize. Default is NULL

summarize_by

Name of the variable to summarize the summarize_var by. Default is NULL.

variables_dt

List of variable locations and descriptions in data.table format.

data

List of household, person, vehicle, day, and trip tables in data.table format.

id_cols

name of unique identifier for each table in hts_data

weighted

Whether the data is weighted. Default is TRUE.

wt_cols

weight name for each table in hts_data

remove_outliers

Whether to remove outliers for numeric variable. Default is TRUE.

threshold

Threshold to define outliers. Default is 0.975.

remove_missing

Whether to remove missing values from the summary. Default is TRUE.

missing_values

Missing values to remove. Default is 995.

not_imputable

Value representing 'Not imputable' to remove. Default is -1.

strataname

Name of strata name to bring in. Default is NULL.

Value

List containing the categorical and numeric datasets of the summary variables and key columns, and either whether the summarize variable is shared or a breakdown of outliers, depending on if the summarize variable is categorical or numeric.

Examples

require(data.table)
require(stringr)
hts_prep_data(summarize_var = 'age',
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> Creating a summary of person age 
#> $cat
#>          hh_id  person_id age person_weight
#>    1: 23000173 2300017301   7           196
#>    2: 23000173 2300017302   3           671
#>    3: 23000173 2300017303   2           577
#>    4: 23000173 2300017304   6           688
#>    5: 23000405 2300040501   5           609
#>   ---                                      
#> 1980: 23417656 2341765603   1           878
#> 1981: 23418585 2341858501   8           586
#> 1982: 23418585 2341858502   8           199
#> 1983: 23422311 2342231101   8           546
#> 1984: 23422311 2342231102   8           617
#> 
#> $num
#> NULL
#> 
hts_prep_data(summarize_var = 'speed_mph',
              summarize_by = 'age',
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> Creating a summary of trip speed_mph broken down by person age
#> $cat
#>           hh_id  person_id       day_id      trip_id person_weight trip_weight
#>     1: 23000173 2300017301 230001730102 2.300017e+12           196         968
#>     2: 23000173 2300017301 230001730102 2.300017e+12           196         906
#>     3: 23000173 2300017301 230001730102 2.300017e+12           196         387
#>     4: 23000173 2300017301 230001730102 2.300017e+12           196         684
#>     5: 23000173 2300017301 230001730103 2.300017e+12           196          64
#>    ---                                                                        
#> 14681: 23422311 2342231101 234223110101 2.342231e+12           546         624
#> 14682: 23422311 2342231102 234223110201 2.342231e+12           617         322
#> 14683: 23422311 2342231102 234223110201 2.342231e+12           617          22
#> 14684: 23422311 2342231102 234223110201 2.342231e+12           617         596
#> 14685: 23422311 2342231102 234223110201 2.342231e+12           617          95
#>        speed_mph age
#>     1:     17-25   7
#>     2:     17-25   7
#>     3:       1-9   7
#>     4:       1-9   7
#>     5:     33-41   7
#>    ---              
#> 14681:     25-33   8
#> 14682:      9-17   8
#> 14683:     25-33   8
#> 14684:      9-17   8
#> 14685:     25-33   8
#> 
#> $num
#>           hh_id  person_id       day_id      trip_id person_weight trip_weight
#>     1: 23000173 2300017301 230001730102 2.300017e+12           196         968
#>     2: 23000173 2300017301 230001730102 2.300017e+12           196         906
#>     3: 23000173 2300017301 230001730102 2.300017e+12           196         387
#>     4: 23000173 2300017301 230001730102 2.300017e+12           196         684
#>     5: 23000173 2300017301 230001730103 2.300017e+12           196          64
#>    ---                                                                        
#> 14681: 23422311 2342231101 234223110101 2.342231e+12           546         624
#> 14682: 23422311 2342231102 234223110201 2.342231e+12           617         322
#> 14683: 23422311 2342231102 234223110201 2.342231e+12           617          22
#> 14684: 23422311 2342231102 234223110201 2.342231e+12           617         596
#> 14685: 23422311 2342231102 234223110201 2.342231e+12           617          95
#>        speed_mph age
#>     1: 21.480572   7
#>     2: 23.720178   7
#>     3:  6.927242   7
#>     4:  6.146634   7
#>     5: 34.657421   7
#>    ---              
#> 14681: 25.812478   8
#> 14682: 16.040986   8
#> 14683: 26.901306   8
#> 14684: 16.720703   8
#> 14685: 25.812478   8
#> 
#> $outliers
#>    threshold num_removed min_outlier max_outlier
#> 1:     0.975         377    113.0093     5198308
#> 
                            
                            
hts_prep_data(summarize_var = 'employment',
              summarize_by = c('age', 'race'),
              variables_dt = variable_list,
              data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> Creating a summary of person employment broken down by person age and person race
#> $cat
#>          hh_id  person_id person_weight employment age                 race
#>    1: 23000173 2300017301           196          1   7                White
#>    2: 23000173 2300017304           688          3   6                White
#>    3: 23000405 2300040501           609          1   5                White
#>    4: 23000405 2300040502            90          1   5                White
#>    5: 23000466 2300046601            62          5   9                White
#>   ---                                                                      
#> 1622: 23417656 2341765602           856          1   6                Asian
#> 1623: 23418585 2341858501           586          1   8                White
#> 1624: 23418585 2341858502           199          2   8                White
#> 1625: 23422311 2342231101           546          7   8 Prefer not to answer
#> 1626: 23422311 2342231102           617          5   8 Prefer not to answer
#> 
#> $num
#> NULL
#>