Prepare datasets for trip rate calculations

Prepare datasets for trip rate calculations

Usage

hts_prep_triprate(
  summarize_by = NULL,
  variables_dt = variable_list,
  trip_name = "trip",
  day_name = "day",
  ids = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id"),
  wts = c("hh_weight", "person_weight", "day_weight", "trip_weight", "hh_weight"),
  remove_outliers = TRUE,
  threshold = 0.975,
  weighted = TRUE,
  hts_data
)

Arguments

summarize_by

Name of the variable to summarize trip rates by. Default is NULL.

variables_dt

List of variable locations and descriptions in data.table format.

trip_name

Name of the trip dataset in hts_data.

day_name

Name of the day dataset in hts_data.

ids

name of unique identifier in each table in hts_data

wts

name of weight column in each table in hts_data

remove_outliers

Boolean whether or not to remove outliers from dataset. Default is TRUE.

threshold

Threshold to define outliers. Default is 0.975.

weighted

Whether the data is weighted. Default is TRUE.

hts_data

List containing household, person, day, trip, and vehicle datasets in data.table format.

Value

List of binned number of trips with key columns and summarize by variable, unbinned number of trips with key columns and summarize by variable, and a breakdown of outliers if removed.

Examples

require(data.table)
require(stringr)
hts_prep_triprate(variables_dt = variable_list,
                  trip_name = 'trip',
                  day_name = 'day',
                  hts_data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> $cat
#> Key: <hh_id, person_id, day_id>
#>          hh_id  person_id       day_id day_weight num_trips_wtd
#>          <int>      <num>        <num>      <int>        <fctr>
#>    1: 23000173 2300017301 230001730101        694     Exactly 0
#>    2: 23000173 2300017301 230001730102        296          7-14
#>    3: 23000173 2300017301 230001730103        352           0-7
#>    4: 23000173 2300017301 230001730104        478           0-7
#>    5: 23000173 2300017301 230001730105        300           0-7
#>   ---                                                          
#> 4181: 23417656 2341765603 234176560301        971           0-7
#> 4182: 23418585 2341858501 234185850101        892           0-7
#> 4183: 23418585 2341858502 234185850201        268     Exactly 0
#> 4184: 23422311 2342231101 234223110101        161          7-14
#> 4185: 23422311 2342231102 234223110201        350           0-7
#> 
#> $num
#> Key: <hh_id, person_id, day_id>
#>          hh_id  person_id       day_id day_weight num_trips_wtd
#>          <int>      <num>        <num>      <int>         <num>
#>    1: 23000173 2300017301 230001730101        694      0.000000
#>    2: 23000173 2300017301 230001730102        296      9.949324
#>    3: 23000173 2300017301 230001730103        352      2.906250
#>    4: 23000173 2300017301 230001730104        478      1.757322
#>    5: 23000173 2300017301 230001730105        300      3.303333
#>   ---                                                          
#> 4181: 23417656 2341765603 234176560301        971      1.685891
#> 4182: 23418585 2341858501 234185850101        892      1.050448
#> 4183: 23418585 2341858502 234185850201        268      0.000000
#> 4184: 23422311 2342231101 234223110101        161     10.329193
#> 4185: 23422311 2342231102 234223110201        350      2.957143
#> 
#> $outliers
#>    threshold num_removed min_outlier max_outlier
#>        <num>       <int>       <num>       <num>
#> 1:     0.975         108    63.74286    340.3529
#> 
hts_prep_triprate(summarize_by = 'age',
                  variables_dt = variable_list,
                  trip_name = 'trip',
                  day_name = 'day',
                  hts_data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> $cat
#>          hh_id  person_id       day_id day_weight person_weight   age
#>          <int>      <num>        <num>      <int>         <int> <int>
#>    1: 23000173 2300017301 230001730101        694           196     7
#>    2: 23000173 2300017301 230001730102        296           196     7
#>    3: 23000173 2300017301 230001730103        352           196     7
#>    4: 23000173 2300017301 230001730104        478           196     7
#>    5: 23000173 2300017301 230001730105        300           196     7
#>   ---                                                                
#> 4181: 23417656 2341765603 234176560301        971           878     1
#> 4182: 23418585 2341858501 234185850101        892           586     8
#> 4183: 23418585 2341858502 234185850201        268           199     8
#> 4184: 23422311 2342231101 234223110101        161           546     8
#> 4185: 23422311 2342231102 234223110201        350           617     8
#>       num_trips_wtd
#>              <fctr>
#>    1:     Exactly 0
#>    2:          7-14
#>    3:           0-7
#>    4:           0-7
#>    5:           0-7
#>   ---              
#> 4181:           0-7
#> 4182:           0-7
#> 4183:     Exactly 0
#> 4184:          7-14
#> 4185:           0-7
#> 
#> $num
#>          hh_id  person_id       day_id day_weight person_weight   age
#>          <int>      <num>        <num>      <int>         <int> <int>
#>    1: 23000173 2300017301 230001730101        694           196     7
#>    2: 23000173 2300017301 230001730102        296           196     7
#>    3: 23000173 2300017301 230001730103        352           196     7
#>    4: 23000173 2300017301 230001730104        478           196     7
#>    5: 23000173 2300017301 230001730105        300           196     7
#>   ---                                                                
#> 4181: 23417656 2341765603 234176560301        971           878     1
#> 4182: 23418585 2341858501 234185850101        892           586     8
#> 4183: 23418585 2341858502 234185850201        268           199     8
#> 4184: 23422311 2342231101 234223110101        161           546     8
#> 4185: 23422311 2342231102 234223110201        350           617     8
#>       num_trips_wtd
#>               <num>
#>    1:      0.000000
#>    2:      9.949324
#>    3:      2.906250
#>    4:      1.757322
#>    5:      3.303333
#>   ---              
#> 4181:      1.685891
#> 4182:      1.050448
#> 4183:      0.000000
#> 4184:     10.329193
#> 4185:      2.957143
#> 
#> $outliers
#>    threshold num_removed min_outlier max_outlier
#>        <num>       <int>       <num>       <num>
#> 1:     0.975         108    63.74286    340.3529
#>