Prepare datasets for trip rate calculations

Prepare datasets for trip rate calculations

Usage

hts_prep_triprate(
  summarize_by = NULL,
  variables_dt = variable_list,
  trip_name = "trip",
  day_name = "day",
  ids = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id"),
  wts = c("hh_weight", "person_weight", "day_weight", "trip_weight", "hh_weight"),
  remove_outliers = TRUE,
  threshold = 0.975,
  weighted = TRUE,
  hts_data
)

Arguments

summarize_by

Name of the variable to summarize trip rates by. Default is NULL.

variables_dt

List of variable locations and descriptions in data.table format.

trip_name

Name of the trip dataset in hts_data.

day_name

Name of the day dataset in hts_data.

ids

name of unique identifier in each table in hts_data

wts

name of weight column in each table in hts_data

remove_outliers

Boolean whether or not to remove outliers from dataset. Default is TRUE.

threshold

Threshold to define outliers. Default is 0.975.

weighted

Whether the data is weighted. Default is TRUE.

hts_data

List containing household, person, day, trip, and vehicle datasets in data.table format.

Value

List of binned number of trips with key columns and summarize by variable, unbinned number of trips with key columns and summarize by variable, and a breakdown of outliers if removed.

Examples

require(data.table)
require(stringr)
hts_prep_triprate(variables_dt = variable_list,
                  trip_name = 'trip',
                  day_name = 'day',
                  hts_data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> $cat
#>          hh_id  person_id       day_id day_weight num_trips_wtd
#>    1: 23000173 2300017301 230001730101        775     Exactly 0
#>    2: 23000173 2300017301 230001730102        244          7-14
#>    3: 23000173 2300017301 230001730103         77    36 or more
#>    4: 23000173 2300017301 230001730104         91           0-7
#>    5: 23000173 2300017301 230001730106         55    36 or more
#>   ---                                                          
#> 4213: 23418712 2341871203 234187120301        119          7-14
#> 4214: 23419141 2341914101 234191410101        626          7-14
#> 4215: 23419141 2341914102 234191410201        595           0-7
#> 4216: 23419141 2341914103 234191410301         49    36 or more
#> 4217: 23419703 2341970301 234197030101         69          7-14
#> 
#> $num
#>          hh_id  person_id       day_id day_weight num_trips_wtd
#>    1: 23000173 2300017301 230001730101        775      0.000000
#>    2: 23000173 2300017301 230001730102        244      8.635246
#>    3: 23000173 2300017301 230001730103         77     39.935065
#>    4: 23000173 2300017301 230001730104         91      4.681319
#>    5: 23000173 2300017301 230001730106         55     51.763636
#>   ---                                                          
#> 4213: 23418712 2341871203 234187120301        119     11.235294
#> 4214: 23419141 2341914101 234191410101        626      8.731629
#> 4215: 23419141 2341914102 234191410201        595      5.848739
#> 4216: 23419141 2341914103 234191410301         49     58.387755
#> 4217: 23419703 2341970301 234197030101         69     11.898551
#> 
#> $outliers
#>    threshold num_removed min_outlier max_outlier
#> 1:     0.975         109     59.2663       469.7
#> 
hts_prep_triprate(summarize_by = 'age',
                  variables_dt = variable_list,
                  trip_name = 'trip',
                  day_name = 'day',
                  hts_data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))
#> $cat
#>          hh_id  person_id       day_id day_weight person_weight age
#>    1: 23000173 2300017301 230001730101        775           139   7
#>    2: 23000173 2300017301 230001730102        244           139   7
#>    3: 23000173 2300017301 230001730103         77           139   7
#>    4: 23000173 2300017301 230001730104         91           139   7
#>    5: 23000173 2300017301 230001730106         55           139   7
#>   ---                                                              
#> 4213: 23418712 2341871203 234187120301        119           809   2
#> 4214: 23419141 2341914101 234191410101        626           136   6
#> 4215: 23419141 2341914102 234191410201        595           170   5
#> 4216: 23419141 2341914103 234191410301         49           490   1
#> 4217: 23419703 2341970301 234197030101         69           645   9
#>       num_trips_wtd
#>    1:     Exactly 0
#>    2:          7-14
#>    3:    36 or more
#>    4:           0-7
#>    5:    36 or more
#>   ---              
#> 4213:          7-14
#> 4214:          7-14
#> 4215:           0-7
#> 4216:    36 or more
#> 4217:          7-14
#> 
#> $num
#>          hh_id  person_id       day_id day_weight person_weight age
#>    1: 23000173 2300017301 230001730101        775           139   7
#>    2: 23000173 2300017301 230001730102        244           139   7
#>    3: 23000173 2300017301 230001730103         77           139   7
#>    4: 23000173 2300017301 230001730104         91           139   7
#>    5: 23000173 2300017301 230001730106         55           139   7
#>   ---                                                              
#> 4213: 23418712 2341871203 234187120301        119           809   2
#> 4214: 23419141 2341914101 234191410101        626           136   6
#> 4215: 23419141 2341914102 234191410201        595           170   5
#> 4216: 23419141 2341914103 234191410301         49           490   1
#> 4217: 23419703 2341970301 234197030101         69           645   9
#>       num_trips_wtd
#>    1:      0.000000
#>    2:      8.635246
#>    3:     39.935065
#>    4:      4.681319
#>    5:     51.763636
#>   ---              
#> 4213:     11.235294
#> 4214:      8.731629
#> 4215:      5.848739
#> 4216:     58.387755
#> 4217:     11.898551
#> 
#> $outliers
#>    threshold num_removed min_outlier max_outlier
#> 1:     0.975         109     59.2663       469.7
#>