Remove missing data for summary variables

Remove missing data for summary variables

Usage

hts_remove_missing_data(
  hts_data,
  variables_dt,
  summarize_var,
  ids = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id"),
  summarize_by = NULL,
  missing_values = c("Missing Response", "995"),
  not_imputable = -1
)

Arguments

hts_data

List containing household, person, day, trip, and vehicle datasets in data.table format.

variables_dt

A variable list with descriptions and table locations of variables.

summarize_var

Variable to be summarized that has it's missing data removed.

ids

names of unique identifiers for each table in hts_data

summarize_by

Variable being summarized by that has it's missing data removed. Default is NULL.

missing_values

Missing values that will be removed. Defaults are 995 and 'Missing Response'.

not_imputable

Value meaning not_imputable that will be removed. Default is -1.

Value

Inputted list of datasets without missing values for specified variables.

Examples

require(data.table)
hts_remove_missing_data(hts_data = list('hh' = hh,
                                        'person' = person,
                                        'day' = day,
                                        'trip' = trip,
                                        'vehicle' = vehicle),
                        variables_dt = variable_list,
                        summarize_var = 'speed_mph',
                        summarize_by = 'mode_type')
#> $hh
#>          hh_id sample_segment num_trips num_people income_detailed
#>          <int>          <int>     <int>      <int>           <int>
#>    1: 23000173              8        66          4               8
#>    2: 23000405             16        80          2              10
#>    3: 23000466             16        81          3               5
#>    4: 23000799              4        45          2               8
#>    5: 23000858             16        10          2              10
#>   ---                                                             
#>  996: 23414932             14         0          1               7
#>  997: 23416563             14        16          6               7
#>  998: 23417656              5        11          3              10
#>  999: 23418585              5         2          2               5
#> 1000: 23422311             14         7          2               8
#>       income_followup  home_lon home_lat home_county residence_type hh_weight
#>                 <int>     <num>    <num>       <int>          <int>     <int>
#>    1:             995 -94.77477 38.86486           2              1       190
#>    2:             995 -91.39139 36.38438           2            997        62
#>    3:             995 -89.40941 35.37538           3              4       313
#>    4:             995 -88.98899 36.18819           1              1       425
#>    5:             995 -91.41141 35.99199           3              2        85
#>   ---                                                                        
#>  996:             995 -82.72272 35.94294           3              1        13
#>  997:             995 -88.36837 39.86687           1              2        94
#>  998:             995 -94.37437 33.04204           1              1       759
#>  999:             995 -88.14815 35.20020           2              1       663
#> 1000:             995 -93.57357 37.16216           2              1       888
#> 
#> $person
#>       ethnicity_1 ethnicity_2 ethnicity_3 ethnicity_4 ethnicity_997
#>             <int>       <int>       <int>       <int>         <int>
#>    1:           1           0           0           0             0
#>    2:         995         995         995         995           995
#>    3:         995         995         995         995           995
#>    4:           1           0           0           0             0
#>    5:           1           0           0           0             0
#>   ---                                                              
#> 1980:         995         995         995         995           995
#> 1981:           0           1           0           0             0
#> 1982:           0           1           0           0             0
#> 1983:           1           0           0           0             0
#> 1984:           1           0           0           0             0
#>       ethnicity_999 race_1 race_2 race_3 race_4 race_5 race_997 race_999
#>               <int>  <int>  <int>  <int>  <int>  <int>    <int>    <int>
#>    1:             0      0      0      0      0      1        0        0
#>    2:           995    995    995    995    995    995      995      995
#>    3:           995    995    995    995    995    995      995      995
#>    4:             0      0      0      0      0      1        0        0
#>    5:             0      0      0      0      0      1        0        0
#>   ---                                                                   
#> 1980:           995    995    995    995    995    995      995      995
#> 1981:             0      0      0      0      0      1        0        0
#> 1982:             0      0      0      0      0      1        0        0
#> 1983:             0      0      0      0      0      0        0        1
#> 1984:             0      0      0      0      0      0        0        1
#>          hh_id num_trips  person_id   age gender employment job_type education
#>          <int>     <int>      <num> <int>  <int>      <int>    <int>     <int>
#>    1: 23000173        21 2300017301     7      2          1        5         6
#>    2: 23000173         5 2300017302     3      2          5      995       995
#>    3: 23000173         2 2300017303     2      2        995      995       995
#>    4: 23000173        38 2300017304     6      1          3        1         6
#>    5: 23000405        30 2300040501     5      2          1        5         6
#>   ---                                                                         
#> 1980: 23417656         3 2341765603     1      1        995      995       995
#> 1981: 23418585         2 2341858501     8      2          1        1         3
#> 1982: 23418585         0 2341858502     8      1          2        1         4
#> 1983: 23422311         3 2342231101     8      1          7        3         7
#> 1984: 23422311         4 2342231102     8      2          5      995         3
#>       person_weight
#>               <int>
#>    1:           196
#>    2:           671
#>    3:           577
#>    4:           688
#>    5:           609
#>   ---              
#> 1980:           878
#> 1981:           586
#> 1982:           199
#> 1983:           546
#> 1984:           617
#> 
#> $day
#>       delivery_2 delivery_3 delivery_4 delivery_5 delivery_6 delivery_7
#>            <int>      <int>      <int>      <int>      <int>      <int>
#>    1:          0          0          0          0          0          0
#>    2:        995        995        995        995        995        995
#>    3:        995        995        995        995        995        995
#>    4:          0          0          0          0          0          0
#>    5:          0          0          0          0          0          0
#>   ---                                                                  
#> 4289:        995        995        995        995        995        995
#> 4290:          0          0          0          1          0          0
#> 4291:        995        995        995        995        995        995
#> 4292:          0          0          0          1          0          0
#> 4293:        995        995        995        995        995        995
#>       delivery_8 delivery_996    hh_id num_trips  person_id       day_id
#>            <int>        <int>    <int>     <int>      <num>        <num>
#>    1:          0            1 23000173         0 2300017301 230001730101
#>    2:        995          995 23000173        NA 2300017302 230001730201
#>    3:        995          995 23000173        NA 2300017303 230001730301
#>    4:          0            1 23000173         3 2300017304 230001730401
#>    5:          0            1 23000173         4 2300017301 230001730102
#>   ---                                                                   
#> 4289:        995          995 23417656         3 2341765603 234176560301
#> 4290:          0            0 23418585         2 2341858501 234185850101
#> 4291:        995          995 23418585         0 2341858502 234185850201
#> 4292:          0            0 23422311         3 2342231101 234223110101
#> 4293:        995          995 23422311         4 2342231102 234223110201
#>       travel_date begin_day end_day day_weight
#>            <Date>     <int>   <int>      <int>
#>    1:  2023-04-19         1       1        694
#>    2:  2023-04-19       995     995         90
#>    3:  2023-04-19       995     995        728
#>    4:  2023-04-19         1       1        507
#>    5:  2023-04-20         1       1        296
#>   ---                                         
#> 4289:  2023-06-06         1       1        971
#> 4290:  2023-06-06         1       1        892
#> 4291:  2023-06-06         1       1        268
#> 4292:  2023-06-01         1       1        161
#> 4293:  2023-06-01         1       1        350
#> 
#> $trip
#>           hh_id  person_id       day_id travel_date      trip_id speed_mph
#>           <int>      <num>        <num>      <Date>        <num>     <num>
#>     1: 23000173 2300017301 230001730102  2023-04-20 2.300017e+12 21.480572
#>     2: 23000173 2300017301 230001730102  2023-04-20 2.300017e+12 23.720178
#>     3: 23000173 2300017301 230001730102  2023-04-20 2.300017e+12  6.927242
#>     4: 23000173 2300017301 230001730102  2023-04-20 2.300017e+12  6.146634
#>     5: 23000173 2300017301 230001730103  2023-04-21 2.300017e+12 34.657421
#>    ---                                                                    
#> 15362: 23422311 2342231101 234223110101  2023-06-01 2.342231e+12 25.812478
#> 15363: 23422311 2342231102 234223110201  2023-06-01 2.342231e+12 16.040986
#> 15364: 23422311 2342231102 234223110201  2023-06-01 2.342231e+12 26.901306
#> 15365: 23422311 2342231102 234223110201  2023-06-01 2.342231e+12 16.720703
#> 15366: 23422311 2342231102 234223110201  2023-06-01 2.342231e+12 25.812478
#>        distance_miles mode_type mode_1 mode_2 num_travelers d_purpose_category
#>                 <num>     <int>  <int>  <int>         <int>              <int>
#>     1:       4.755560         8      7    995             2                  9
#>     2:       4.856047         8      7    995             3                  1
#>     3:       1.516296         1      1    995             1                  9
#>     4:       1.393237         1      1    995             1                  1
#>     5:       8.933913         8      7    995             1                  3
#>    ---                                                                        
#> 15362:      20.219774         8      6    995             2                  1
#> 15363:       6.683744         8      7    995             1                 10
#> 15364:      20.175979         8      7    995             1                 13
#> 15365:       5.573568         8      6    995             2                  6
#> 15366:      20.219774         8      6    995             2                  1
#>        trip_weight
#>              <int>
#>     1:         968
#>     2:         906
#>     3:         387
#>     4:         684
#>     5:          64
#>    ---            
#> 15362:         624
#> 15363:         322
#> 15364:          22
#> 15365:         596
#> 15366:          95
#> 
#> $vehicle
#> Key: <hh_id>
#>          hh_id vehicle_id fuel_type hh_weight
#>          <int>      <num>     <int>     <int>
#>    1: 23000173 2300017301         1       190
#>    2: 23000173 2300017302         1       190
#>    3: 23000173 2300017303         1       190
#>    4: 23000405 2300040501         1        62
#>    5: 23000466 2300046601         1       313
#>   ---                                        
#> 1384: 23417656 2341765601         1       759
#> 1385: 23418585 2341858501         1       663
#> 1386: 23418585 2341858502         1       663
#> 1387: 23422311 2342231101         4       888
#> 1388: 23422311 2342231102         1       888
#>