Remove missing data for summary variables

Remove missing data for summary variables

hts_remove_missing_data(
  hts_data,
  variables_dt,
  summarize_var,
  summarize_by = NULL,
  missing_values = c("Missing Response", "995"),
  not_imputable = -1
)

Arguments

hts_data

List containing household, person, day, trip, and vehicle datasets in data.table format.

variables_dt

A variable list with descriptions and table locations of variables.

summarize_var

Variable to be summarized that has it's missing data removed.

summarize_by

Variable being summarized by that has it's missing data removed. Default is NULL.

missing_values

Missing values that will be removed. Defaults are 995 and 'Missing Response'.

not_imputable

Value meaning not_imputable that will be removed. Default is -1.

Value

Inputted list of datasets without missing values for specified variables.

Examples

require(data.table)
hts_remove_missing_data(hts_data = list('hh' = hh,
                                        'person' = person,
                                        'day' = day,
                                        'trip' = trip,
                                        'vehicle' = vehicle),
                        variables_dt = variable_list,
                        summarize_var = 'speed_mph',
                        summarize_by = 'mode_type')
#> $hh
#>          hh_id sample_segment num_trips num_people income_detailed
#>    1: 23000173              8        66          4               8
#>    2: 23000738              4        16          1             999
#>    3: 23000858             16        10          2              10
#>    4: 23001054              8        75          2               8
#>    5: 23001072             16        27          1               8
#>   ---                                                             
#>  996: 23415053              5         9          2               8
#>  997: 23416563             14        16          6               7
#>  998: 23418712              5         7          3             999
#>  999: 23419141             14        19          3              10
#> 1000: 23419703             14         2          1               2
#>       income_followup residence_type hh_weight
#>    1:             995              1       190
#>    2:             999              1        54
#>    3:             995              2       912
#>    4:             995              1       182
#>    5:             995              2       958
#>   ---                                         
#>  996:             995              4       232
#>  997:             995              2       928
#>  998:               6              1       314
#>  999:             995              1       811
#> 1000:             995              1       407
#> 
#> $person
#>       ethnicity_1 ethnicity_2 ethnicity_3 ethnicity_4 ethnicity_997
#>    1:           1           0           0           0             0
#>    2:         995         995         995         995           995
#>    3:         995         995         995         995           995
#>    4:           1           0           0           0             0
#>    5:           1           0           0           0             0
#>   ---                                                              
#> 2038:         995         995         995         995           995
#> 2039:           1           0           0           0             0
#> 2040:           1           0           0           0             0
#> 2041:         995         995         995         995           995
#> 2042:           1           0           0           0             0
#>       ethnicity_999 race_1 race_2 race_3 race_4 race_5 race_997 race_999
#>    1:             0      0      0      0      0      1        0        0
#>    2:           995    995    995    995    995    995      995      995
#>    3:           995    995    995    995    995    995      995      995
#>    4:             0      0      0      0      0      1        0        0
#>    5:             0      0      0      0      0      1        0        0
#>   ---                                                                   
#> 2038:           995    995    995    995    995    995      995      995
#> 2039:             0      0      0      0      0      1        0        0
#> 2040:             0      0      0      0      0      1        0        0
#> 2041:           995    995    995    995    995    995      995      995
#> 2042:             0      0      0      0      0      1        0        0
#>          hh_id num_trips  person_id age gender employment education
#>    1: 23000173        21 2300017301   7      2          1         6
#>    2: 23000173         5 2300017302   3      2          5       995
#>    3: 23000173         2 2300017303   2      2        995       995
#>    4: 23000173        38 2300017304   6      1          3         6
#>    5: 23000738        16 2300073801   8      1          1         7
#>   ---                                                              
#> 2038: 23418712         2 2341871203   2      1        995       995
#> 2039: 23419141         9 2341914101   6      2          1         6
#> 2040: 23419141         5 2341914102   5      1          5         2
#> 2041: 23419141         5 2341914103   1      1        995       995
#> 2042: 23419703         2 2341970301   9      2          5         6
#>       person_weight
#>    1:           139
#>    2:            17
#>    3:           400
#>    4:           747
#>    5:           298
#>   ---              
#> 2038:           809
#> 2039:           136
#> 2040:           170
#> 2041:           490
#> 2042:           645
#> 
#> $day
#>       delivery_2 delivery_3 delivery_4 delivery_5 delivery_6 delivery_7
#>    1:          0          0          0          0          0          0
#>    2:        995        995        995        995        995        995
#>    3:        995        995        995        995        995        995
#>    4:          0          0          0          0          0          0
#>    5:          0          0          0          0          0          0
#>   ---                                                                  
#> 4322:        995        995        995        995        995        995
#> 4323:          1          1          0          0          0          0
#> 4324:        995        995        995        995        995        995
#> 4325:        995        995        995        995        995        995
#> 4326:          0          0          0          1          0          0
#>       delivery_8 delivery_996    hh_id num_trips  person_id       day_id
#>    1:          0            1 23000173         0 2300017301 230001730101
#>    2:        995          995 23000173        NA 2300017302 230001730201
#>    3:        995          995 23000173        NA 2300017303 230001730301
#>    4:          0            1 23000173         3 2300017304 230001730401
#>    5:          0            1 23000173         4 2300017301 230001730102
#>   ---                                                                   
#> 4322:        995          995 23418712         2 2341871203 234187120301
#> 4323:          0            0 23419141         9 2341914101 234191410101
#> 4324:        995          995 23419141         5 2341914102 234191410201
#> 4325:        995          995 23419141         5 2341914103 234191410301
#> 4326:          0            0 23419703         2 2341970301 234197030101
#>       travel_date begin_day end_day day_weight
#>    1:  2023-04-19         1       1        775
#>    2:  2023-04-19       995     995         61
#>    3:  2023-04-19       995     995        818
#>    4:  2023-04-19         1       1        673
#>    5:  2023-04-20         1       1        244
#>   ---                                         
#> 4322:  2023-05-31         1       1        119
#> 4323:  2023-05-31         1       1        626
#> 4324:  2023-05-31         1       1        595
#> 4325:  2023-05-31         1       1         49
#> 4326:  2023-05-31         1       1         69
#> 
#> $trip
#>           hh_id  person_id       day_id travel_date      trip_id speed_mph
#>     1: 23000173 2300017301 230001730102  2023-04-20 2.300017e+12 21.480572
#>     2: 23000173 2300017301 230001730102  2023-04-20 2.300017e+12 23.720178
#>     3: 23000173 2300017301 230001730102  2023-04-20 2.300017e+12  6.927242
#>     4: 23000173 2300017301 230001730102  2023-04-20 2.300017e+12  6.146634
#>     5: 23000173 2300017301 230001730103  2023-04-21 2.300017e+12 34.657421
#>    ---                                                                    
#> 15620: 23419141 2341914103 234191410301  2023-05-31 2.341914e+12  8.904454
#> 15621: 23419141 2341914103 234191410301  2023-05-31 2.341914e+12 43.284089
#> 15622: 23419141 2341914103 234191410301  2023-05-31 2.341914e+12 25.518845
#> 15623: 23419703 2341970301 234197030101  2023-05-31 2.341970e+12 13.131100
#> 15624: 23419703 2341970301 234197030101  2023-05-31 2.341970e+12 16.413875
#>        mode_type d_purpose_category trip_weight
#>     1:         8                  9         356
#>     2:         8                  1         679
#>     3:         1                  9         758
#>     4:         1                  1         314
#>     5:         8                  3         883
#>    ---                                         
#> 15620:         8                  1         950
#> 15621:         8                  6         123
#> 15622:         8                  1         720
#> 15623:         8                 13         523
#> 15624:         8                  1         298
#> 
#> $vehicle
#>          hh_id vehicle_id fuel_type hh_weight
#>    1: 23000173 2300017301         1       190
#>    2: 23000173 2300017302         1       190
#>    3: 23000173 2300017303         1       190
#>    4: 23000738 2300073801         5        54
#>    5: 23000858 2300085801         1       912
#>   ---                                        
#> 1453: 23418712 2341871202         1       314
#> 1454: 23419141 2341914101         2       811
#> 1455: 23419141 2341914102         1       811
#> 1456: 23419703 2341970301         1       407
#> 1457: 23419703 2341970302         1       407
#>