Summarize a categorical variable

Summarize a categorical variable

Usage

hts_summary_cat(
  prepped_dt,
  summarize_var = NULL,
  summarize_by = NULL,
  weighted = TRUE,
  se = FALSE,
  wtname = NULL,
  strataname = NULL,
  checkbox_valname = "value",
  checkbox_yesval = 1,
  summarize_vartype = "categorical"
)

Arguments

prepped_dt

Dataset containing the summary variables and key columns in data.table format.

summarize_var

Name of the categorical variable to summarize. Default is NULL.

summarize_by

Name of the variable to summarize the summarize_var by. Default is NULL.

weighted

Whether the data is weighted. Default is TRUE.

se

Whether to calculate standard error. Default is FALSE.

wtname

Name of the weight column to use. Default is NULL.

strataname

Name of strata name to bring in. Default is NULL.

checkbox_valname

Name of the column with the checkbox value. Default is 'value'. Must be provided if summarize_var is a checkbox variable.

checkbox_yesval

Value of checkbox_valname that indicates it was selected. Default is 1. Must be provided if summarize_var is a checkbox variable.

summarize_vartype

String; one of either 'categorical' (when the variable being summarized is categorical) or 'checkbox' (when the variable being summarized is derived from a multiple response, aka select-all-that-apply question).

Value

List of unweighted and weighted categorical summaries including counts and proportions.

Examples

require(data.table)
require(stringr)
require(dplyr)
require(srvyr)
DT = hts_prep_data(summarize_var = 'age',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
#> Creating a summary of person age 
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'age',
                wtname = 'person_weight')
#> $unwtd
#>     age count        prop
#>  1:   1    80 0.039177277
#>  2:   2   187 0.091576885
#>  3:   3    30 0.014691479
#>  4:   4   118 0.057786484
#>  5:   5   454 0.222331048
#>  6:   6   364 0.178256611
#>  7:   7   232 0.113614104
#>  8:   8   252 0.123408423
#>  9:   9   217 0.106268364
#> 10:  10    96 0.047012733
#> 11:  11    12 0.005876592
#> 
#> $wtd
#>     age count        prop    est
#>  1:   1    80 0.035449836  36529
#>  2:   2   187 0.094923344  97813
#>  3:   3    30 0.015961112  16447
#>  4:   4   118 0.056409774  58127
#>  5:   5   454 0.217321305 223937
#>  6:   6   364 0.183085511 188659
#>  7:   7   232 0.110731123 114102
#>  8:   8   252 0.124048709 127825
#>  9:   9   217 0.108907634 112223
#> 10:  10    96 0.046645032  48065
#> 11:  11    12 0.006516621   6715
#> 
#> $weight_name
#> [1] "person_weight"
#> 
                
DT = hts_prep_data(summarize_var = 'race',
                   summarize_by = 'employment',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
#> Creating a summary of person race broken down by person employment
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'race',
                summarize_by = 'employment',
                wtname = 'person_weight',
                summarize_vartype = 'checkbox')
#> Error in hts_summary_cat(prepped_dt = DT, summarize_var = "race", summarize_by = "employment",     wtname = "person_weight", summarize_vartype = "checkbox"): unused argument (summarize_vartype = "checkbox")
                
DT = hts_prep_data(summarize_var = 'employment',
                summarize_by = c('race', 'income_detailed', 'gender'),
                variables_dt = variable_list,
                data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))$cat
#> Creating a summary of person employment broken down by person race and hh income_detailed and person gender

hts_summary_cat(prepped_dt = DT,
                summarize_var = 'employment',
                summarize_by = c('race', 'income_detailed', 'gender'),
                wtname = 'person_weight')$unwtd
#>                           race income_detailed gender employment count
#>   1: African American or Black               1      1          5     1
#>   2: African American or Black               1      2          6     1
#>   3: African American or Black               1      2          8     1
#>   4: African American or Black               1    999          6     1
#>   5: African American or Black               2      1          1     1
#>  ---                                                                  
#> 393:      Prefer not to answer             999      2          8     1
#> 394:      Prefer not to answer             999    999          1    22
#> 395:      Prefer not to answer             999    999          3     4
#> 396:      Prefer not to answer             999    999          5     6
#> 397:      Prefer not to answer             999    999          8     1
#>            prop
#>   1: 1.00000000
#>   2: 0.50000000
#>   3: 0.50000000
#>   4: 1.00000000
#>   5: 0.20000000
#>  ---           
#> 393: 0.05882353
#> 394: 0.66666667
#> 395: 0.12121212
#> 396: 0.18181818
#> 397: 0.03030303