Summarize a categorical variable

Summarize a categorical variable

Usage

hts_summary_cat(
  prepped_dt,
  summarize_var = NULL,
  summarize_by = NULL,
  weighted = TRUE,
  se = FALSE,
  wtname = NULL,
  strataname = NULL,
  checkbox_valname = "value",
  checkbox_yesval = 1,
  summarize_vartype = "categorical"
)

Arguments

prepped_dt

Dataset containing the summary variables and key columns in data.table format.

summarize_var

Name of the categorical variable to summarize. Default is NULL.

summarize_by

Name of the variable to summarize the summarize_var by. Default is NULL.

weighted

Whether the data is weighted. Default is TRUE.

se

Whether to calculate standard error. Default is FALSE.

wtname

Name of the weight column to use. Default is NULL.

strataname

Name of strata name to bring in. Default is NULL.

checkbox_valname

Name of the column with the checkbox value. Default is 'value'. Must be provided if summarize_var is a checkbox variable.

checkbox_yesval

Value of checkbox_valname that indicates it was selected. Default is 1. Must be provided if summarize_var is a checkbox variable.

summarize_vartype

String; one of either 'categorical' (when the variable being summarized is categorical) or 'checkbox' (when the variable being summarized is derived from a multiple response, aka select-all-that-apply question).

Value

List of unweighted and weighted categorical summaries including counts and proportions.

Examples

require(data.table)
require(stringr)
require(dplyr)
require(srvyr)
DT = hts_prep_data(summarize_var = 'age',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
#> Creating a summary of person age 
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'age',
                wtname = 'person_weight')
#> $unwtd
#> Key: <age>
#>       age count        prop
#>     <int> <int>       <num>
#>  1:     1    79 0.039818548
#>  2:     2   194 0.097782258
#>  3:     3    33 0.016633065
#>  4:     4    99 0.049899194
#>  5:     5   389 0.196068548
#>  6:     6   354 0.178427419
#>  7:     7   234 0.117943548
#>  8:     8   259 0.130544355
#>  9:     9   218 0.109879032
#> 10:    10   119 0.059979839
#> 11:    11     6 0.003024194
#> 
#> $wtd
#>       age count        prop    est
#>     <int> <int>       <num>  <int>
#>  1:     1    79 0.042732174  43036
#>  2:     2   194 0.092259038  92915
#>  3:     3    33 0.018525285  18657
#>  4:     4    99 0.049041316  49390
#>  5:     5   389 0.198440091 199851
#>  6:     6   354 0.171938517 173161
#>  7:     7   234 0.117023960 117856
#>  8:     8   259 0.134759857 135718
#>  9:     9   218 0.109464706 110243
#> 10:    10   119 0.062543317  62988
#> 11:    11     6 0.003271738   3295
#> 
#> $weight_name
#> [1] "person_weight"
#> 
                
DT = hts_prep_data(summarize_var = 'race',
                   summarize_by = 'employment',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
#> Creating a summary of person race broken down by person employment
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'race',
                summarize_by = 'employment',
                wtname = 'person_weight',
                summarize_vartype = 'checkbox')
#> $unwtd
#> Key: <employment>
#>     employment                                      race count        prop
#>          <int>                                    <fctr> <int>       <num>
#>  1:          1                 African American or Black    24 0.028103044
#>  2:          1          American Indian or Alaska Native     6 0.007025761
#>  3:          1                                     Asian   203 0.237704918
#>  4:          1 Native Hawaiian or other Pacific Islander    11 0.012880562
#>  5:          1                                     White   505 0.591334895
#>  6:          1                                Other race    13 0.015222482
#>  7:          1                      Prefer not to answer    92 0.107728337
#>  8:          2                 African American or Black    10 0.064516129
#>  9:          2          American Indian or Alaska Native     2 0.012903226
#> 10:          2                                     Asian    25 0.161290323
#> 11:          2 Native Hawaiian or other Pacific Islander     1 0.006451613
#> 12:          2                                     White   101 0.651612903
#> 13:          2                      Prefer not to answer    16 0.103225806
#> 14:          3                 African American or Black     1 0.009433962
#> 15:          3          American Indian or Alaska Native     3 0.028301887
#> 16:          3                                     Asian    15 0.141509434
#> 17:          3                                     White    76 0.716981132
#> 18:          3                                Other race     2 0.018867925
#> 19:          3                      Prefer not to answer     9 0.084905660
#> 20:          5                 African American or Black    10 0.020964361
#> 21:          5          American Indian or Alaska Native    11 0.023060797
#> 22:          5                                     Asian    69 0.144654088
#> 23:          5 Native Hawaiian or other Pacific Islander     1 0.002096436
#> 24:          5                                     White   343 0.719077568
#> 25:          5                                Other race     8 0.016771488
#> 26:          5                      Prefer not to answer    35 0.073375262
#> 27:          6                 African American or Black     8 0.100000000
#> 28:          6                                     Asian     9 0.112500000
#> 29:          6                                     White    51 0.637500000
#> 30:          6                                Other race     2 0.025000000
#> 31:          6                      Prefer not to answer    10 0.125000000
#> 32:          7                 African American or Black     3 0.150000000
#> 33:          7          American Indian or Alaska Native     1 0.050000000
#> 34:          7                                     Asian     4 0.200000000
#> 35:          7                                     White     8 0.400000000
#> 36:          7                                Other race     1 0.050000000
#> 37:          7                      Prefer not to answer     3 0.150000000
#> 38:          8                                     Asian     1 0.111111111
#> 39:          8                                     White     4 0.444444444
#> 40:          8                                Other race     1 0.111111111
#> 41:          8                      Prefer not to answer     3 0.333333333
#>     employment                                      race count        prop
#> 
#> $wtd
#>     employment                                      race count        prop
#>          <int>                                    <fctr> <int>       <num>
#>  1:          1                 African American or Black    24 0.027352051
#>  2:          1          American Indian or Alaska Native     6 0.008841747
#>  3:          1                                     Asian   203 0.246026788
#>  4:          1 Native Hawaiian or other Pacific Islander    11 0.016833461
#>  5:          1                                     White   505 0.591495944
#>  6:          1                                Other race    13 0.017100548
#>  7:          1                      Prefer not to answer    92 0.092349462
#>  8:          2                 African American or Black    10 0.059962428
#>  9:          2          American Indian or Alaska Native     2 0.014888369
#> 10:          2                                     Asian    25 0.146174392
#> 11:          2 Native Hawaiian or other Pacific Islander     1 0.006134264
#> 12:          2                                     White   101 0.644775013
#> 13:          2                      Prefer not to answer    16 0.128065534
#> 14:          3                 African American or Black     1 0.014977871
#> 15:          3          American Indian or Alaska Native     3 0.020563168
#> 16:          3                                     Asian    15 0.139458477
#> 17:          3                                     White    76 0.716947220
#> 18:          3                                Other race     2 0.012639391
#> 19:          3                      Prefer not to answer     9 0.095413872
#> 20:          5                 African American or Black    10 0.024816957
#> 21:          5          American Indian or Alaska Native    11 0.025288943
#> 22:          5                                     Asian    69 0.138990029
#> 23:          5 Native Hawaiian or other Pacific Islander     1 0.003165084
#> 24:          5                                     White   343 0.709776064
#> 25:          5                                Other race     8 0.020767394
#> 26:          5                      Prefer not to answer    35 0.077195529
#> 27:          6                 African American or Black     8 0.087148966
#> 28:          6                                     Asian     9 0.088099227
#> 29:          6                                     White    51 0.674810573
#> 30:          6                                Other race     2 0.021605942
#> 31:          6                      Prefer not to answer    10 0.128335292
#> 32:          7                 African American or Black     3 0.102527544
#> 33:          7          American Indian or Alaska Native     1 0.031626701
#> 34:          7                                     Asian     4 0.157355800
#> 35:          7                                     White     8 0.427478937
#> 36:          7                                Other race     1 0.031626701
#> 37:          7                      Prefer not to answer     3 0.249384316
#> 38:          8                                     Asian     1 0.166990763
#> 39:          8                                     White     4 0.401798736
#> 40:          8                                Other race     1 0.095527467
#> 41:          8                      Prefer not to answer     3 0.335683034
#>     employment                                      race count        prop
#>        est
#>      <int>
#>  1:  11777
#>  2:   3807
#>  3: 105932
#>  4:   7248
#>  5: 254681
#>  6:   7363
#>  7:  39763
#>  8:   4692
#>  9:   1165
#> 10:  11438
#> 11:    480
#> 12:  50453
#> 13:  10021
#> 14:    775
#> 15:   1064
#> 16:   7216
#> 17:  37097
#> 18:    654
#> 19:   4937
#> 20:   6257
#> 21:   6376
#> 22:  35043
#> 23:    798
#> 24: 178953
#> 25:   5236
#> 26:  19463
#> 27:   3485
#> 28:   3523
#> 29:  26985
#> 30:    864
#> 31:   5132
#> 32:    791
#> 33:    244
#> 34:   1214
#> 35:   3298
#> 36:    244
#> 37:   1924
#> 38:    687
#> 39:   1653
#> 40:    393
#> 41:   1381
#>        est
#> 
#> $weight_name
#> [1] "person_weight"
#> 
                
DT = hts_prep_data(summarize_var = 'employment',
                summarize_by = c('race', 'income_detailed', 'gender'),
                variables_dt = variable_list,
                data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))$cat
#> Creating a summary of person employment broken down by person race and hh income_detailed and person gender

hts_summary_cat(prepped_dt = DT,
                summarize_var = 'employment',
                summarize_by = c('race', 'income_detailed', 'gender'),
                wtname = 'person_weight')$unwtd
#> Key: <race, income_detailed, gender>
#>                           race income_detailed gender employment count
#>                         <fctr>           <int>  <int>      <int> <int>
#>   1: African American or Black               1      1          5     2
#>   2: African American or Black               1      1          6     1
#>   3: African American or Black               1      2          6     1
#>   4: African American or Black               2      1          1     1
#>   5: African American or Black               2      1          5     2
#>  ---                                                                  
#> 394:      Prefer not to answer             999    999          1    24
#> 395:      Prefer not to answer             999    999          2     1
#> 396:      Prefer not to answer             999    999          3     1
#> 397:      Prefer not to answer             999    999          5     5
#> 398:      Prefer not to answer             999    999          6     1
#>           prop
#>          <num>
#>   1: 0.6666667
#>   2: 0.3333333
#>   3: 1.0000000
#>   4: 0.2000000
#>   5: 0.4000000
#>  ---          
#> 394: 0.7500000
#> 395: 0.0312500
#> 396: 0.0312500
#> 397: 0.1562500
#> 398: 0.0312500