Summarize a categorical variable

Summarize a categorical variable

Usage

hts_summary_cat(
  prepped_dt,
  summarize_var = NULL,
  summarize_by = NULL,
  weighted = TRUE,
  se = FALSE,
  wtname = NULL,
  strataname = NULL,
  checkbox_valname = NULL,
  checkbox_yesval = NULL
)

Arguments

prepped_dt

Dataset containing the summary variables and key columns in data.table format.

summarize_var

Name of the categorical variable to summarize. Default is NULL.

summarize_by

Name of the variable to summarize the summarize_var by. Default is NULL.

weighted

Whether the data is weighted. Default is TRUE.

se

Whether to calculate standard error. Default is FALSE.

wtname

Name of the weight column to use. Default is NULL.

strataname

Name of strata name to bring in. Default is NULL.

checkbox_valname

Name of the column with the checkbox value. Default is NULL. Must be provided if summarize_var is a checkbox variable.

checkbox_yesval

Value of checkbox_valname that indicates it was selected. Default is NULL. Must be provided if summarize_var is a checkbox variable.

Value

List of unweighted and weighted categorical summaries including counts and proportions.

Examples

require(data.table)
require(stringr)
require(dplyr)
require(srvyr)
DT = hts_prep_data(summarize_var = 'age',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
#> Creating a summary of person age 
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'age',
                wtname = 'person_weight')
#> $unwtd
#>     age count        prop
#>  1:   1    80 0.039177277
#>  2:   2   187 0.091576885
#>  3:   3    30 0.014691479
#>  4:   4   118 0.057786484
#>  5:   5   454 0.222331048
#>  6:   6   364 0.178256611
#>  7:   7   232 0.113614104
#>  8:   8   252 0.123408423
#>  9:   9   217 0.106268364
#> 10:  10    96 0.047012733
#> 11:  11    12 0.005876592
#> 
#> $wtd
#>     age count        prop    est
#>  1:   1    80 0.035449836  36529
#>  2:   2   187 0.094923344  97813
#>  3:   3    30 0.015961112  16447
#>  4:   4   118 0.056409774  58127
#>  5:   5   454 0.217321305 223937
#>  6:   6   364 0.183085511 188659
#>  7:   7   232 0.110731123 114102
#>  8:   8   252 0.124048709 127825
#>  9:   9   217 0.108907634 112223
#> 10:  10    96 0.046645032  48065
#> 11:  11    12 0.006516621   6715
#> 
#> $weight_name
#> [1] "person_weight"
#> 
                
DT = hts_prep_data(summarize_var = 'race',
                   summarize_by = 'employment',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
#> Creating a summary of person race broken down by person employment
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'race',
                summarize_by = 'employment',
                wtname = 'person_weight',
                checkbox_valname = 'value',
                checkbox_yesval = 1)
#> $unwtd
#>     employment                                      race count        prop
#>  1:          1                 African American or Black    33 0.035143770
#>  2:          1          American Indian or Alaska Native    12 0.012779553
#>  3:          1                                     Asian   214 0.227902023
#>  4:          1 Native Hawaiian or other Pacific Islander    10 0.010649627
#>  5:          1                                     White   550 0.585729499
#>  6:          1                                Other race    21 0.022364217
#>  7:          1                      Prefer not to answer    99 0.105431310
#>  8:          2                 African American or Black     9 0.056962025
#>  9:          2          American Indian or Alaska Native     1 0.006329114
#> 10:          2                                     Asian    18 0.113924051
#> 11:          2 Native Hawaiian or other Pacific Islander     3 0.018987342
#> 12:          2                                     White   107 0.677215190
#> 13:          2                                Other race     1 0.006329114
#> 14:          2                      Prefer not to answer    19 0.120253165
#> 15:          3                 African American or Black     1 0.011111111
#> 16:          3          American Indian or Alaska Native     1 0.011111111
#> 17:          3                                     Asian    13 0.144444444
#> 18:          3                                     White    67 0.744444444
#> 19:          3                      Prefer not to answer     8 0.088888889
#> 20:          5                 African American or Black    12 0.025586354
#> 21:          5          American Indian or Alaska Native    11 0.023454158
#> 22:          5                                     Asian    70 0.149253731
#> 23:          5 Native Hawaiian or other Pacific Islander     4 0.008528785
#> 24:          5                                     White   314 0.669509595
#> 25:          5                                Other race     8 0.017057569
#> 26:          5                      Prefer not to answer    50 0.106609808
#> 27:          6                 African American or Black    10 0.126582278
#> 28:          6          American Indian or Alaska Native     3 0.037974684
#> 29:          6                                     Asian    13 0.164556962
#> 30:          6                                     White    43 0.544303797
#> 31:          6                                Other race     2 0.025316456
#> 32:          6                      Prefer not to answer     8 0.101265823
#> 33:          7                 African American or Black     3 0.157894737
#> 34:          7          American Indian or Alaska Native     2 0.105263158
#> 35:          7                                     Asian     3 0.157894737
#> 36:          7                                     White    10 0.526315789
#> 37:          7                                Other race     1 0.052631579
#> 38:          8                 African American or Black     2 0.153846154
#> 39:          8                                     Asian     1 0.076923077
#> 40:          8 Native Hawaiian or other Pacific Islander     1 0.076923077
#> 41:          8                                     White     5 0.384615385
#> 42:          8                                Other race     1 0.076923077
#> 43:          8                      Prefer not to answer     3 0.230769231
#>     employment                                      race count        prop
#> 
#> $wtd
#>     employment                                      race count        prop
#>  1:          1                 African American or Black    33 0.039852519
#>  2:          1          American Indian or Alaska Native    12 0.012715506
#>  3:          1                                     Asian   214 0.213507915
#>  4:          1 Native Hawaiian or other Pacific Islander    10 0.010313354
#>  5:          1                                     White   550 0.587831102
#>  6:          1                                Other race    21 0.021503343
#>  7:          1                      Prefer not to answer    99 0.114276260
#>  8:          2                 African American or Black     9 0.061423340
#>  9:          2          American Indian or Alaska Native     1 0.009333809
#> 10:          2                                     Asian    18 0.095426962
#> 11:          2 Native Hawaiian or other Pacific Islander     3 0.011330134
#> 12:          2                                     White   107 0.696809847
#> 13:          2                                Other race     1 0.001546821
#> 14:          2                      Prefer not to answer    19 0.124129087
#> 15:          3                 African American or Black     1 0.005067356
#> 16:          3          American Indian or Alaska Native     1 0.016223881
#> 17:          3                                     Asian    13 0.146035784
#> 18:          3                                     White    67 0.739646328
#> 19:          3                      Prefer not to answer     8 0.093026651
#> 20:          5                 African American or Black    12 0.030971790
#> 21:          5          American Indian or Alaska Native    11 0.027242159
#> 22:          5                                     Asian    70 0.135451519
#> 23:          5 Native Hawaiian or other Pacific Islander     4 0.011485094
#> 24:          5                                     White   314 0.670586812
#> 25:          5                                Other race     8 0.016741621
#> 26:          5                      Prefer not to answer    50 0.107521005
#> 27:          6                 African American or Black    10 0.108340051
#> 28:          6          American Indian or Alaska Native     3 0.026655918
#> 29:          6                                     Asian    13 0.188749902
#> 30:          6                                     White    43 0.553246821
#> 31:          6                                Other race     2 0.019894417
#> 32:          6                      Prefer not to answer     8 0.103112891
#> 33:          7                 African American or Black     3 0.158050047
#> 34:          7          American Indian or Alaska Native     2 0.105878187
#> 35:          7                                     Asian     3 0.093130312
#> 36:          7                                     White    10 0.610717658
#> 37:          7                                Other race     1 0.032223796
#> 38:          8                 African American or Black     2 0.107026524
#> 39:          8                                     Asian     1 0.080153560
#> 40:          8 Native Hawaiian or other Pacific Islander     1 0.103420195
#> 41:          8                                     White     5 0.413331782
#> 42:          8                                Other race     1 0.102838530
#> 43:          8                      Prefer not to answer     3 0.193229409
#>     employment                                      race count        prop
#>        est
#>  1:  18548
#>  2:   5918
#>  3:  99370
#>  4:   4800
#>  5: 273586
#>  6:  10008
#>  7:  53186
#>  8:   4646
#>  9:    706
#> 10:   7218
#> 11:    857
#> 12:  52706
#> 13:    117
#> 14:   9389
#> 15:    243
#> 16:    778
#> 17:   7003
#> 18:  35469
#> 19:   4461
#> 20:   7424
#> 21:   6530
#> 22:  32468
#> 23:   2753
#> 24: 160741
#> 25:   4013
#> 26:  25773
#> 27:   4166
#> 28:   1025
#> 29:   7258
#> 30:  21274
#> 31:    765
#> 32:   3965
#> 33:   1339
#> 34:    897
#> 35:    789
#> 36:   5174
#> 37:    273
#> 38:    920
#> 39:    689
#> 40:    889
#> 41:   3553
#> 42:    884
#> 43:   1661
#>        est
#> 
#> $weight_name
#> [1] "person_weight"
#> 
                
DT = hts_prep_data(summarize_var = 'employment',
                summarize_by = c('race', 'income_detailed', 'gender'),
                variables_dt = variable_list,
                data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))$cat
#> Creating a summary of person employment broken down by person race and hh income_detailed and person gender

hts_summary_cat(prepped_dt = DT,
                summarize_var = 'employment',
                summarize_by = c('race', 'income_detailed', 'gender'),
                wtname = 'person_weight')$unwtd
#>                           race income_detailed gender employment count
#>   1: African American or Black               1      1          5     1
#>   2: African American or Black               1      2          6     1
#>   3: African American or Black               1      2          8     1
#>   4: African American or Black               1    999          6     1
#>   5: African American or Black               2      1          1     1
#>  ---                                                                  
#> 393:      Prefer not to answer             999      2          8     1
#> 394:      Prefer not to answer             999    999          1    22
#> 395:      Prefer not to answer             999    999          3     4
#> 396:      Prefer not to answer             999    999          5     6
#> 397:      Prefer not to answer             999    999          8     1
#>            prop
#>   1: 1.00000000
#>   2: 0.50000000
#>   3: 0.50000000
#>   4: 1.00000000
#>   5: 0.20000000
#>  ---           
#> 393: 0.05882353
#> 394: 0.66666667
#> 395: 0.12121212
#> 396: 0.18181818
#> 397: 0.03030303