Summarize a categorical variable

Summarize a categorical variable

Usage

hts_summary_cat(
  prepped_dt,
  summarize_var = NULL,
  summarize_by = NULL,
  weighted = TRUE,
  se = FALSE,
  wtname = NULL,
  strataname = NULL,
  checkbox_valname = NULL,
  checkbox_yesval = NULL
)

Arguments

prepped_dt

Dataset containing the summary variables and key columns in data.table format.

summarize_var

Name of the categorical variable to summarize. Default is NULL.

summarize_by

Name of the variable to summarize the summarize_var by. Default is NULL.

weighted

Whether the data is weighted. Default is TRUE.

se

Whether to calculate standard error. Default is FALSE.

wtname

Name of the weight column to use. Default is NULL.

strataname

Name of strata name to bring in. Default is NULL.

checkbox_valname

Name of the column with the checkbox value. Default is NULL. Must be provided if summarize_var is a checkbox variable.

checkbox_yesval

Value of checkbox_valname that indicates it was selected. Default is NULL. Must be provided if summarize_var is a checkbox variable.

Value

List of unweighted and weighted categorical summaries including counts and proportions.

Examples

require(data.table)
require(stringr)
require(dplyr)
require(srvyr)
DT = hts_prep_data(summarize_var = 'age',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
#> Creating a summary of person age 
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'age',
                wtname = 'person_weight')
#> $unwtd
#>     age count        prop
#>  1:   1    80 0.039177277
#>  2:   2   187 0.091576885
#>  3:   3    30 0.014691479
#>  4:   4   118 0.057786484
#>  5:   5   454 0.222331048
#>  6:   6   364 0.178256611
#>  7:   7   232 0.113614104
#>  8:   8   252 0.123408423
#>  9:   9   217 0.106268364
#> 10:  10    96 0.047012733
#> 11:  11    12 0.005876592
#> 
#> $wtd
#>     age count        prop    est
#>  1:   1    80 0.035449836  36529
#>  2:   2   187 0.094923344  97813
#>  3:   3    30 0.015961112  16447
#>  4:   4   118 0.056409774  58127
#>  5:   5   454 0.217321305 223937
#>  6:   6   364 0.183085511 188659
#>  7:   7   232 0.110731123 114102
#>  8:   8   252 0.124048709 127825
#>  9:   9   217 0.108907634 112223
#> 10:  10    96 0.046645032  48065
#> 11:  11    12 0.006516621   6715
#> 
#> $weight_name
#> [1] "person_weight"
#> 
DT = hts_prep_data(summarize_var = 'age',
                   summarize_by = 'employment',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
#> Creating a summary of person age broken down by person employment
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'age',
                summarize_by = 'employment',
                wtname = 'person_weight')
#> $unwtd
#>     employment age count        prop
#>  1:          1   4    37 0.039827772
#>  2:          1   5   345 0.371367061
#>  3:          1   6   264 0.284176534
#>  4:          1   7   146 0.157158235
#>  5:          1   8   116 0.124865447
#>  6:          1   9    19 0.020452099
#>  7:          1  10     2 0.002152853
#>  8:          2   3     7 0.040935673
#>  9:          2   4    34 0.198830409
#> 10:          2   5    31 0.181286550
#> 11:          2   6    31 0.181286550
#> 12:          2   7    19 0.111111111
#> 13:          2   8    22 0.128654971
#> 14:          2   9    22 0.128654971
#> 15:          2  10     5 0.029239766
#> 16:          3   4     1 0.011111111
#> 17:          3   5    15 0.166666667
#> 18:          3   6    22 0.244444444
#> 19:          3   7    20 0.222222222
#> 20:          3   8    23 0.255555556
#> 21:          3   9     9 0.100000000
#> 22:          5   3    21 0.043933054
#> 23:          5   4    33 0.069037657
#> 24:          5   5    27 0.056485356
#> 25:          5   6    24 0.050209205
#> 26:          5   7    30 0.062761506
#> 27:          5   8    80 0.167364017
#> 28:          5   9   163 0.341004184
#> 29:          5  10    88 0.184100418
#> 30:          5  11    12 0.025104603
#> 31:          6   3     2 0.024691358
#> 32:          6   4    10 0.123456790
#> 33:          6   5    29 0.358024691
#> 34:          6   6    15 0.185185185
#> 35:          6   7    15 0.185185185
#> 36:          6   8     8 0.098765432
#> 37:          6   9     1 0.012345679
#> 38:          6  10     1 0.012345679
#> 39:          7   4     2 0.153846154
#> 40:          7   5     3 0.230769231
#> 41:          7   6     3 0.230769231
#> 42:          7   7     2 0.153846154
#> 43:          7   8     1 0.076923077
#> 44:          7   9     2 0.153846154
#> 45:          8   4     1 0.076923077
#> 46:          8   5     4 0.307692308
#> 47:          8   6     5 0.384615385
#> 48:          8   8     2 0.153846154
#> 49:          8   9     1 0.076923077
#> 50:        995   1    80 0.299625468
#> 51:        995   2   187 0.700374532
#>     employment age count        prop
#> 
#> $wtd
#>     employment age count        prop    est
#>  1:          1   4    37 0.043813401  20360
#>  2:          1   5   345 0.365732153 169955
#>  3:          1   6   264 0.289476176 134519
#>  4:          1   7   146 0.151894349  70585
#>  5:          1   8   116 0.126103405  58600
#>  6:          1   9    19 0.020949090   9735
#>  7:          1  10     2 0.002031427    944
#>  8:          2   3     7 0.044806200   3735
#>  9:          2   4    34 0.185822767  15490
#> 10:          2   5    31 0.143319857  11947
#> 11:          2   6    31 0.214313991  17865
#> 12:          2   7    19 0.109598244   9136
#> 13:          2   8    22 0.159730803  13315
#> 14:          2   9    22 0.121546564  10132
#> 15:          2  10     5 0.020861575   1739
#> 16:          3   4     1 0.006874947    327
#> 17:          3   5    15 0.166007905   7896
#> 18:          3   6    22 0.202590194   9636
#> 19:          3   7    20 0.226641998  10780
#> 20:          3   8    23 0.267303002  12714
#> 21:          3   9     9 0.130581953   6211
#> 22:          5   3    21 0.046289050  11298
#> 23:          5   4    33 0.061272150  14955
#> 24:          5   5    27 0.064517054  15747
#> 25:          5   6    24 0.054749565  13363
#> 26:          5   7    30 0.066000205  16109
#> 27:          5   8    80 0.152178634  37143
#> 28:          5   9   163 0.345020998  84211
#> 29:          5  10    88 0.182460309  44534
#> 30:          5  11    12 0.027512035   6715
#> 31:          6   3     2 0.034445798   1414
#> 32:          6   4    10 0.140414129   5764
#> 33:          6   5    29 0.342679659  14067
#> 34:          6   6    15 0.177880633   7302
#> 35:          6   7    15 0.169987820   6978
#> 36:          6   8     8 0.095907430   3937
#> 37:          6   9     1 0.018026797    740
#> 38:          6  10     1 0.020657734    848
#> 39:          7   4     2 0.154918668   1000
#> 40:          7   5     3 0.191789311   1238
#> 41:          7   6     3 0.368706429   2380
#> 42:          7   7     2 0.079628195    514
#> 43:          7   8     1 0.119287374    770
#> 44:          7   9     2 0.085670023    553
#> 45:          8   4     1 0.025957973    231
#> 46:          8   5     4 0.346892909   3087
#> 47:          8   6     5 0.403865603   3594
#> 48:          8   8     2 0.151252950   1346
#> 49:          8   9     1 0.072030565    641
#> 50:        995   1    80 0.271910497  36529
#> 51:        995   2   187 0.728089503  97813
#>     employment age count        prop    est
#> 
#> $weight_name
#> [1] "person_weight"
#> 
                
DT = hts_prep_data(summarize_var = 'employment',
                summarize_by = c('race', 'income_detailed', 'gender'),
                variables_dt = variable_list,
                data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))$cat
#> Creating a summary of person employment broken down by person race and hh income_detailed and person gender

hts_summary_cat(prepped_dt = DT,
                summarize_var = 'employment',
                summarize_by = c('race', 'income_detailed', 'gender'),
                wtname = 'person_weight')$unwtd
#>                           race income_detailed gender employment count
#>   1: African American or Black               1      1          5     1
#>   2: African American or Black               1      2          6     1
#>   3: African American or Black               1      2          8     1
#>   4: African American or Black               1    999          6     1
#>   5: African American or Black               2      1          1     1
#>  ---                                                                  
#> 393:      Prefer not to answer             999      2          8     1
#> 394:      Prefer not to answer             999    999          1    22
#> 395:      Prefer not to answer             999    999          3     4
#> 396:      Prefer not to answer             999    999          5     6
#> 397:      Prefer not to answer             999    999          8     1
#>            prop
#>   1: 1.00000000
#>   2: 0.50000000
#>   3: 0.50000000
#>   4: 1.00000000
#>   5: 0.20000000
#>  ---           
#> 393: 0.05882353
#> 394: 0.66666667
#> 395: 0.12121212
#> 396: 0.18181818
#> 397: 0.03030303