Summarize a categorical variable

Summarize a categorical variable

Usage

hts_summary_cat(
  prepped_dt,
  summarize_var = NULL,
  summarize_by = NULL,
  weighted = TRUE,
  se = FALSE,
  wtname = NULL,
  strataname = NULL,
  checkbox_valname = "value",
  checkbox_yesval = 1,
  summarize_vartype = "categorical"
)

Arguments

prepped_dt

Dataset containing the summary variables and key columns in data.table format.

summarize_var

Name of the categorical variable to summarize. Default is NULL.

summarize_by

Name of the variable to summarize the summarize_var by. Default is NULL.

weighted

Whether the data is weighted. Default is TRUE.

se

Whether to calculate standard error. Default is FALSE.

wtname

Name of the weight column to use. Default is NULL.

strataname

Name of strata name to bring in. Default is NULL.

checkbox_valname

Name of the column with the checkbox value. Default is 'value'. Must be provided if summarize_var is a checkbox variable.

checkbox_yesval

Value of checkbox_valname that indicates it was selected. Default is 1. Must be provided if summarize_var is a checkbox variable.

summarize_vartype

String; one of either 'categorical' (when the variable being summarized is categorical) or 'checkbox' (when the variable being summarized is derived from a multiple response, aka select-all-that-apply question).

Value

List of unweighted and weighted categorical summaries including counts and proportions.

Examples

require(data.table)
require(stringr)
require(dplyr)
require(srvyr)
DT = hts_prep_data(summarize_var = 'age',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
#> Creating a summary of person age 
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'age',
                wtname = 'person_weight')
#> $unwtd
#>     age count        prop
#>  1:   1    90 0.044334975
#>  2:   2   182 0.089655172
#>  3:   3    24 0.011822660
#>  4:   4   114 0.056157635
#>  5:   5   451 0.222167488
#>  6:   6   347 0.170935961
#>  7:   7   235 0.115763547
#>  8:   8   261 0.128571429
#>  9:   9   214 0.105418719
#> 10:  10    98 0.048275862
#> 11:  11    14 0.006896552
#> 
#> $wtd
#>     age count        prop    est
#>  1:   1    90 0.048069348  48998
#>  2:   2   182 0.090921488  92678
#>  3:   3    24 0.012522086  12764
#>  4:   4   114 0.064091810  65330
#>  5:   5   451 0.215398712 219560
#>  6:   6   347 0.181303400 184806
#>  7:   7   235 0.114738369 116955
#>  8:   8   261 0.116686729 118941
#>  9:   9   214 0.100840856 102789
#> 10:  10    98 0.046158268  47050
#> 11:  11    14 0.009268933   9448
#> 
#> $weight_name
#> [1] "person_weight"
#> 
                
DT = hts_prep_data(summarize_var = 'race',
                   summarize_by = 'employment',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$cat
#> Creating a summary of person race broken down by person employment
hts_summary_cat(prepped_dt = DT,
                summarize_var = 'race',
                summarize_by = 'employment',
                wtname = 'person_weight',
                summarize_vartype = 'checkbox')
#> $unwtd
#>     employment                                      race count        prop
#>  1:          1                 African American or Black    22 0.024070022
#>  2:          1          American Indian or Alaska Native    11 0.012035011
#>  3:          1                                     Asian   186 0.203501094
#>  4:          1 Native Hawaiian or other Pacific Islander    11 0.012035011
#>  5:          1                                     White   575 0.629102845
#>  6:          1                                Other race    17 0.018599562
#>  7:          1                      Prefer not to answer    92 0.100656455
#>  8:          2                 African American or Black     6 0.047244094
#>  9:          2                                     Asian    20 0.157480315
#> 10:          2 Native Hawaiian or other Pacific Islander     4 0.031496063
#> 11:          2                                     White    82 0.645669291
#> 12:          2                      Prefer not to answer    15 0.118110236
#> 13:          3                 African American or Black     1 0.008547009
#> 14:          3          American Indian or Alaska Native     4 0.034188034
#> 15:          3                                     Asian    16 0.136752137
#> 16:          3 Native Hawaiian or other Pacific Islander     1 0.008547009
#> 17:          3                                     White    80 0.683760684
#> 18:          3                                Other race     3 0.025641026
#> 19:          3                      Prefer not to answer    12 0.102564103
#> 20:          5                 African American or Black    13 0.027600849
#> 21:          5          American Indian or Alaska Native    10 0.021231423
#> 22:          5                                     Asian    55 0.116772824
#> 23:          5 Native Hawaiian or other Pacific Islander     5 0.010615711
#> 24:          5                                     White   326 0.692144374
#> 25:          5                                Other race     4 0.008492569
#> 26:          5                      Prefer not to answer    58 0.123142251
#> 27:          6                 African American or Black     8 0.097560976
#> 28:          6          American Indian or Alaska Native     1 0.012195122
#> 29:          6                                     Asian    11 0.134146341
#> 30:          6 Native Hawaiian or other Pacific Islander     1 0.012195122
#> 31:          6                                     White    49 0.597560976
#> 32:          6                                Other race     1 0.012195122
#> 33:          6                      Prefer not to answer    11 0.134146341
#> 34:          7                 African American or Black     2 0.100000000
#> 35:          7          American Indian or Alaska Native     2 0.100000000
#> 36:          7                                     Asian     3 0.150000000
#> 37:          7                                     White    10 0.500000000
#> 38:          7                                Other race     1 0.050000000
#> 39:          7                      Prefer not to answer     2 0.100000000
#> 40:          8                 African American or Black     1 0.055555556
#> 41:          8          American Indian or Alaska Native     1 0.055555556
#> 42:          8                                     Asian     4 0.222222222
#> 43:          8                                     White    11 0.611111111
#> 44:          8                      Prefer not to answer     1 0.055555556
#>     employment                                      race count        prop
#> 
#> $wtd
#>     employment                                      race count        prop
#>  1:          1                 African American or Black    22 0.021638708
#>  2:          1          American Indian or Alaska Native    11 0.013179012
#>  3:          1                                     Asian   186 0.203922848
#>  4:          1 Native Hawaiian or other Pacific Islander    11 0.012823843
#>  5:          1                                     White   575 0.636004431
#>  6:          1                                Other race    17 0.020433351
#>  7:          1                      Prefer not to answer    92 0.091997807
#>  8:          2                 African American or Black     6 0.041024015
#>  9:          2                                     Asian    20 0.172465721
#> 10:          2 Native Hawaiian or other Pacific Islander     4 0.035903939
#> 11:          2                                     White    82 0.642799398
#> 12:          2                      Prefer not to answer    15 0.107806927
#> 13:          3                 African American or Black     1 0.007079765
#> 14:          3          American Indian or Alaska Native     4 0.025402872
#> 15:          3                                     Asian    16 0.135779786
#> 16:          3 Native Hawaiian or other Pacific Islander     1 0.011968175
#> 17:          3                                     White    80 0.664081990
#> 18:          3                                Other race     3 0.025065741
#> 19:          3                      Prefer not to answer    12 0.130621671
#> 20:          5                 African American or Black    13 0.024707040
#> 21:          5          American Indian or Alaska Native    10 0.023532166
#> 22:          5                                     Asian    55 0.120088614
#> 23:          5 Native Hawaiian or other Pacific Islander     5 0.017011831
#> 24:          5                                     White   326 0.673532383
#> 25:          5                                Other race     4 0.009823856
#> 26:          5                      Prefer not to answer    58 0.131304110
#> 27:          6                 African American or Black     8 0.087900372
#> 28:          6          American Indian or Alaska Native     1 0.018215225
#> 29:          6                                     Asian    11 0.111809273
#> 30:          6 Native Hawaiian or other Pacific Islander     1 0.019054532
#> 31:          6                                     White    49 0.618886671
#> 32:          6                                Other race     1 0.006238091
#> 33:          6                      Prefer not to answer    11 0.137895835
#> 34:          7                 African American or Black     2 0.050441655
#> 35:          7          American Indian or Alaska Native     2 0.095536960
#> 36:          7                                     Asian     3 0.183054393
#> 37:          7                                     White    10 0.523128777
#> 38:          7                                Other race     1 0.001743375
#> 39:          7                      Prefer not to answer     2 0.146094840
#> 40:          8                 African American or Black     1 0.105333038
#> 41:          8          American Indian or Alaska Native     1 0.093383492
#> 42:          8                                     Asian     4 0.215866342
#> 43:          8                                     White    11 0.572914362
#> 44:          8                      Prefer not to answer     1 0.012502766
#>     employment                                      race count        prop
#>        est
#>  1:   9748
#>  2:   5937
#>  3:  91865
#>  4:   5777
#>  5: 286513
#>  6:   9205
#>  7:  41444
#>  8:   2588
#>  9:  10880
#> 10:   2265
#> 11:  40551
#> 12:   6801
#> 13:    420
#> 14:   1507
#> 15:   8055
#> 16:    710
#> 17:  39396
#> 18:   1487
#> 19:   7749
#> 20:   5699
#> 21:   5428
#> 22:  27700
#> 23:   3924
#> 24: 155359
#> 25:   2266
#> 26:  30287
#> 27:   3875
#> 28:    803
#> 29:   4929
#> 30:    840
#> 31:  27283
#> 32:    275
#> 33:   6079
#> 34:    434
#> 35:    822
#> 36:   1575
#> 37:   4501
#> 38:     15
#> 39:   1257
#> 40:    952
#> 41:    844
#> 42:   1951
#> 43:   5178
#> 44:    113
#>        est
#> 
#> $weight_name
#> [1] "person_weight"
#> 
                
DT = hts_prep_data(summarize_var = 'employment',
                summarize_by = c('race', 'income_detailed', 'gender'),
                variables_dt = variable_list,
                data = list('hh' = hh,
                            'person' = person,
                            'day' = day,
                            'trip' = trip,
                            'vehicle' = vehicle))$cat
#> Creating a summary of person employment broken down by person race and hh income_detailed and person gender

hts_summary_cat(prepped_dt = DT,
                summarize_var = 'employment',
                summarize_by = c('race', 'income_detailed', 'gender'),
                wtname = 'person_weight')$unwtd
#>                           race income_detailed gender employment count
#>   1: African American or Black               1      1          5     2
#>   2: African American or Black               2      1          1     1
#>   3: African American or Black               2      1          5     2
#>   4: African American or Black               2      1          6     2
#>   5: African American or Black               2      1          7     1
#>  ---                                                                  
#> 392:      Prefer not to answer             999    999          2     2
#> 393:      Prefer not to answer             999    999          3     1
#> 394:      Prefer not to answer             999    999          5    19
#> 395:      Prefer not to answer             999    999          6     1
#> 396:      Prefer not to answer             999    999          7     1
#>            prop
#>   1: 1.00000000
#>   2: 0.16666667
#>   3: 0.33333333
#>   4: 0.33333333
#>   5: 0.16666667
#>  ---           
#> 392: 0.05405405
#> 393: 0.02702703
#> 394: 0.51351351
#> 395: 0.02702703
#> 396: 0.02702703