Summarize a numeric variable

Summarize a numeric variable

Usage

hts_summary_num(
  prepped_dt,
  summarize_var = NULL,
  summarize_by = NULL,
  weighted = TRUE,
  se = FALSE,
  wtname = NULL,
  strataname = NULL
)

Arguments

prepped_dt

A prepared dataset in data.table format with the variable to summarize, the variable to summarize by, and the weights, if used.

summarize_var

Name of the variable to summarize. Default is NULL.

summarize_by

Name of the variable to summarize the summarize_var by. Default is NULL.

weighted

Whether the data is weighted. Default is TRUE.

se

Whether to calculate standard error. Default is FALSE. Will be set to FALSE if weighted is FALSE.

wtname

Name of the weight column to use. Default is NULL. Must be specified when weighted = TRUE.

strataname

Name of strata name to bring in. Default is NULL.

Value

List of unweighted and weighted numeric summaries including count, min, max, mean, se, and median.

Examples

require(data.table)
require(stringr)
require(dplyr)
require(srvyr)
DT = hts_prep_data(summarize_var = 'speed_mph',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$num
#> Creating a summary of trip speed_mph 
hts_summary_num(prepped_dt = DT,
                summarize_var = 'speed_mph',
                wtname = 'trip_weight')
#> $unwtd
#>    count min      max     mean   median
#> 1: 16283   0 123.3683 12.71216 9.213738
#> 
#> $wtd
#>    count min      max     mean   median
#> 1: 16283   0 123.3683 12.77496 9.335348
#> 
#> $weight_name
#> [1] "trip_weight"
#> 
DT = hts_prep_data(summarize_var = 'speed_mph',
                   summarize_by = 'age', 
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$num
#> Creating a summary of trip speed_mph broken down by person age
hts_summary_num(prepped_dt = DT,
                summarize_var = 'speed_mph',
                summarize_by = 'age',
                wtname = 'trip_weight')
#> $unwtd
#>     age count         min       max      mean    median
#>  1:   1   431 0.114271649 119.93525 12.881785 10.088685
#>  2:   2   855 0.000000000  99.06068 12.598717  9.111278
#>  3:   3    71 0.339308169  98.43097 13.007454  8.185791
#>  4:   4  1042 0.000000000 121.10244 13.917912 10.286320
#>  5:   5  4814 0.000000000 119.93525 11.756395  7.844880
#>  6:   6  3556 0.008117147 123.36831 12.788130  9.499362
#>  7:   7  1811 0.000000000 106.72662 13.441124 10.112949
#>  8:   8  1923 0.015928992 120.84837 13.803242 10.340619
#>  9:   9  1241 0.000000000 118.12607 12.298467  9.393039
#> 10:  10   511 0.024460902 122.43479 13.380015 12.028255
#> 11:  11    28 0.000000000  33.39149  6.683246  1.347399
#> 
#> $wtd
#>     age count         min       max      mean     median
#>  1:   1   431 0.114271649 119.93525 13.472854 10.5119474
#>  2:   2   855 0.000000000  99.06068 13.063057  9.1689195
#>  3:   3    71 0.339308169  98.43097 14.671470  8.1857913
#>  4:   4  1042 0.000000000 121.10244 13.992131 10.3692066
#>  5:   5  4814 0.000000000 119.93525 11.800110  7.9223551
#>  6:   6  3556 0.008117147 123.36831 12.863970  9.6812125
#>  7:   7  1811 0.000000000 106.72662 13.225017 10.1207500
#>  8:   8  1923 0.015928992 120.84837 13.844358 10.3730573
#>  9:   9  1241 0.000000000 118.12607 12.211063  9.4153208
#> 10:  10   511 0.024460902 122.43479 13.446605 11.9591687
#> 11:  11    28 0.000000000  33.39149  7.170318  0.9284994
#> 
#> $weight_name
#> [1] "trip_weight"
#>