Summarize a numeric variable

Summarize a numeric variable

Usage

hts_summary_num(
  prepped_dt,
  summarize_var = NULL,
  summarize_by = NULL,
  weighted = TRUE,
  se = FALSE,
  wtname = NULL,
  strataname = NULL
)

Arguments

prepped_dt

A prepared dataset in data.table format with the variable to summarize, the variable to summarize by, and the weights, if used.

summarize_var

Name of the variable to summarize. Default is NULL.

summarize_by

Name of the variable to summarize the summarize_var by. Default is NULL.

weighted

Whether the data is weighted. Default is TRUE.

se

Whether to calculate standard error. Default is FALSE. Will be set to FALSE if weighted is FALSE.

wtname

Name of the weight column to use. Default is NULL. Must be specified when weighted = TRUE.

strataname

Name of strata name to bring in. Default is NULL.

Value

List of unweighted and weighted numeric summaries including count, min, max, mean, se, and median.

Examples

require(data.table)
require(stringr)
require(dplyr)
require(srvyr)
DT = hts_prep_data(summarize_var = 'speed_mph',
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$num
#> Creating a summary of trip speed_mph 
hts_summary_num(prepped_dt = DT,
                summarize_var = 'speed_mph',
                wtname = 'trip_weight')
#> $unwtd
#>    count min      max     mean  median
#> 1: 14685   0 112.9918 11.83172 8.63421
#> 
#> $wtd
#>    count min      max    mean   median
#> 1: 14685   0 112.9918 11.8349 8.618818
#> 
#> $weight_name
#> [1] "trip_weight"
#> 
DT = hts_prep_data(summarize_var = 'speed_mph',
                   summarize_by = 'age', 
                   variables_dt = variable_list,
                   data = list('hh' = hh,
                               'person' = person,
                               'day' = day,
                               'trip' = trip,
                               'vehicle' = vehicle))$num
#> Creating a summary of trip speed_mph broken down by person age
hts_summary_num(prepped_dt = DT,
                summarize_var = 'speed_mph',
                summarize_by = 'age',
                wtname = 'trip_weight')
#> $unwtd
#>     age count        min       max      mean    median
#>  1:   1   296 0.08714984 101.97575 11.133769  7.385581
#>  2:   2   743 0.11910113 109.83115 10.434704  8.880725
#>  3:   3   101 0.00000000  82.63329 13.845243 11.990414
#>  4:   4   980 0.00000000 112.99185  9.535263  5.138501
#>  5:   5  3006 0.00000000 108.02176 11.713822  7.568860
#>  6:   6  3518 0.00000000 112.53715 10.896587  7.745927
#>  7:   7  1825 0.00000000 112.37068 13.201686 10.388775
#>  8:   8  1759 0.00000000 108.05151 13.253130 10.530855
#>  9:   9  1767 0.00000000 107.93792 12.783168  9.682409
#> 10:  10   674 0.00000000  63.09036 12.136437  9.821970
#> 11:  11    16 0.82500585  88.79454 14.887919  6.119036
#> 
#> $wtd
#>     age count        min       max      mean    median
#>  1:   1   296 0.08714984 101.97575 11.757763  7.713921
#>  2:   2   743 0.11910113 109.83115 10.448693  9.024466
#>  3:   3   101 0.00000000  82.63329 14.582794 12.965059
#>  4:   4   980 0.00000000 112.99185  9.090289  4.795173
#>  5:   5  3006 0.00000000 108.02176 11.616469  7.425500
#>  6:   6  3518 0.00000000 112.53715 10.985165  7.769085
#>  7:   7  1825 0.00000000 112.37068 12.990898 10.022160
#>  8:   8  1759 0.00000000 108.05151 13.138908 10.593106
#>  9:   9  1767 0.00000000 107.93792 13.124612 10.114323
#> 10:  10   674 0.00000000  63.09036 12.642415 10.661429
#> 11:  11    16 0.82500585  88.79454 15.323818  6.541481
#> 
#> $weight_name
#> [1] "trip_weight"
#>