Make household travel survey summaries-- runs hts_prep_variable and hts_summary

Make household travel survey summaries-- runs hts_prep_variable and hts_summary

Usage

hts_summary_wrapper(
  summarize_var = NULL,
  summarize_by = NULL,
  variables_dt = variable_list,
  vals_df = value_labels,
  data = list(hh = hh, person = person, day = day, trip = trip, vehicle = vehicle),
  id_cols = c("hh_id", "person_id", "day_id", "trip_id", "vehicle_id"),
  weighted = TRUE,
  wt_cols = c("hh_weight", "person_weight", "day_weight", "trip_weight", "hh_weight"),
  trip_name = "trip",
  day_name = "day",
  strataname = NULL,
  se = FALSE,
  wtname = NULL,
  checkbox_valname = "value",
  checkbox_yesval = 1,
  value_label_colname = "label",
  remove_outliers = TRUE,
  threshold = 0.975,
  remove_missing = TRUE,
  not_imputable = -1,
  missing_values = c("Missing Response", "995")
)

Arguments

summarize_var

Name of the variable to summarize. Default is NULL

summarize_by

Name of the variable to summarize the summarize_var by. Default is NULL.

variables_dt

List of variable locations and descriptions in data.table format.

vals_df

A dataframe of variable labels (i.e., factor levels and labels) with the format as specified below. Passed to factorize_column function.

data

List of household, person, vehicle, day, and trip tables in data.table format.

id_cols

name of unique identifier for each table in hts_data

weighted

Whether the data is weighted. Default is TRUE.

wt_cols

weight name for each table in hts_data

trip_name

Name of the trip dataset in hts_data.

day_name

Name of the day dataset in hts_data.

strataname

Name of strata name to bring in. Default is NULL.

se

Whether to calculate standard error. Default is FALSE. Will be set to FALSE if weighted is FALSE.

wtname

Name of the weight column to use. Default is NULL. Must be specified when weighted = TRUE.

checkbox_valname

Name of the column with the checkbox value. Default is 'value'. Must be provided if summarize_var is a checkbox variable.

checkbox_yesval

Value of checkbox_valname that indicates it was selected. Default is 1. Must be provided if summarize_var is a checkbox variable.

value_label_colname

The name of the value label column in vals_df. Default is 'value_label'

remove_outliers

Whether to remove outliers for numeric variable. Default is TRUE.

threshold

Threshold to define outliers. Default is 0.975.

remove_missing

Whether to remove missing values from the summary. Default is TRUE.

not_imputable

Value representing 'Not imputable' to remove. Default is -1.

missing_values

Missing values to remove. Default is 995.

Value

A list containing (if applicable) categorical and numeric summaries of the specified variable(s), as well as sample sizes and whether or not the summarized variable is a shared checkbox variable. To access the categorical/numeric df use output$summary. To access the weighted df use output$summary$wtd, and output$summary$unwtd for the unweighted df. To access the weight name use output$summary$weight_name. To access sample sizes use output$n_ls. To access weighted and unweighted sample sizes respectively, use output$n_ls$wtd and output$n_ls$unwtd.

Examples

hts_summary_wrapper(
summarize_var = 'employment',
summarize_by = 'age',
wtname = 'person_weight')
#> $cat
#> $cat$n_ls
#> $cat$n_ls$unwtd
#> $cat$n_ls$unwtd$`Count of unique hh_id`
#> [1] 843
#> 
#> $cat$n_ls$unwtd$`Count of unique person_id`
#> [1] 1770
#> 
#> 
#> $cat$n_ls$wtd
#> Sum of person_weight 
#>               902225 
#> 
#> 
#> $cat$summary
#> $cat$summary$unwtd
#>             age
#>           <ord>
#>  1:     Under 5
#>  2:     Under 5
#>  3:     Under 5
#>  4:     Under 5
#>  5:     Under 5
#>  6:     Under 5
#>  7:     Under 5
#>  8:        5-11
#>  9:        5-11
#> 10:        5-11
#> 11:        5-11
#> 12:        5-11
#> 13:        5-11
#> 14:        5-11
#> 15:       12-15
#> 16:       12-15
#> 17:       12-15
#> 18:       12-15
#> 19:       12-15
#> 20:       12-15
#> 21:       12-15
#> 22:       16-17
#> 23:       16-17
#> 24:       16-17
#> 25:       16-17
#> 26:       16-17
#> 27:       16-17
#> 28:       18-24
#> 29:       18-24
#> 30:       18-24
#> 31:       18-24
#> 32:       18-24
#> 33:       18-24
#> 34:       18-24
#> 35:       25-34
#> 36:       25-34
#> 37:       25-34
#> 38:       25-34
#> 39:       25-34
#> 40:       25-34
#> 41:       35-44
#> 42:       35-44
#> 43:       35-44
#> 44:       35-44
#> 45:       35-44
#> 46:       35-44
#> 47:       45-54
#> 48:       45-54
#> 49:       45-54
#> 50:       45-54
#> 51:       45-54
#> 52:       45-54
#> 53:       55-64
#> 54:       55-64
#> 55:       55-64
#> 56:       55-64
#> 57:       55-64
#> 58:       55-64
#> 59:       55-64
#> 60:       65-74
#> 61:       65-74
#> 62:       65-74
#> 63:       65-74
#> 64:       65-74
#> 65:       65-74
#> 66:       65-74
#> 67:       75-84
#> 68:       75-84
#> 69:       75-84
#> 70:       75-84
#> 71:       75-84
#> 72:       75-84
#> 73:       75-84
#> 74: 85 or older
#> 75: 85 or older
#> 76: 85 or older
#> 77: 85 or older
#> 78: 85 or older
#>             age
#>                                                                              employment
#>                                                                                   <ord>
#>  1:                                           Employed full-time (35+ hours/week, paid)
#>  2:                                 Employed part-time (fewer than 35 hours/week, paid)
#>  3:                                                                       Self-employed
#>  4: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#>  5:                                                     Unemployed and looking for work
#>  6:                                                          Unpaid volunteer or intern
#>  7:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#>  8:                                           Employed full-time (35+ hours/week, paid)
#>  9:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 10:                                                                       Self-employed
#> 11: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 12:                                                     Unemployed and looking for work
#> 13:                                                          Unpaid volunteer or intern
#> 14:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 15:                                           Employed full-time (35+ hours/week, paid)
#> 16:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 17:                                                                       Self-employed
#> 18: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 19:                                                     Unemployed and looking for work
#> 20:                                                          Unpaid volunteer or intern
#> 21:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 22:                                           Employed full-time (35+ hours/week, paid)
#> 23:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 24:                                                                       Self-employed
#> 25: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 26:                                                     Unemployed and looking for work
#> 27:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 28:                                           Employed full-time (35+ hours/week, paid)
#> 29:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 30:                                                                       Self-employed
#> 31: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 32:                                                     Unemployed and looking for work
#> 33:                                                          Unpaid volunteer or intern
#> 34:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 35:                                           Employed full-time (35+ hours/week, paid)
#> 36:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 37:                                                                       Self-employed
#> 38: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 39:                                                     Unemployed and looking for work
#> 40:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 41:                                           Employed full-time (35+ hours/week, paid)
#> 42:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 43:                                                                       Self-employed
#> 44: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 45:                                                     Unemployed and looking for work
#> 46:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 47:                                           Employed full-time (35+ hours/week, paid)
#> 48:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 49:                                                                       Self-employed
#> 50: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 51:                                                     Unemployed and looking for work
#> 52:                                                          Unpaid volunteer or intern
#> 53:                                           Employed full-time (35+ hours/week, paid)
#> 54:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 55:                                                                       Self-employed
#> 56: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 57:                                                     Unemployed and looking for work
#> 58:                                                          Unpaid volunteer or intern
#> 59:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 60:                                           Employed full-time (35+ hours/week, paid)
#> 61:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 62:                                                                       Self-employed
#> 63: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 64:                                                     Unemployed and looking for work
#> 65:                                                          Unpaid volunteer or intern
#> 66:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 67:                                           Employed full-time (35+ hours/week, paid)
#> 68:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 69:                                                                       Self-employed
#> 70: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 71:                                                     Unemployed and looking for work
#> 72:                                                          Unpaid volunteer or intern
#> 73:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 74:                                           Employed full-time (35+ hours/week, paid)
#> 75:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 76:                                                                       Self-employed
#> 77: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 78:                                                     Unemployed and looking for work
#>                                                                              employment
#>     count        prop
#>     <int>       <num>
#>  1:    68 0.444444444
#>  2:    16 0.104575163
#>  3:     9 0.058823529
#>  4:    50 0.326797386
#>  5:     8 0.052287582
#>  6:     1 0.006535948
#>  7:     1 0.006535948
#>  8:    73 0.540740741
#>  9:    12 0.088888889
#> 10:     8 0.059259259
#> 11:    36 0.266666667
#> 12:     4 0.029629630
#> 13:     1 0.007407407
#> 14:     1 0.007407407
#> 15:    77 0.538461538
#> 16:     6 0.041958042
#> 17:     8 0.055944056
#> 18:    37 0.258741259
#> 19:     7 0.048951049
#> 20:     4 0.027972028
#> 21:     4 0.027972028
#> 22:    82 0.546666667
#> 23:    12 0.080000000
#> 24:     3 0.020000000
#> 25:    43 0.286666667
#> 26:     9 0.060000000
#> 27:     1 0.006666667
#> 28:    80 0.522875817
#> 29:    10 0.065359477
#> 30:    11 0.071895425
#> 31:    42 0.274509804
#> 32:     6 0.039215686
#> 33:     3 0.019607843
#> 34:     1 0.006535948
#> 35:    70 0.503597122
#> 36:    14 0.100719424
#> 37:     6 0.043165468
#> 38:    42 0.302158273
#> 39:     5 0.035971223
#> 40:     2 0.014388489
#> 41:    78 0.516556291
#> 42:    10 0.066225166
#> 43:    10 0.066225166
#> 44:    44 0.291390728
#> 45:     8 0.052980132
#> 46:     1 0.006622517
#> 47:    79 0.484662577
#> 48:    15 0.092024540
#> 49:    13 0.079754601
#> 50:    51 0.312883436
#> 51:     4 0.024539877
#> 52:     1 0.006134969
#> 53:    69 0.496402878
#> 54:    14 0.100719424
#> 55:    11 0.079136691
#> 56:    36 0.258992806
#> 57:     5 0.035971223
#> 58:     1 0.007194245
#> 59:     3 0.021582734
#> 60:    68 0.515151515
#> 61:    10 0.075757576
#> 62:     9 0.068181818
#> 63:    32 0.242424242
#> 64:    10 0.075757576
#> 65:     2 0.015151515
#> 66:     1 0.007575758
#> 67:    82 0.512500000
#> 68:    15 0.093750000
#> 69:    13 0.081250000
#> 70:    40 0.250000000
#> 71:     8 0.050000000
#> 72:     1 0.006250000
#> 73:     1 0.006250000
#> 74:    71 0.467105263
#> 75:    17 0.111842105
#> 76:     9 0.059210526
#> 77:    48 0.315789474
#> 78:     7 0.046052632
#>     count        prop
#> 
#> $cat$summary$wtd
#>             age
#>           <ord>
#>  1:     Under 5
#>  2:     Under 5
#>  3:     Under 5
#>  4:     Under 5
#>  5:     Under 5
#>  6:     Under 5
#>  7:     Under 5
#>  8:        5-11
#>  9:        5-11
#> 10:        5-11
#> 11:        5-11
#> 12:        5-11
#> 13:        5-11
#> 14:        5-11
#> 15:       12-15
#> 16:       12-15
#> 17:       12-15
#> 18:       12-15
#> 19:       12-15
#> 20:       12-15
#> 21:       12-15
#> 22:       16-17
#> 23:       16-17
#> 24:       16-17
#> 25:       16-17
#> 26:       16-17
#> 27:       16-17
#> 28:       18-24
#> 29:       18-24
#> 30:       18-24
#> 31:       18-24
#> 32:       18-24
#> 33:       18-24
#> 34:       18-24
#> 35:       25-34
#> 36:       25-34
#> 37:       25-34
#> 38:       25-34
#> 39:       25-34
#> 40:       25-34
#> 41:       35-44
#> 42:       35-44
#> 43:       35-44
#> 44:       35-44
#> 45:       35-44
#> 46:       35-44
#> 47:       45-54
#> 48:       45-54
#> 49:       45-54
#> 50:       45-54
#> 51:       45-54
#> 52:       45-54
#> 53:       55-64
#> 54:       55-64
#> 55:       55-64
#> 56:       55-64
#> 57:       55-64
#> 58:       55-64
#> 59:       55-64
#> 60:       65-74
#> 61:       65-74
#> 62:       65-74
#> 63:       65-74
#> 64:       65-74
#> 65:       65-74
#> 66:       65-74
#> 67:       75-84
#> 68:       75-84
#> 69:       75-84
#> 70:       75-84
#> 71:       75-84
#> 72:       75-84
#> 73:       75-84
#> 74: 85 or older
#> 75: 85 or older
#> 76: 85 or older
#> 77: 85 or older
#> 78: 85 or older
#>             age
#>                                                                              employment
#>                                                                                   <ord>
#>  1:                                           Employed full-time (35+ hours/week, paid)
#>  2:                                 Employed part-time (fewer than 35 hours/week, paid)
#>  3:                                                                       Self-employed
#>  4: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#>  5:                                                     Unemployed and looking for work
#>  6:                                                          Unpaid volunteer or intern
#>  7:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#>  8:                                           Employed full-time (35+ hours/week, paid)
#>  9:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 10:                                                                       Self-employed
#> 11: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 12:                                                     Unemployed and looking for work
#> 13:                                                          Unpaid volunteer or intern
#> 14:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 15:                                           Employed full-time (35+ hours/week, paid)
#> 16:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 17:                                                                       Self-employed
#> 18: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 19:                                                     Unemployed and looking for work
#> 20:                                                          Unpaid volunteer or intern
#> 21:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 22:                                           Employed full-time (35+ hours/week, paid)
#> 23:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 24:                                                                       Self-employed
#> 25: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 26:                                                     Unemployed and looking for work
#> 27:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 28:                                           Employed full-time (35+ hours/week, paid)
#> 29:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 30:                                                                       Self-employed
#> 31: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 32:                                                     Unemployed and looking for work
#> 33:                                                          Unpaid volunteer or intern
#> 34:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 35:                                           Employed full-time (35+ hours/week, paid)
#> 36:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 37:                                                                       Self-employed
#> 38: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 39:                                                     Unemployed and looking for work
#> 40:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 41:                                           Employed full-time (35+ hours/week, paid)
#> 42:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 43:                                                                       Self-employed
#> 44: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 45:                                                     Unemployed and looking for work
#> 46:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 47:                                           Employed full-time (35+ hours/week, paid)
#> 48:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 49:                                                                       Self-employed
#> 50: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 51:                                                     Unemployed and looking for work
#> 52:                                                          Unpaid volunteer or intern
#> 53:                                           Employed full-time (35+ hours/week, paid)
#> 54:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 55:                                                                       Self-employed
#> 56: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 57:                                                     Unemployed and looking for work
#> 58:                                                          Unpaid volunteer or intern
#> 59:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 60:                                           Employed full-time (35+ hours/week, paid)
#> 61:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 62:                                                                       Self-employed
#> 63: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 64:                                                     Unemployed and looking for work
#> 65:                                                          Unpaid volunteer or intern
#> 66:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 67:                                           Employed full-time (35+ hours/week, paid)
#> 68:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 69:                                                                       Self-employed
#> 70: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 71:                                                     Unemployed and looking for work
#> 72:                                                          Unpaid volunteer or intern
#> 73:               Employed, but not currently working (e.g., on leave, furloughed 100%)
#> 74:                                           Employed full-time (35+ hours/week, paid)
#> 75:                                 Employed part-time (fewer than 35 hours/week, paid)
#> 76:                                                                       Self-employed
#> 77: Not employed and not looking for work (e.g., retired, stay-at-home parent, student)
#> 78:                                                     Unemployed and looking for work
#>                                                                              employment
#>     count        prop   est
#>     <int>       <num> <int>
#>  1:    68 0.431768082 35060
#>  2:    16 0.116033054  9422
#>  3:     9 0.076045862  6175
#>  4:    50 0.291708230 23687
#>  5:     8 0.062179037  5049
#>  6:     1 0.009987562   811
#>  7:     1 0.012278174   997
#>  8:    73 0.510666817 31621
#>  9:    12 0.108509230  6719
#> 10:     8 0.078115664  4837
#> 11:    36 0.239159574 14809
#> 12:     4 0.045428853  2813
#> 13:     1 0.002696985   167
#> 14:     1 0.015422878   955
#> 15:    77 0.522054934 40465
#> 16:     6 0.045696740  3542
#> 17:     8 0.043503503  3372
#> 18:    37 0.262814310 20371
#> 19:     7 0.055463096  4299
#> 20:     4 0.037710777  2923
#> 21:     4 0.032756641  2539
#> 22:    82 0.518540589 38288
#> 23:    12 0.112679108  8320
#> 24:     3 0.019542783  1443
#> 25:    43 0.276131531 20389
#> 26:     9 0.069002411  5095
#> 27:     1 0.004103578   303
#> 28:    80 0.516957618 43472
#> 29:    10 0.065618608  5518
#> 30:    11 0.079603292  6694
#> 31:    42 0.284759549 23946
#> 32:     6 0.025995338  2186
#> 33:     3 0.015768444  1326
#> 34:     1 0.011297151   950
#> 35:    70 0.539207351 36149
#> 36:    14 0.081741024  5480
#> 37:     6 0.041168837  2760
#> 38:    42 0.288480184 19340
#> 39:     5 0.042183142  2828
#> 40:     2 0.007219463   484
#> 41:    78 0.527234938 39066
#> 42:    10 0.050704491  3757
#> 43:    10 0.093432844  6923
#> 44:    44 0.274333297 20327
#> 45:     8 0.048342691  3582
#> 46:     1 0.005951738   441
#> 47:    79 0.479060279 35461
#> 48:    15 0.092337413  6835
#> 49:    13 0.080030261  5924
#> 50:    51 0.313825619 23230
#> 51:     4 0.030355840  2247
#> 52:     1 0.004390587   325
#> 53:    69 0.531648697 36587
#> 54:    14 0.078191752  5381
#> 55:    11 0.075212880  5176
#> 56:    36 0.245706065 16909
#> 57:     5 0.043418873  2988
#> 58:     1 0.005725246   394
#> 59:     3 0.020096486  1383
#> 60:    68 0.459676752 33333
#> 61:    10 0.088479466  6416
#> 62:     9 0.081956588  5943
#> 63:    32 0.251937557 18269
#> 64:    10 0.082094492  5953
#> 65:     2 0.022243980  1613
#> 66:     1 0.013611165   987
#> 67:    82 0.517030042 45555
#> 68:    15 0.108082035  9523
#> 69:    13 0.070344687  6198
#> 70:    40 0.260722514 22972
#> 71:     8 0.031790169  2801
#> 72:     1 0.010759400   948
#> 73:     1 0.001271153   112
#> 74:    71 0.462725456 36584
#> 75:    17 0.106701070  8436
#> 76:     9 0.063886570  5051
#> 77:    48 0.316789355 25046
#> 78:     7 0.049897549  3945
#>     count        prop   est
#> 
#> $cat$summary$weight_name
#> [1] "person_weight"
#> 
#> 
#> 
#> $num
#> NULL
#> 

hts_summary_wrapper(
summarize_var = 'race',
summarize_by = c('age', 'employment'),
wtname = 'person_weight',
)
#> $cat
#> $cat$n_ls
#> $cat$n_ls$unwtd
#> $cat$n_ls$unwtd$`Count of unique hh_id`
#> [1] 843
#> 
#> $cat$n_ls$unwtd$`Count of unique person_id`
#> [1] 1770
#> 
#> 
#> $cat$n_ls$wtd
#> Sum of person_weight 
#>              6315575 
#> 
#> 
#> $cat$summary
#> $cat$summary$unwtd
#>              age                                employment
#>            <ord>                                     <ord>
#>   1:     Under 5 Employed full-time (35+ hours/week, paid)
#>   2:     Under 5 Employed full-time (35+ hours/week, paid)
#>   3:     Under 5 Employed full-time (35+ hours/week, paid)
#>   4:     Under 5 Employed full-time (35+ hours/week, paid)
#>   5:     Under 5 Employed full-time (35+ hours/week, paid)
#>  ---                                                      
#> 427: 85 or older           Unemployed and looking for work
#> 428: 85 or older           Unemployed and looking for work
#> 429: 85 or older           Unemployed and looking for work
#> 430: 85 or older           Unemployed and looking for work
#> 431: 85 or older           Unemployed and looking for work
#>                                           race count      prop
#>                                         <fctr> <int>     <num>
#>   1:                 African American or Black    19 0.2794118
#>   2:          American Indian or Alaska Native    18 0.2647059
#>   3:                                     Asian    21 0.3088235
#>   4: Native Hawaiian or other Pacific Islander    20 0.2941176
#>   5:                                     White    19 0.2794118
#>  ---                                                          
#> 427:          American Indian or Alaska Native     2 0.2857143
#> 428:                                     Asian     2 0.2857143
#> 429: Native Hawaiian or other Pacific Islander     2 0.2857143
#> 430:                                     White     2 0.2857143
#> 431:                      Prefer not to answer     3 0.4285714
#> 
#> $cat$summary$wtd
#>              age                                employment
#>            <ord>                                     <ord>
#>   1:     Under 5 Employed full-time (35+ hours/week, paid)
#>   2:     Under 5 Employed full-time (35+ hours/week, paid)
#>   3:     Under 5 Employed full-time (35+ hours/week, paid)
#>   4:     Under 5 Employed full-time (35+ hours/week, paid)
#>   5:     Under 5 Employed full-time (35+ hours/week, paid)
#>  ---                                                      
#> 427: 85 or older           Unemployed and looking for work
#> 428: 85 or older           Unemployed and looking for work
#> 429: 85 or older           Unemployed and looking for work
#> 430: 85 or older           Unemployed and looking for work
#> 431: 85 or older           Unemployed and looking for work
#>                                           race count      prop   est
#>                                         <fctr> <int>     <num> <int>
#>   1:                 African American or Black    19 0.2678266  9390
#>   2:          American Indian or Alaska Native    18 0.2697946  9459
#>   3:                                     Asian    21 0.2495722  8750
#>   4: Native Hawaiian or other Pacific Islander    20 0.2790645  9784
#>   5:                                     White    19 0.2383058  8355
#>  ---                                                                
#> 427:          American Indian or Alaska Native     2 0.2094723   774
#> 428:                                     Asian     2 0.2094723   774
#> 429: Native Hawaiian or other Pacific Islander     2 0.3161028  1168
#> 430:                      Prefer not to answer     3 0.5412720  2000
#> 431:                                     White     2 0.3918809  1448
#>      wtd_group_n
#>            <int>
#>   1:       35060
#>   2:       35060
#>   3:       35060
#>   4:       35060
#>   5:       35060
#>  ---            
#> 427:        3695
#> 428:        3695
#> 429:        3695
#> 430:        3695
#> 431:        3695
#> 
#> $cat$summary$weight_name
#> [1] "person_weight"
#> 
#> 
#> 
#> $num
#> NULL
#> 

hts_summary_wrapper(
summarize_var = 'num_trips',
summarize_by = 'age',
weighted = FALSE)
#> $cat
#> $cat$n_ls
#> $cat$n_ls$unwtd
#> $cat$n_ls$unwtd$`Count of unique hh_id`
#> [1] 819
#> 
#> $cat$n_ls$unwtd$`Count of unique person_id`
#> [1] 1733
#> 
#> $cat$n_ls$unwtd$`Count of unique day_id`
#> [1] 3958
#> 
#> 
#> $cat$n_ls$wtd
#> NULL
#> 
#> 
#> $cat$summary
#> $cat$summary$unwtd
#>             age num_trips count       prop
#>           <ord>    <fctr> <int>      <num>
#>  1:     Under 5 1 or less    43 0.11977716
#>  2:     Under 5       1-2    62 0.17270195
#>  3:     Under 5       2-3    72 0.20055710
#>  4:     Under 5       3-4    65 0.18105850
#>  5:     Under 5       4-5    59 0.16434540
#>  6:     Under 5       5-6    40 0.11142061
#>  7:     Under 5       6-7    18 0.05013928
#>  8:        5-11 1 or less    35 0.12152778
#>  9:        5-11       1-2    50 0.17361111
#> 10:        5-11       2-3    74 0.25694444
#> 11:        5-11       3-4    43 0.14930556
#> 12:        5-11       4-5    39 0.13541667
#> 13:        5-11       5-6    31 0.10763889
#> 14:        5-11       6-7    16 0.05555556
#> 15:       12-15 1 or less    29 0.09666667
#> 16:       12-15       1-2    42 0.14000000
#> 17:       12-15       2-3    63 0.21000000
#> 18:       12-15       3-4    59 0.19666667
#> 19:       12-15       4-5    54 0.18000000
#> 20:       12-15       5-6    34 0.11333333
#> 21:       12-15       6-7    19 0.06333333
#> 22:       16-17 1 or less    43 0.12250712
#> 23:       16-17       1-2    60 0.17094017
#> 24:       16-17       2-3    61 0.17378917
#> 25:       16-17       3-4    75 0.21367521
#> 26:       16-17       4-5    64 0.18233618
#> 27:       16-17       5-6    31 0.08831909
#> 28:       16-17       6-7    17 0.04843305
#> 29:       18-24 1 or less    34 0.09855072
#> 30:       18-24       1-2    48 0.13913043
#> 31:       18-24       2-3    81 0.23478261
#> 32:       18-24       3-4    67 0.19420290
#> 33:       18-24       4-5    55 0.15942029
#> 34:       18-24       5-6    33 0.09565217
#> 35:       18-24       6-7    27 0.07826087
#> 36:       25-34 1 or less    35 0.10174419
#> 37:       25-34       1-2    65 0.18895349
#> 38:       25-34       2-3    68 0.19767442
#> 39:       25-34       3-4    70 0.20348837
#> 40:       25-34       4-5    49 0.14244186
#> 41:       25-34       5-6    39 0.11337209
#> 42:       25-34       6-7    18 0.05232558
#> 43:       35-44 1 or less    34 0.09770115
#> 44:       35-44       1-2    45 0.12931034
#> 45:       35-44       2-3    83 0.23850575
#> 46:       35-44       3-4    59 0.16954023
#> 47:       35-44       4-5    68 0.19540230
#> 48:       35-44       5-6    39 0.11206897
#> 49:       35-44       6-7    20 0.05747126
#> 50:       45-54 1 or less    45 0.14469453
#> 51:       45-54       1-2    46 0.14790997
#> 52:       45-54       2-3    67 0.21543408
#> 53:       45-54       3-4    66 0.21221865
#> 54:       45-54       4-5    41 0.13183280
#> 55:       45-54       5-6    27 0.08681672
#> 56:       45-54       6-7    19 0.06109325
#> 57:       55-64 1 or less    42 0.13291139
#> 58:       55-64       1-2    49 0.15506329
#> 59:       55-64       2-3    68 0.21518987
#> 60:       55-64       3-4    70 0.22151899
#> 61:       55-64       4-5    44 0.13924051
#> 62:       55-64       5-6    29 0.09177215
#> 63:       55-64       6-7    14 0.04430380
#> 64:       65-74 1 or less    20 0.06666667
#> 65:       65-74       1-2    42 0.14000000
#> 66:       65-74       2-3    58 0.19333333
#> 67:       65-74       3-4    68 0.22666667
#> 68:       65-74       4-5    59 0.19666667
#> 69:       65-74       5-6    33 0.11000000
#> 70:       65-74       6-7    20 0.06666667
#> 71:       75-84 1 or less    37 0.10571429
#> 72:       75-84       1-2    66 0.18857143
#> 73:       75-84       2-3    65 0.18571429
#> 74:       75-84       3-4    62 0.17714286
#> 75:       75-84       4-5    67 0.19142857
#> 76:       75-84       5-6    38 0.10857143
#> 77:       75-84       6-7    15 0.04285714
#> 78: 85 or older 1 or less    39 0.11271676
#> 79: 85 or older       1-2    71 0.20520231
#> 80: 85 or older       2-3    78 0.22543353
#> 81: 85 or older       3-4    60 0.17341040
#> 82: 85 or older       4-5    49 0.14161850
#> 83: 85 or older       5-6    34 0.09826590
#> 84: 85 or older       6-7    15 0.04335260
#>             age num_trips count       prop
#> 
#> 
#> 
#> $num
#> $num$n_ls
#> $num$n_ls$unwtd
#> $num$n_ls$unwtd$`Count of unique hh_id`
#> [1] 819
#> 
#> $num$n_ls$unwtd$`Count of unique person_id`
#> [1] 1733
#> 
#> $num$n_ls$unwtd$`Count of unique day_id`
#> [1] 3958
#> 
#> 
#> $num$n_ls$wtd
#> NULL
#> 
#> 
#> $num$summary
#> $num$summary$unwtd
#>             age count   min   max     mean median
#>           <ord> <int> <int> <int>    <num>  <num>
#>  1:     Under 5   359     0     7 3.607242      4
#>  2:        5-11   288     0     7 3.534722      3
#>  3:       12-15   300     0     7 3.796667      4
#>  4:       16-17   351     0     7 3.595442      4
#>  5:       18-24   345     0     7 3.762319      4
#>  6:       25-34   344     0     7 3.619186      4
#>  7:       35-44   348     0     7 3.793103      4
#>  8:       45-54   311     0     7 3.533762      3
#>  9:       55-64   316     0     7 3.509494      3
#> 10:       65-74   300     0     7 3.916667      4
#> 11:       75-84   350     0     7 3.634286      4
#> 12: 85 or older   346     0     7 3.471098      3
#> 
#> 
#>